% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asics_helpers.R
\name{alps_asics}
\alias{alps_asics}
\title{Export data for the spectral quantification library ASICS}
\usage{
alps_asics(samples, ...)
}
\arguments{
\item{samples}{An \code{nmr_dataset_family} 1D object}

\item{...}{Additional arguments are passed
directly to \code{ASICS::createSpectra}, which (in theory) provide an
opportunity to use distinct normalisation methods.}
}
\value{
An \code{ASICS::Spectra} object
}
\description{
A simple helper function for mangling data in the right format for the
spectral quantification library ASICS.
}
\examples{
# forAsics <- alps_asics(dataset)
# ASICS(forAsics)

}
