% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier_MUVR.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Confusion matrix of the MUVR model}
\usage{
confusion_matrix(MVObj, model = "mid")
}
\arguments{
\item{MVObj}{A MUVR object (classification analysis)}

\item{model}{min, mid or max model}
}
\value{
A confusion matrix of the model comparing actual vs predicted class
}
\description{
After creating a model with the \code{rdCV_PLS_RF} function, you can run
\code{confusion_matrix} on the model to make a confusion matrix from MUVR. This
gives information about the model performance (e.g. classification rate).
}
\examples{
\dontrun{
# 1.Build a model with the X data from your nmr object and your class:
MVObj <- rdCV_PLS_RF(nmr_data(nmr_peak_table),
Y = nmr_peak_table_completed$Timepoint)


# 2.Model performance
confusion_matrix(MVObj)

# 3.Plotting the model
MUVR_model_plot(MVObj)

# 4.Permutation test
permutations <- permutation_test_model(MVObj, nPerm = 50)

# 5.Plotting permutation test results
permutation_test_plot(MVObj, permutations, model = "Mid", type = "t")

# 6.p-Value
p.value <- p_value_perm(MVObj$miss[["mid"]], permutations[, "Mid"])

# 7.Significant variables
VIPs <- model_VIP(MVObj)

# 8.Identification
results <- nmr_identify_regions_blood(ppm_VIP_vector(VIPs))
}
}
