% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_MTBLS242_demo.R
\name{download_demo}
\alias{download_demo}
\title{Downloadind the MTBLS242 dataset}
\usage{
download_demo(to = ".")
}
\arguments{
\item{to}{A directory}
}
\value{
A folder with demo samples
}
\description{
Downloadind the MTBLS242 dataset
}
