% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset.R
\name{filter_.nmr_dataset}
\alias{filter_.nmr_dataset}
\title{Extract samples from the nmr_dataset based on metadata column criteria}
\usage{
\method{filter_}{nmr_dataset}(.data, ..., .dots)
}
\arguments{
\item{.data}{An nmr_dataset object}

\item{...}{conditions, as in dplyr}

\item{.dots}{Used to work around non-standard evaluation. See
\code{vignette("nse", package = "dplyr")}}
}
\value{
an nmr_dataset object, with the matching rows
}
\description{
Extract samples from the nmr_dataset based on metadata column criteria
}

