% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batman_helpers.R
\docType{data}
\name{hmdb}
\alias{hmdb}
\title{The Human Metabolome DataBase multiplet table}
\description{
The Human Metabolome DataBase multiplet table
}
\examples{
# Get all the 1-Methylhistidine peaks:
data("hmdb")
hmdb[hmdb$Metabolite == "1-Methylhistidine", ]
}
\references{
\url{https://hmdb.ca/}
}
\keyword{data}
