% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_identify_regions.R
\docType{data}
\name{HMDB_blood}
\alias{HMDB_blood}
\title{The Human Metabolome DataBase multiplet table: blood metabolites normally found in NMR-based metabolomics}
\description{
The Human Metabolome DataBase multiplet table: blood metabolites normally found in NMR-based metabolomics
}
\references{
\url{hmdb.ca}
}
\keyword{data}
