% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_identify_regions.R
\docType{data}
\name{HMDB_cell}
\alias{HMDB_cell}
\title{The Human Metabolome DataBase multiplet table: cell metabolites normally found in NMR-based metabolomics}
\description{
The Human Metabolome DataBase multiplet table: cell metabolites normally found in NMR-based metabolomics
}
\examples{
data("HMDB_cell")
HMDB_cell[HMDB_cell$Metabolite == "Acetone", ]
}
\references{
\url{https://hmdb.ca/}
}
\keyword{data}
