% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruker.R
\name{infer_dim_pulse_nuclei}
\alias{infer_dim_pulse_nuclei}
\title{Infer dimension, pulse sequence and nuclei given the acqus file information}
\usage{
infer_dim_pulse_nuclei(acqus_list)
}
\arguments{
\item{acqus_list}{The list of information retrieved by \code{read_acqus_file}}
}
\value{
A list with dimension, pulse_sequence and nuclei
}
\description{
This function determines the dimensionality of the sample, infers the pulse
sequence (based on the experiment name, that can be read on the acqus file
-in the EXP- field) and sets the sample nuclei field.
}
\details{
The dimension is based on the number of acqus, acqu2s, acqu3s existant files

The pulse sequence assignment is based on parsing the experiment name:
experiment name contains -> pulse sequence assigned
\itemize{
\item{NOESY -> NOESY}
\item{NOEZY -> NOESY (alias)}
\item{CPMG -> CPMG}
\item{DIFF -> DIFF}
\item{JRES -> JRES}
\item{COSY -> COSY}
\item{TOCSY -> TOCSY}
\item{DIPSI -> TOCSY (alias)}
\item{MLEV -> TOCSY (alias)}
\item{HSQC -> HSQC}
\item{HMBC -> HMBC}
}

The nuclei in the sample are set according to the pulse sequence and the
acqus parameters, following Sofia's advice.
}
\keyword{internal}

