% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset.R
\name{is.nmr_dataset}
\alias{is.nmr_dataset}
\title{Object is of \link{nmr_dataset} class}
\usage{
is.nmr_dataset(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object is an \link{nmr_dataset}, \code{FALSE} otherwise
}
\description{
Object is of \link{nmr_dataset} class
}
\examples{
dir_to_demo_dataset <- system.file("dataset-demo", package = "AlpsNMR")
dataset <- nmr_read_samples_dir(dir_to_demo_dataset)
is(dataset)

}
\concept{nmr_dataset manipulation functions}
