% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_of_lists_to_tibble}
\alias{list_of_lists_to_tibble}
\title{Convert a list of lists to a tibble}
\usage{
list_of_lists_to_tibble(ls)
}
\arguments{
\item{ls}{A list}
}
\value{
A tibble
}
\description{
Each element of \code{ls} corresponds to 1 row of the output tibble. This
element is a list, and each subelement corresponds to a column of the
output tibble.
}
\details{
One nice thing about this function is that if there are elements with missing
columns, those values will be filled with \code{NA}.

The other nice thing of this function is that if any of those elements is
itself a list or a vector, it will be nicely preserved inside the tibble cell.
}
\examples{
\dontrun{
data_as_list <- list(list(Gender = "Male", Height = 170),
                     list(Gender = "Female", Height = 160),
                     list(Gender = "Male", Weight = 80))
mydata <- list_of_lists_to_tibble(data_as_list)
}
}
\keyword{internal}
