% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{log10_minus_min_trans}
\alias{log10_minus_min_trans}
\title{Custom scale transformation for 2D plots of nmr_dataset objects}
\usage{
log10_minus_min_trans()
}
\description{
Custom scale transformation for 2D plots of nmr_dataset objects
}

