% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier_MUVR.R
\name{machine_learning}
\alias{machine_learning}
\title{Machine learning (delete if everything works properly, kept as example)}
\usage{
machine_learning(nmr_peak_table, groups, scale = T)
}
\arguments{
\item{nmr_peak_table}{a dataset with integration values}

\item{groups}{a vector with the class groups}

\item{scale}{If TRUE, the predictor variable matrix is scaled to unit variance for PLS modelling.}
}
\value{
a MUVR model
}
\description{
Machine learning on a integration object \code{nmr_peak_table}
}
