% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier_MUVR.R
\name{model_VIP}
\alias{model_VIP}
\title{model VIP values}
\usage{
model_VIP(MVObj, model = "mid")
}
\arguments{
\item{MVObj}{a MUVR model}

\item{model}{Which model to use ("min", "mid" (default), or "max")}
}
\value{
a data frame with the order, name and average rank of selected variables
}
\description{
The function extracts autoselected ranked variables from the model (MUVR object)
}
\examples{
\dontrun{
VIPs = model_VIP(MVObj)
}

}
