% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier_MUVR.R
\name{MUVR_model_plot}
\alias{MUVR_model_plot}
\title{Model plot}
\usage{
MUVR_model_plot(MVObj, model = "mid", factCols, sampLabels,
  ylim = NULL)
}
\arguments{
\item{MVObj}{An `MVobject` obtained from the MVWrap function}

\item{model}{What type of model to plot ('min', 'mid' or 'max'). Defaults to 'mid'.}

\item{factCols}{An optional vector with colors for the factor levels (in the same order as the levels)}

\item{sampLabels}{Sample labels (optional; implemented for classification)}

\item{ylim}{Optional for imposing y-limits for regression and classification analysis}
}
\value{
A plot with the model performance
}
\description{
Plot the model obtained from \code{rdCV_PLS_RF} function (a \code{MUVR} object)
}
\examples{
\dontrun{
model = rdCV_PLS_RF(X = nmr_data(nmr_peak_table),
                    Y = nmr_meta_get(nmr_peak_table, groups = "external")$Timepoint,
                    ML = F, method = "PLS", 
                    fitness = "AUROC", 
                    nRep = 12, 
                    nOuter = 4, 
                    varRatio = 0.8,
                    scale = T)

MUVR_model_plot(model)
}

}
