% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIHSnmr.R
\docType{package}
\name{NIHSnmr-package}
\alias{NIHSnmr}
\alias{NIHSnmr-package}
\title{NIHSnmr: Routines for importing and processing NMR data}
\description{
NIHSnmr allows you to import NMR samples into R and applying basic data
processing to them. See an interactive example with
\code{\link[=NIHSnmr_interactive]{NIHSnmr_interactive()}}
}
\details{
The following functions can be combined with the \link[=\%>\%]{pipe}. They
create or modify the \code{\link{nmr_dataset}} object.

\itemize{
 \item{\code{\link{nmr_read_samples_dir}} (or \code{\link{nmr_read_samples}})}
 \item{\code{\link{nmr_interpolate}}}
 \item{\code{\link{nmr_exclude_region}}}
 \item{\code{\link{nmr_normalize}}}
 \item{\code{\link[=plot.nmr_dataset]{plot}}}
}

There are also functions to extract the metadata and submit the samples to
irods, see the example below.

The \code{nmr_dataset} object is essentially a list, so it is easy to access
its components for further analysis.
}
\examples{
\dontrun{
library(NIHSnmr)
sample_names <- c("dataset/10/", "dataset/20/", "dataset/30/")
my_nmr_dataset <- nmr_read_samples(sample_names) \%>\%
  nmr_interpolate(axis1 = c(0.4, 10)) \%>\%
  nmr_exclude_region(exclude = list(water = c(4.7, 5))) \%>\%
  nmr_normalize(method = "area") \%>\%
  plot
}
\dontrun{
library(NIHSnmr)
sample_names <- c("dataset/10/", "dataset/20/", "dataset/30/")
nmrdata <- nmr_read_samples(sample_names, metadata_only = TRUE)
irods_metadata <- nmr_get_irods_meta(nmrdata)
# explore and confirm everything is in irods_metadata
# Now, let's compress the samples to zip files for irods
irods_metadata <- nmr_prepare_zip_files(meta_irods = irods_metadata,
                                        workdir = 'my_dataset_zip_files')
# Finally push the data to the irods directory
nmr_push_to_irods(irods_metadata, "/NIHSData/DUND-XXXXXX/study/Metabolomics")
}
}

