% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset.R
\name{nmr_add_metadata}
\alias{nmr_add_metadata}
\title{Add metadata to an nmr_dataset object}
\usage{
nmr_add_metadata(nmr_data, metadata, by = "NMRExperiment")
}
\arguments{
\item{nmr_data}{an \link{nmr_dataset} object}

\item{metadata}{A data frame with metadata to add}

\item{by}{A column name of both the \code{nmr_dataset$metadata} and the metadata
data.frame. If you want to merge two columns with different headers you can
use a named character vector \code{c("NMRExperiment" = "ExperimentNMR")} where
the left side is the column name of the \code{nmr_dataset$metadata} and the right side is
the column name of the metadata data frame.}
}
\value{
The nmr_dataset object with the added metadata
}
\description{
This is useful to add metadata to datasets that can be later used for
plotting spectra or further analysis (PCA...).
}
