% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_data_analysis.R
\name{nmr_data_analysis}
\alias{nmr_data_analysis}
\title{Data analysis}
\usage{
nmr_data_analysis(
  dataset,
  y_column,
  identity_column,
  external_val,
  internal_val,
  data_analysis_method
)
}
\arguments{
\item{dataset}{An \link{nmr_dataset_family} object}

\item{y_column}{A string with the name of the y column (present in the
metadata of the dataset)}

\item{identity_column}{\code{NULL} or a string with the name of the identity column (present in the
metadata of the dataset).}

\item{external_val, internal_val}{A list with two elements: \code{iterations} and \code{test_size}.
See \link{random_subsampling} for further details}

\item{data_analysis_method}{An \link{nmr_data_analysis_method} object}
}
\value{
A list with the following elements:
\itemize{
\item \code{train_test_partitions}: A list with the indices used in train and test on each of the cross-validation iterations
\item \code{inner_cv_results}: The output returned by \code{train_evaluate_model} on each inner cross-validation
\item \code{inner_cv_results_digested}: The output returned by \code{choose_best_inner}.
\item \code{outer_cv_results}: The output returned by \code{train_evaluate_model} on each outer cross-validation
\item \code{outer_cv_results_digested}: The output returned by \code{train_evaluate_model_digest_outer}.
}
}
\description{
Data analysis
}
