% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asics_helpers.R
\name{nmr_dataset_autophase}
\alias{nmr_dataset_autophase}
\title{Perform an automatic 1D NMR phasing}
\usage{
nmr_dataset_autophase(
  samples,
  method = c("NLS", "MPC_DANM", "MPC_EMP", "SPC_DANM", "SPC_EMP", "SPC_AAM", "SPC_DSM"),
  withBC = F
)
}
\arguments{
\item{samples}{An \code{nmr_dataset_family} 1D object}

\item{method}{The autophasing method -- see \code{NMRphasing::NMRphasing} for details.}

\item{withBC}{\code{NMRphasing::NMRphasing} performs an integrated baseline correction -- this parameter enables or disables it.}
}
\value{
A (hopefully better phased) \code{nmr_dataset_family} 1D object, with updated real and imaginary parts.
}
\description{
It is quite often the case that (e.g. Bruker's) 1D autophase isn't quite right.
This uses \code{NMRphasing::NMRphasing} to automatically rephase data in the spectral domain.
A number of algorithms are available (see NMRphasing's documentation),
of which NLS, MPC_DANM and SPC_DANM are the most recent.

Perhaps obviously, you are likely to get better results loading the complex data
than having real-only spectra (i.e. pass \code{all_components = T} to \code{nmr_read_samples}).

Run this before spectral interpolation.
}
\examples{
dir_to_demo_dataset <- system.file("dataset-demo", package = "AlpsNMR")
dataset <- nmr_read_samples_dir(dir_to_demo_dataset)
dataset <- nmr_dataset_autophase(dataset,method="MPC_DANM")
dataset <- nmr_interpolate_1D(dataset, axis = c(min = 1, max = 2, by = 0.002))
plot(dataset)

}
