% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIHSnmr.R
\name{nmr_dataset_load}
\alias{nmr_dataset_load}
\title{Load a \link{nmr_dataset} from a file}
\usage{
nmr_dataset_load(file_name)
}
\arguments{
\item{file_name}{The file name with the \link{nmr_dataset}}
}
\value{
the \link{nmr_dataset} object stored in \code{file_name}
}
\description{
Loads a \link{nmr_dataset} saved with \code{\link[=nmr_dataset_save]{nmr_dataset_save()}}
}
