% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset_load_save.R
\name{nmr_dataset_save}
\alias{nmr_dataset_save}
\title{Save an \link{nmr_dataset} object}
\usage{
nmr_dataset_save(nmr_dataset, file_name, ...)
}
\arguments{
\item{nmr_dataset}{The \link{nmr_dataset} object to save}

\item{file_name}{The output file name.}

\item{...}{Optional arguments passed to \link{saveRDS}.}
}
\value{
the passed \link{nmr_dataset} object
}
\description{
Wraps \code{save} to save the \link{nmr_dataset} object in \code{.RDS} format
}
