% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_detect_peaks_align.R
\name{nmr_detect_peaks_plot_peaks}
\alias{nmr_detect_peaks_plot_peaks}
\title{Plot multiple peaks from a peak list}
\usage{
nmr_detect_peaks_plot_peaks(
  nmr_dataset,
  peak_data,
  peak_ids,
  caption = paste("{peak_id}", "(NMRExp.\u00A0{NMRExperiment},",
    "\u03B3(ppb)\u00a0=\u00a0{gamma_ppb},", "\narea\u00a0=\u00a0{area},",
    "nrmse\u00a0=\u00a0{norm_rmse})")
)
}
\arguments{
\item{nmr_dataset}{The \code{nmr_dataset_1D} object with the spectra}

\item{peak_data}{A data frame, the peak list}

\item{peak_ids}{The peak ids to plot}

\item{caption}{The caption for each subplot. The default's non-breaking
spaces and the gamma sign only render correctly in a UTF-8 locale; if
\code{caption} is left at its default AND the session's locale isn't UTF-8
(\code{l10n_info()[["UTF-8"]]} is not \code{TRUE}), they're swapped for a regular space
and a plain "g" respectively, since some plotting devices otherwise
render them as their raw byte sequence instead of the intended
character. This fallback only applies to the built-in default -- an
explicitly supplied \code{caption} is always used as given.}
}
\value{
A plot object
}
\description{
Plot multiple peaks from a peak list
}
