% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_diagnose.R
\name{nmr_diagnose-defunct}
\alias{nmr_diagnose-defunct}
\alias{nmr_diagnose}
\title{Defunct function}
\usage{
nmr_diagnose()
}
\value{
nmr_normalize_extra_info will return a plot of the info
}
\description{
Defunct function
}
\keyword{internal}
