% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_diagnose.R
\name{nmr_diagnose}
\alias{nmr_diagnose}
\title{Deprecated function}
\usage{
nmr_diagnose(samples)
}
\arguments{
\item{samples}{An nmr_dataset}
}
\value{
normalized nmr
}
\description{
Deprecated function
}
\examples{
nmr_dataset <- nmr_dataset_load(system.file("extdata", "nmr_dataset.rds", package = "AlpsNMR"))
nmr_diagnose(nmr_dataset)
}
\seealso{
nmr_normalize
}
