% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_get_metadata.R
\name{nmr_export_metadata}
\alias{nmr_export_metadata}
\title{Export Metadata to an Excel file}
\usage{
nmr_export_metadata(nmr_dataset, xlsx_file, groups = c("info", "orig",
  "title", "external"))
}
\arguments{
\item{nmr_dataset}{An nmr_dataset object}

\item{xlsx_file}{"The .xlsx excel file"}

\item{groups}{A character vector. Use \code{"external"} for the external metadata or
the default for a more generic solution}
}
\value{
The Excel file name
}
\description{
Export Metadata to an Excel file
}
