% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{nmr_get_long_df}
\alias{nmr_get_long_df}
\title{Get a long data frame from nmr_data object}
\usage{
nmr_get_long_df(nmr_data, sample_idx = NULL, chemshift_range = NULL)
}
\arguments{
\item{nmr_data}{a \code{\link{nmr_dataset}} object}

\item{sample_idx}{numeric vector with the sample indices to include}

\item{chemshift_range}{range of the chemical shifts to be included. Can be of length 3
to include the resolution in the third element (e.g. \code{c(0.2, 0.8, 0.05)})}
}
\description{
This dataframe is useful for plotting with ggplot, although it may be very
long and therefore eat a lot of RAM.
}
