% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_identify_regions.R
\name{nmr_identify_regions_urine}
\alias{nmr_identify_regions_urine}
\title{NMR peak identification (urine samples)}
\usage{
nmr_identify_regions_urine(
  ppm_to_assign,
  num_proposed_compounds = 5,
  verbose = FALSE
)
}
\arguments{
\item{ppm_to_assign}{A vector with the ppm regions to assign}

\item{num_proposed_compounds}{set the number of proposed metabolites sorted by the number times reported in the HMDB: \code{HMDB_urine}.}

\item{verbose}{Logical value. Set it to TRUE to print additional information}
}
\value{
a data frame with plausible assignations.
}
\description{
Identify given regions and return a data frame with plausible assignations
in human urine samples. The data frame contains the column "Bouatra_2013" showing if
the proposed metabolite was reported in this publication as regular urinary metabolite.
}
\examples{
# We identify regions from from the corresponding ppm storaged in a vector.
ppm_to_assign <- c(
    4.060960203, 3.048970634, 2.405935596,
    3.24146865, 0.990616851, 1.002075066, 0.955325548
)
identification <- nmr_identify_regions_urine(ppm_to_assign, num_proposed_compounds = 5)
}
\seealso{
Other peak detection functions:
\code{\link{Pipelines}},
\code{\link[=nmr_baseline_threshold]{nmr_baseline_threshold()}},
\code{\link[=nmr_detect_peaks]{nmr_detect_peaks()}},
\code{\link[=nmr_detect_peaks_plot]{nmr_detect_peaks_plot()}},
\code{\link[=nmr_detect_peaks_plot_overview]{nmr_detect_peaks_plot_overview()}},
\code{\link[=nmr_detect_peaks_tune_snr]{nmr_detect_peaks_tune_snr()}},
\code{\link[=nmr_identify_regions_blood]{nmr_identify_regions_blood()}},
\code{\link[=nmr_identify_regions_cell]{nmr_identify_regions_cell()}},
\code{\link[=nmr_integrate_regions]{nmr_integrate_regions()}}

Other peak integration functions:
\code{\link{Pipelines}},
\code{\link[=get_integration_with_metadata]{get_integration_with_metadata()}},
\code{\link[=nmr_identify_regions_blood]{nmr_identify_regions_blood()}},
\code{\link[=nmr_identify_regions_cell]{nmr_identify_regions_cell()}},
\code{\link[=nmr_integrate_peak_positions]{nmr_integrate_peak_positions()}},
\code{\link[=nmr_integrate_regions]{nmr_integrate_regions()}}
}
\concept{peak detection functions}
\concept{peak integration functions}
