% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIHSnmr.R
\name{nmr_interpolate}
\alias{nmr_interpolate}
\title{Interpolate samples}
\usage{
nmr_interpolate(samples, axis1 = c(min = 0.4, max = 10, by = 8e-04),
  axis2 = NULL)
}
\arguments{
\item{samples}{A \link{nmr_dataset} object}

\item{axis1, axis2}{Axis given as \code{c(minimum, maximum, step)}}
}
\value{
The \link{nmr_dataset} object, with interpolated samples
}
\description{
Interpolate the samples so they share the same axis and can be placed
in the same matrix
}
\details{
For 2-D samples, we use two 1-D linear interpolations.
}
