% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nihs_specific.R
\name{nmr_irods_search}
\alias{nmr_irods_search}
\title{Search NMR samples in irods}
\usage{
nmr_irods_search(project_NPDI_ID = NA, study_nickname = NA,
  assay_ID = NA, master_sample_accession = NA,
  operational_sample_accession = NA, run_ID = NA,
  software_platform = NA, taxonomy = NA, hardware_platform = NA,
  assay_technique = NA, file_type = "Raw data file",
  file_format = "Zipped Bruker NMR directory",
  assay_platform = "Metabolomics", author = NA, exclude_trash = TRUE)
}
\arguments{
\item{project_NPDI_ID}{The NPDI project ID (e.g. 'DUND-700713')}

\item{study_nickname}{The study nickname, typically defined by a Project Manager
(e.g. 'st_etienne'). A study nickname belongs to a
project ID.}

\item{assay_ID}{An identifier which must be unique within a study and which corresponds
to a piece of work in the project management sense. It may be defined by
the project manager or the platform, but should be known to the project manager.
For an analytical platform it corresponds to a request to perform analysis on a
set of samples from a Project. e.g. "Plasma NMR metabolites"
or "Urine NMR metabolites".}

\item{master_sample_accession}{The master sample accession. It is a code like \code{'R0000123456'}.}

\item{operational_sample_accession}{The operational sample accession. It is a code like \code{'R0000123456'}.}

\item{run_ID}{Optional field for operational use to separate distinct batches
of samples within an Assay ID.}

\item{software_platform}{The software platform used to acquire or process the data.
One of \code{NIHSrods::getValidTerms('Software platform')}}

\item{taxonomy}{The biological taxonomy of the sample. For instance \code{'Homo sapiens (Human)'}.
The list of defined biological taxonomies at NIHS is available here:
\code{NIHSrods::getValidTerms('Taxonomy')}.}

\item{hardware_platform}{Machine which created the raw data. One of
\code{NIHSrods::getValidTerms('Hardware platform')}}

\item{assay_technique}{The analytical technique used for creating the sample.
One of \code{NIHSrods::getValidTerms('Assay technique')}}

\item{file_type}{By default: \code{"Raw data file"}.
The full list of accepted file types is available here:
\code{NIHSrods::getValidTerms('File type')}}

\item{file_format}{By default \code{"Zipped Bruker NMR directory"}. The full list of
accepted file formats is available here:
\code{NIHSrods::getValidTerms('File format')}}

\item{assay_platform}{By default: \code{'Metabolomics'}.
The full list of defined assay platforms is available here:
\code{NIHSrods::getValidTerms('Assay platform')}.}

\item{author}{This will be the person who performed the processing or analysis and made the decisions.
Required for processed data, the author who did the processing.}

\item{exclude_trash}{Exclude samples in \code{/NIHSData/trash} (Default: \code{TRUE})}
}
\value{
a data frame with the irods metadata that results from the search
}
\description{
Search for NMR samples inside irods.
}
\details{
This function is slow because NIHSrods::imeta_ls does not accept multiple files
at once.
}
\seealso{
Other NIHS specific infrastructure functions: \code{\link{nmr_get_irods_meta}},
  \code{\link{nmr_irods_validate_meta}},
  \code{\link{nmr_prepare_zip_files}},
  \code{\link{nmr_push_to_irods}},
  \code{\link{nmr_read_samples_irods}},
  \code{\link{read_exported_slims}}
}
\concept{NIHS specific infrastructure functions}
