% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nihs_specific.R
\name{nmr_irods_validate_meta}
\alias{nmr_irods_validate_meta}
\title{Validate irods_meta}
\usage{
nmr_irods_validate_meta(irods_meta)
}
\arguments{
\item{irods_meta}{A data frame with irods metadata}
}
\value{
a logical. \code{TRUE} if checked columns are in the ontology, \code{FALSE} otherwise.
}
\description{
Validate irods_meta
}
