% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nihs_specific.R
\name{nmr_prepare_zip_files}
\alias{nmr_prepare_zip_files}
\title{Create one zip file for each sample}
\usage{
nmr_prepare_zip_files(meta_irods, workdir, overwrite = FALSE, ...)
}
\arguments{
\item{meta_irods}{Data frame given by \link{nmr_get_irods_meta}}

\item{workdir}{Directory to store zip files}

\item{overwrite}{Should existing zip files be overwritten?}

\item{...}{Passed to \link[utils:zip]{utils::zip}}
}
\value{
A meta_irods data frame with an additional \code{src_path} column
containing the zip file path location.
}
\description{
Create one zip file for each sample
}
\seealso{
Other NIHS specific infrastructure functions: \code{\link{nmr_get_irods_meta}},
  \code{\link{nmr_irods_search}},
  \code{\link{nmr_irods_validate_meta}},
  \code{\link{nmr_push_to_irods}},
  \code{\link{nmr_read_samples_irods}},
  \code{\link{read_exported_slims}}
}
\concept{NIHS specific infrastructure functions}
