% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nihs_specific.R
\name{nmr_push_to_irods}
\alias{nmr_push_to_irods}
\title{Push NMR data to irods}
\usage{
nmr_push_to_irods(meta_irods, dest_path, check_metadata_issues = TRUE,
  extra_columns = NULL)
}
\arguments{
\item{meta_irods}{a dataframe as obtained from \code{\link{nmr_get_irods_meta}} with
these columns: \code{injection_id}, \code{sample_path}, \code{"src_path",
"project_NPDI_ID", "study_nickname", "assay_ID", "file_type",
"file_format", "master_sample_accession", "operational_sample_accession",
"run_ID", "assay_platform", "software_platform", "taxonomy",
"hardware_platform", "assay_technique", "author", "relates_to"}}

\item{dest_path}{The irods destination directory}

\item{check_metadata_issues}{if \code{FALSE} missing values in mandatory
\code{meta_irods} columns are allowed.}

\item{extra_columns}{Column names from \code{meta_irods} that will
be stored into irods as additional metadata values.
If \code{extra_columns} has names, then those names will be
used as attribute names.}
}
\value{
The meta_irods data frame with an additional column named \code{full_irods_path}.
}
\description{
Push NMR data to irods
}
\examples{
\dontrun{
 nmrdata <- nmr_read_samples_dir("/dir/DUND-100713-Constitutive Thinness-plasma/",
                                 metadata_only = TRUE)
 irods_metadata <- nmr_get_irods_meta(nmrdata)
 # explore and confirm everything is in there
 View(irods_metadata)
 # optionally add extra columns (but try to adhere to a standard if possible):
 irods_metadata$looks_weird <- NA
 irods_metadata$looks_weird[1] <- TRUE
 # Now, let's compress the samples to zip files for irods
 irods_metadata <- nmr_prepare_zip_files(meta_irods = irods_metadata,
                                         workdir = 'ct_plasma_zip_files')
 # Finally push the data to the irods directory
 nmr_push_to_irods(irods_metadata, "/NIHSData/DUND-100713/ct/Data/Data_raw/Metabolomics",
                   extra_columns = "looks_weird")
}
}
