% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nihs_specific.R
\name{nmr_read_samples_irods}
\alias{nmr_read_samples_irods}
\title{Read samples from irods}
\usage{
nmr_read_samples_irods(irods_search_results, ...)
}
\arguments{
\item{irods_search_results}{The data frame resulting of nmr_irods_search that at least
must have the \code{full_irods_path} column with the
irods paths that look like
\code{"/NIHSData/home/rdollermse/DUND-100713/plasma/10.zip"}}

\item{...}{passed to \code{\link{nmr_read_samples}}}
}
\value{
a \code{\link{nmr_dataset}} object
}
\description{
Read samples from irods
}
\examples{
\dontrun{
irods_results <-
  nmr_irods_search(project_NPDI_ID = "DUND-100713",
                   study_nickname = "st_etienne",
                   assay_ID = "Plasma NMR metabolites",
                   assay_technique = "Nuclear Magnetic Resonance-1D-1H-NOESY")
plasmadata <- nmr_read_samples_irods(irods_results)
}
}
\seealso{
Other NIHS specific infrastructure functions: \code{\link{nmr_get_irods_meta}},
  \code{\link{nmr_irods_search}},
  \code{\link{nmr_irods_validate_meta}},
  \code{\link{nmr_prepare_zip_files}},
  \code{\link{nmr_push_to_irods}},
  \code{\link{read_exported_slims}}
}
\concept{NIHS specific infrastructure functions}
