% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipe_exclude_regions}
\alias{pipe_exclude_regions}
\title{Title}
\usage{
pipe_exclude_regions(nmr_dataset_rds, exclude, output_dir)
}
\arguments{
\item{nmr_dataset_rds}{}

\item{output_dir}{}

\item{exclude_regions}{}
}
\description{
Title
}
