% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipe_filter_samples}
\alias{pipe_filter_samples}
\title{Pipeline: Filter samples according to metadata conditions}
\usage{
pipe_filter_samples(nmr_dataset_rds, conditions, output_dir)
}
\arguments{
\item{nmr_dataset_rds}{The nmr_dataset.rds file name coming from previous nodes}

\item{conditions}{A character vector with conditions to filter metadata.

The \code{conditions} parameter should be a character vector of valid R logical conditions.
Some examples:
\itemize{
\item conditions <- 'Gender == "Female"'
\item conditions <- 'Cohort == "Chuv"'
\item conditions <- 'TimePoint %in% c("T0", "T31")'
\item conditions <- c(Cohort == "Chuv", 'TimePoint %in% c("T0", "T31")')
}

Only samples fullfilling all the given conditions are kept in further analysis.}

\item{output_dir}{The output directory for this pipe element}
}
\description{
Pipeline: Filter samples according to metadata conditions
}
