% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipe_interpolate_1D}
\alias{pipe_interpolate_1D}
\title{Pipeline: Interpolate 1D samples}
\usage{
pipe_interpolate_1D(nmr_dataset_rds, axis1, output_dir)
}
\arguments{
\item{nmr_dataset_rds}{}

\item{axis1}{The desired ppm axis range and optionally the ppm step}

\item{output_dir}{The output directory for this pipe element}
}
\value{
This function saves the result to the output directory
}
\description{
Pipeline: Interpolate 1D samples
}
\seealso{
Other pipeline functions: \code{\link{pipe_add_metadata}},
  \code{\link{pipe_exclude_regions}},
  \code{\link{pipe_load_samples}}
}
\concept{pipeline functions}
