% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipe_peak_integration}
\alias{pipe_peak_integration}
\title{Title}
\usage{
pipe_peak_integration(nmr_dataset_rds, peak_det_align_dir, peak_width_ppm,
  output_dir)
}
\arguments{
\item{nmr_dataset_rds}{The nmr_dataset.rds file name coming from previous nodes}

\item{peak_det_align_dir}{Output directory from \link{pipe_peakdet_align}}

\item{peak_width_ppm}{A peak width in ppm}

\item{output_dir}{The output directory for this pipe element}
}
\description{
Title
}
