% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipe_peak_table_normalization}
\alias{pipe_peak_table_normalization}
\title{Pipe: Peak table PQN Normalization}
\usage{
pipe_peak_table_normalization(nmr_dataset_rds, peak_table_no_norm_fn,
  output_dir)
}
\arguments{
\item{nmr_dataset_rds}{The nmr_dataset.rds file name coming from previous nodes}

\item{peak_table_no_norm_fn}{Filename of the CSV file that results from the peak integration pipeline}

\item{output_dir}{The output directory for this pipe element}
}
\description{
Pipe: Peak table PQN Normalization
}
