% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipe_peakdet_align}
\alias{pipe_peakdet_align}
\title{Pipeline: Peak detection and Alignment}
\usage{
pipe_peakdet_align(nmr_dataset_rds, nDivRange = 128, scales = seq(1,
  16, 2), baselineThresh = 0.01, SNR.Th = -1, maxShift = 3,
  acceptLostPeak = FALSE, output_dir = NULL)
}
\arguments{
\item{nmr_dataset_rds}{The nmr_dataset.rds file name coming from previous nodes}

\item{nDivRange}{The size of a single small segment after division of spectra}

\item{scales}{The parameter of peakDetectionCWT function of MassSpecWavelet package, look it up in the original function.}

\item{baselineThresh}{It will remove all peaks under an intensity set by baselineThresh.}

\item{SNR.Th}{The parameter of peakDetectionCWT function of MassSpecWavelet package, look it up in the original function. If you set -1, the function will itself re-compute this value.}

\item{maxShift}{The maximum number of the points for a shift step.}

\item{acceptLostPeak}{This is an option for users, TRUE is the default value. If the users believe that all the peaks in the peak list are true positive, change it to FALSE.}

\item{output_dir}{The output directory for this pipe element}
}
\description{
Pipeline: Peak detection and Alignment
}
