% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.nmr_dataset}
\alias{plot.nmr_dataset}
\title{Plot an nmr_dataset}
\usage{
\method{plot}{nmr_dataset}(x, sample_idx = NULL,
  chemshift_range = NULL, quantile_plot = FALSE, interactive = FALSE,
  ...)
}
\arguments{
\item{x}{a \code{\link{nmr_dataset}} object}

\item{sample_idx}{numeric vector with the sample indices to include. Use "all" to include all}

\item{chemshift_range}{range of the chemical shifts to be included. Can be of length 3
to include the resolution in the third element (e.g. \code{c(0.2, 0.8, 0.05)})}

\item{quantile_plot}{If \code{TRUE} plot the 10\%, 50\%, 90\% percentiles of the spectra as reference.
If two numbers between 0 and 1 are given then a custom percentile can be plotted}

\item{interactive}{if \code{TRUE} return an interactive plotly plot, otherwise return a ggplot one.}

\item{...}{arguments passed to ggplot2::aes.}
}
\value{
The plot
}
\description{
Plot an nmr_dataset
}
