% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda.R
\name{plot_plsda_samples}
\alias{plot_plsda_samples}
\title{Plot PLSDA predictions}
\usage{
plot_plsda_samples(model)
}
\arguments{
\item{model}{A mixOmics plsda model}
}
\value{
A plot of the samples
}
\description{
Plot PLSDA predictions
}
\examples{
\dontrun{
plot_plsda_samples(model)
}
}
