% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda.R
\name{plsda_auroc_vip_compare}
\alias{plsda_auroc_vip_compare}
\title{Compare PLSDA auroc VIP results}
\usage{
plsda_auroc_vip_compare(...)
}
\arguments{
\item{...}{Results of \link{nmr_data_analysis} to be combined. Give each result a name.}
}
\value{
A plot of the AUC for each method
}
\description{
Compare PLSDA auroc VIP results
}
\examples{
\dontrun{
method <- plsda_auroc_vip_method(3)
#analisis missed
plsda_auroc_vip_compare()
}
}
