% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda.R
\name{plsda_auroc_vip_method}
\alias{plsda_auroc_vip_method}
\title{Method for nmr_data_analysis (PLSDA model with AUROC and VIP outputs)}
\usage{
plsda_auroc_vip_method(ncomp, auc_increment_threshold = 0.05)
}
\arguments{
\item{ncomp}{Max. number of latent variables to explore in the PLSDA analysis}

\item{auc_increment_threshold}{Choose the number of latent variables when the
AUC does not increment more than this threshold.

Returns an object to be used with \link{nmr_data_analysis} to perform a (optionally
multilevel) PLS-DA model, using the area under the ROC curve as figure of
merit to determine the optimum number of latent variables.}
}
\description{
Method for nmr_data_analysis (PLSDA model with AUROC and VIP outputs)
}
