% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_detect_peaks_align.R
\name{ppm_resolution}
\alias{ppm_resolution}
\title{Unlisted PPM resolution}
\usage{
ppm_resolution(nmr_dataset)
}
\arguments{
\item{nmr_dataset}{An object containing NMR samples}
}
\value{
A number (the ppm resolution, measured in ppms)

Numeric (the ppm resolution, measured in ppms)
}
\description{
A wrapper to unlist the output from the function
\code{nmr_ppm_resolution(nmr_dataset)} when no interpolation has been applied.
}
\examples{
nmr_dataset <- nmr_dataset_load(system.file("extdata", "nmr_dataset.rds", package = "AlpsNMR"))
nmr_ppm_resolution(nmr_dataset)
}
