% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier_MUVR.R
\name{ppm_VIP_vector}
\alias{ppm_VIP_vector}
\title{Numeric VIPs vector}
\usage{
ppm_VIP_vector(VIPs)
}
\arguments{
\item{VIPs}{a dataframe from the \code{model_VIP(MVObj)} function. It requires a
"ppms" variable}
}
\value{
a numeric ppm vector ready to be identified with
\code{nmr_identify_regions_blood}, \code{nmr_identify_cell} or
\code{nmr_identify_regions_urine}
}
\description{
The function extracts the VIPs vector (numeric) from the \code{model_VIP(MVObj)}
function. It is not necessary if you have the ppm values in a numeric vector.
This is needed in case that an automated pipeline is applied, connecting the
output from \code{model_VIP(MVObj)} to \code{nmr_identify_regions} family functions.
}
\examples{
\dontrun{
# 1.Build a model with the X data from your nmr object and your class:
MVObj <- rdCV_PLS_RF(nmr_data(nmr_peak_table),
Y = nmr_peak_table_completed$Timepoint)


# 2.Model performance
confusion_matrix(MVObj)

# 3.Plotting the model
MUVR_model_plot(MVObj)

# 4.Permutation test
permutations <- permutation_test_model(MVObj, nPerm = 50)

# 5.Plotting permutation test results
permutation_test_plot(MVObj, permutations, model = "Mid", type = "t")

# 6.p-Value
p.value <- p_value_perm(MVObj$miss[["mid"]], permutations[, "Mid"])

# 7.Significant variables
VIPs <- model_VIP(MVObj)

# 8.Identification
results <- nmr_identify_regions_blood(ppm_VIP_vector(VIPs))
}
}
