% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_jdx.R
\name{process_block}
\alias{process_block}
\title{Process a JDX block from a JDX file}
\usage{
process_block(lines, metadata_only = FALSE)
}
\arguments{
\item{lines}{A character vector with the JDX file.}

\item{metadata_only}{ignore data (do not parse it)}
}
\value{
A list with two elements: the new block and the number of lines read from lines
}
\description{
This is a very limited and partial implementation of the JDX specification,
in order to support ACD/Spectrus JCAMP-DX files.
}
\details{
JDX blocks should start with \code{##TITLE} and end with \code{##END}. Blocks
can be nested if they are of the LINK type.

One data field is allowed per block.
}
\keyword{internal}

