% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier_MUVR.R
\name{rdCV_PLS_RF}
\alias{rdCV_PLS_RF}
\title{Deprecated function}
\usage{
rdCV_PLS_RF(
  X,
  Y,
  ID,
  scale = TRUE,
  nRep = 10,
  nOuter = 5,
  nInner,
  varRatio = 0.75,
  DA = FALSE,
  fitness = "MISS",
  method = "PLS",
  nCompMax,
  methParam,
  ML = FALSE,
  modReturn = FALSE,
  logg = FALSE,
  parallel = TRUE
)
}
\description{
Deprecated function
}
\seealso{
nmr_data_analysis
}
