% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier_MUVR.R
\name{rdCV_PLS_RF_ML}
\alias{rdCV_PLS_RF_ML}
\title{Deprecated function}
\usage{
rdCV_PLS_RF_ML(
  nmr_peak_table,
  label,
  scale = TRUE,
  nRep = 10,
  nOuter = 5,
  nInner,
  varRatio = 0.75,
  DA = FALSE,
  fitness = "MISS",
  method = "PLS",
  ML = TRUE,
  modReturn = FALSE,
  logg = FALSE,
  parallel = TRUE
)
}
\description{
Deprecated function
}
