% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_rDolphin_blood.R
\name{rDolphin_plot}
\alias{rDolphin_plot}
\title{rDolphin plot}
\usage{
rDolphin_plot(rDolphin_object)
}
\arguments{
\item{rDolphin_object}{an \code{rDolphin_object} created with \code{to_rDolphin} function}
}
\value{
rDolphin plot
}
\description{
The function provides an interactive plot with the spectra of random samples.
}
\examples{
\dontrun{
## library(AlpsNMR)
rDolphin_object = to_rDolphin(parameters)
rDolphin_plot(rDolphin_object)
}
}
