% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruker.R
\name{read_acqus_file}
\alias{read_acqus_file}
\title{Reads all common acqus files}
\usage{
read_acqus_file(sample_path, acqus_files = NULL)
}
\arguments{
\item{sample_path}{A directory that corresponds to a sample}

\item{acqus_files}{The acqus files that you want to load. if \code{NULL} the
default set of acqus files is used.}
}
\value{
a list with a list of parameters for each acqus file given
}
\description{
Reads all common acqus files
}
\keyword{internal}
