% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruker.R
\name{read_bin_data}
\alias{read_bin_data}
\title{Read binary Bruker NMR data}
\usage{
read_bin_data(file_name, endian)
}
\arguments{
\item{file_name}{The file name to read}

\item{endian}{Passed to \code{readBin}.}
}
\description{
Read binary Bruker NMR data
}
\keyword{internal}

