% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruker.R
\name{read_bruker_metadata}
\alias{read_bruker_metadata}
\title{Read Bruker NMR metadata}
\usage{
read_bruker_metadata(sample_path, pdata_path = "pdata/1",
  read_pdata_title = TRUE)
}
\arguments{
\item{sample_path}{A character path of the sample directory}

\item{pdata_path}{Optional character path of the processed data (to read pdata title file)}

\item{read_pdata_title}{logical. If \code{TRUE} reads the \code{pdata/1/title}
file if it exists.}
}
\value{
an NMR sample
}
\description{
Read Bruker NMR metadata
}
