% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruker.R
\name{read_bruker_param}
\alias{read_bruker_param}
\title{Reads a BRUKER parameter file to a list}
\usage{
read_bruker_param(file_name)
}
\arguments{
\item{file_name}{File name of a BRUKER parameter file}
}
\value{
list of parameter-value pairs
}
\description{
Reads a BRUKER parameter file to a list
}
\keyword{internal}

