% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruker.R
\name{read_bruker_sample}
\alias{read_bruker_sample}
\title{Read a Bruker sample directory}
\usage{
read_bruker_sample(
  sample_path,
  pdata_file = NULL,
  pdata_path = "pdata/1",
  all_components = FALSE
)
}
\arguments{
\item{sample_path}{A character path of the sample directory}

\item{pdata_file}{File name of the binary NMR data to load. Usually "1r".
If it is null it is autodetected and all files are loaded.}

\item{pdata_path}{Path from \code{sample_path} to the preprocessed data}

\item{all_components}{If \code{FALSE} load only the real component. Otherwise load all of them}
}
\value{
a list with all the bruker sample information
}
\description{
Read a Bruker sample directory
}
