% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nihs_specific.R
\name{read_exported_slims}
\alias{read_exported_slims}
\title{Reads an exported xlsx slims spreadsheet}
\usage{
read_exported_slims(file_name, all_character = FALSE)
}
\arguments{
\item{file_name}{The exported slims spreadsheet in excel format}

\item{all_character}{Read all columns as text}
}
\value{
a data frame with the slims information

Key columns on the output data frame are "cntn_cf_fluidx_barcode" and "cntn_cf_subject_id" "cntn_cf_fk_masterSample"
}
\description{
This function reads an xlsx slims spreadsheet and converts it into a data frame
}

