% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_jdx.R
\name{read_jdx}
\alias{read_jdx}
\title{Read a ACD/Spectrus JDX file}
\usage{
read_jdx(file_names, metadata_only = FALSE)
}
\arguments{
\item{file_names}{A character vector with JDX file names}

\item{metadata_only}{A logical to load only metadata and skip the actual
spectrum. (default: \code{FALSE})}
}
\value{
A list with the JDX information
}
\description{
Read a ACD/Spectrus JDX file
}
\keyword{internal}
