% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruker.R
\name{read_orig_file}
\alias{read_orig_file}
\title{Read orig Bruker NMR file}
\usage{
read_orig_file(sample_path)
}
\arguments{
\item{sample_path}{A character path of the sample directory}
}
\value{
a list with name-value pairs
}
\description{
Read orig Bruker NMR file
}
\keyword{internal}

