% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruker.R
\name{read_pdata_title_file}
\alias{read_pdata_title_file}
\title{Read pdata title Bruker NMR file}
\usage{
read_pdata_title_file(sample_path, pdata_path = "pdata/1")
}
\arguments{
\item{sample_path}{A character path of the sample directory}

\item{pdata_path}{Path from \code{sample_path} to the preprocessed data}
}
\value{
a list with name-value pairs. If the title file has no field names,
then fields are named V1, V2...
}
\description{
Read pdata title Bruker NMR file
}
\keyword{internal}
