% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruker.R
\name{read_procs_file}
\alias{read_procs_file}
\title{Reads a Bruker procs file}
\usage{
read_procs_file(full_pdata_path, procs_files = NULL)
}
\arguments{
\item{full_pdata_path}{The path to the procs files}

\item{procs_files}{a character vector with the procs file names. If
procs_files is NULL default sensible names are tested.}
}
\value{
a list with a list of parameters for each procs file given
}
\description{
Reads a Bruker procs file
}
\keyword{internal}

