% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_detect_peaks_align.R
\name{regions_from_peak_table}
\alias{regions_from_peak_table}
\title{Build list of regions for peak integration}
\usage{
regions_from_peak_table(peak_pos_ppm, peak_width_ppm)
}
\arguments{
\item{peak_pos_ppm}{The peak positions, in ppm}

\item{peak_width_ppm}{The peak widths (or a single peak width for all peaks)}
}
\value{
A list of regions suitable for \link{nmr_integrate_regions}
}
\description{
Build list of regions for peak integration
}
\seealso{
Other peak detection functions: 
\code{\link{Pipelines}},
\code{\link{nmr_baseline_threshold}()},
\code{\link{nmr_detect_peaks_plot_overview}()},
\code{\link{nmr_detect_peaks_plot}()},
\code{\link{nmr_detect_peaks_tune_snr}()},
\code{\link{nmr_detect_peaks}()},
\code{\link{nmr_identify_regions_blood}()},
\code{\link{nmr_identify_regions_cell}()},
\code{\link{nmr_identify_regions_urine}()},
\code{\link{nmr_integrate_regions}()}
}
\concept{peak detection functions}
