% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_rDolphin_blood.R
\docType{data}
\name{ROI_blood}
\alias{ROI_blood}
\title{The template ROI_blood contains the targeted list of metabolites to be quantified}
\description{
The template ROI_blood contains the targeted list of metabolites to be quantified
}
\references{
\url{github.com/danielcanueto/rDolphin}
}
\keyword{data}
