% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_rDolphin_blood.R
\docType{data}
\name{ROI_cell}
\alias{ROI_cell}
\title{The template ROI_cell contains the targeted list of metabolites to be quantified (cell samples)}
\description{
The template ROI_cell contains the targeted list of metabolites to be quantified (cell samples)
}
\references{
\url{github.com/danielcanueto/rDolphin}
}
\keyword{data}
