% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_rDolphin_blood.R
\docType{data}
\name{ROI_urine}
\alias{ROI_urine}
\title{ROIs for urine samples}
\description{
The template ROI_urine contains the targeted list of metabolites to be quantified (urine samples)
}
\examples{
data("ROI_urine")
ROI_urine[ROI_urine$Metabolite == "Valine", ]
}
\references{
\url{github.com/danielcanueto/rDolphin}
}
\keyword{data}
