% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_rDolphin_blood.R
\name{save_profiling_output}
\alias{save_profiling_output}
\title{Save rDoplhin output}
\usage{
save_profiling_output(targeted_profiling, output_directory)
}
\arguments{
\item{targeted_profiling}{A list from Automatic_targeted_profiling function}

\item{output_directory}{a directory in which the CSV files are saved}
}
\value{
rDolphin output from Automatic_targeted_profiling function:
\itemize{
\item metabolites_intensity
\item metabolites_quantification
\item ROI_profiles_used
\item chemical_shift
\item fitting_error
\item half_bandwidth
\item signal_area_ratio
}
}
\description{
The function saves the output from Automatic_targeted_profiling function in CSV format.
}
\examples{
\dontrun{
rDolphin_object <- to_rDolphin(parameters)
targeted_profiling <- Automatic_targeted_profiling(rDolphin)
save_profiling_output(targeted_profiling, output_directory)
}
}
\seealso{
Other import/export functions:
\code{\link{Pipelines}},
\code{\link[=files_to_rDolphin]{files_to_rDolphin()}},
\code{\link{load_and_save_functions}},
\code{\link[=nmr_data]{nmr_data()}},
\code{\link[=nmr_meta_export]{nmr_meta_export()}},
\code{\link[=nmr_read_bruker_fid]{nmr_read_bruker_fid()}},
\code{\link[=nmr_read_samples]{nmr_read_samples()}},
\code{\link[=nmr_zip_bruker_samples]{nmr_zip_bruker_samples()}},
\code{\link[=save_files_to_rDolphin]{save_files_to_rDolphin()}},
\code{\link[=to_ChemoSpec]{to_ChemoSpec()}}
}
\concept{import/export functions}
