% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_rDolphin_blood.R
\name{save_profiling_plots}
\alias{save_profiling_plots}
\title{Save rDolphin output and profiling plots}
\usage{
save_profiling_plots(...)
}
\value{
Plots of the fitted compounds in pdf files
}
\description{
The function writes the plots of the adjusted compounds.
}
\examples{
\dontrun{
## library(AlpsNMR)
rDolphin_object = to_rDolphin(parameters)
targeted_profiling = Automatic_targeted_profiling(rDolphin_object)
write_plots(dir_10_rDolphin, targeted_profiling$final_output,
targeted_profiling$reproducibility_data)
}

}
