% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simplify_df}
\alias{simplify_df}
\title{Split data frame into constant and non-constant columns}
\usage{
simplify_df(data)
}
\arguments{
\item{data}{a data frame}
}
\description{
Columns that do not changes across observations are moved to the \code{const}
data frame. The rest are returned in the \code{diff} data frame.
}
\details{
The \code{const} data frame only has one row as all rows would be identical.
}
\keyword{internal}
