% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_jdx.R
\name{strip_comments}
\alias{strip_comments}
\title{Remove comments from JDX files}
\usage{
strip_comments(lines)
}
\arguments{
\item{lines}{A character vector with the lines of the jdx file}
}
\value{
A character vector where the lines have been removed
}
\description{
JDX comments start with \code{$$}.
}
\keyword{internal}
