% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset.R
\name{[.nmr_dataset}
\alias{[.nmr_dataset}
\title{Extract parts of an nmr_dataset}
\usage{
\method{[}{nmr_dataset}(x, i)
}
\arguments{
\item{x}{an \link{nmr_dataset} object}

\item{i}{indices of the samples to keep}
}
\value{
an nmr_dataset with the extracted samples
}
\description{
Extract parts of an nmr_dataset
}
\examples{
dir_to_demo_dataset <- system.file("dataset-demo", package = "AlpsNMR")
dataset <- nmr_read_samples_dir(dir_to_demo_dataset)
dataset2 <- dataset[1:3] # get the first 3 samples

}
\seealso{
Other subsetting functions:
\code{\link[=[.nmr_dataset_1D]{[.nmr_dataset_1D()}},
\code{\link[=[.nmr_dataset_peak_table]{[.nmr_dataset_peak_table()}},
\code{\link[=filter.nmr_dataset_family]{filter.nmr_dataset_family()}},
\code{\link[=nmr_pca_outliers_filter]{nmr_pca_outliers_filter()}}
}
\concept{subsetting functions}
