% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tibble_lists_columns_to_vector_columns}
\alias{tibble_lists_columns_to_vector_columns}
\title{Simplifies a tibble with lists columns of length 1 of the same type}
\usage{
tibble_lists_columns_to_vector_columns(data)
}
\arguments{
\item{data}{a tibble}
}
\value{
a tibble with converted columns
}
\description{
Simplifies a tibble with lists columns of length 1 of the same type
}
\keyword{internal}
