% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_ASICS}
\alias{to_ASICS}
\title{Export data for the ASICS spectral quantification library}
\usage{
to_ASICS(dataset, ...)
}
\arguments{
\item{dataset}{An \link{nmr_dataset_1D} object}

\item{...}{
  Arguments passed on to \code{\link[ASICS:createSpectra]{ASICS::createSpectra}}
  \describe{
    \item{\code{norm.method}}{Character specifying the normalisation method to use on
spectra ONLY if the \code{\link[ASICS]{importSpectra}} function was not used.}
    \item{\code{norm.params}}{List containing normalisation parameteres (see
\code{\link[ASICS]{normaliseSpectra}} for details) ONLY if the
\code{\link[ASICS]{importSpectra}} function was not used.}
  }}
}
\value{
An \link[ASICS:Spectra-class]{ASICS::Spectra} object
}
\description{
Exports the spectra matrix, sample names and chemical shift axis into
an ASICS Spectra object.
}
\examples{
library(ASICS)
dir_to_demo_dataset <- system.file("dataset-demo", package = "AlpsNMR")
dataset <- nmr_read_samples_dir(dir_to_demo_dataset)
dataset <- nmr_interpolate_1D(dataset, axis = c(min = 1, max = 2, by = 0.002))
forAsics <- to_ASICS(dataset)
ASICS(forAsics)
}
