% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_rDolphin_blood.R
\name{to_rDolphin}
\alias{to_rDolphin}
\title{to rDolphin}
\usage{
to_rDolphin(parameters)
}
\arguments{
\item{parameters}{the path in which the parameters CSV file is stored to create an rDolphin object}
}
\value{
an \code{rDolphin_object}
}
\description{
Import an spectra object from the rDolphin package using the information provided by the \code{parameters} file.
The obtained rDolphin object can be used to extract the metabolite profiling using \code{Automatic_targeted_profiling} function.
}
\examples{
\dontrun{
## After running AlpsNMR and get aligned spectra along with the corresponding metadata we run:
library(AlpsNMR)
setwd("your_directory_with_parameters.csv")
rDolphin_object = to_rDolphin(parameters.csv)
targeted_profiling = Automatic_targeted_profiling(rDolphin_object)
save_profiling_output(targeted_profiling, output_directory)
}
}
\seealso{
Other import/export functions: \code{\link{Automatic_targeted_profiling}},
  \code{\link{files_to_rDolphin}},
  \code{\link{load_and_save_functions}},
  \code{\link{nmr_data}}, \code{\link{nmr_meta_export}},
  \code{\link{nmr_read_bruker_fid}},
  \code{\link{nmr_read_samples}},
  \code{\link{nmr_zip_bruker_samples}},
  \code{\link{pipe_load_samples}},
  \code{\link{save_files_to_rDolphin}},
  \code{\link{save_profiling_output}},
  \code{\link{to_ASICS}}, \code{\link{to_ChemoSpec}}

Other nmr_dataset_1D functions: \code{\link{Automatic_targeted_profiling}},
  \code{\link{[.nmr_dataset_1D}},
  \code{\link{computes_peak_width_ppm}},
  \code{\link{files_to_rDolphin}},
  \code{\link{get_integration_with_metadata}},
  \code{\link{is.nmr_dataset_1D}},
  \code{\link{load_and_save_functions}},
  \code{\link{new_nmr_dataset_1D}},
  \code{\link{nmr_align_find_ref}},
  \code{\link{nmr_align}},
  \code{\link{nmr_baseline_removal}},
  \code{\link{nmr_baseline_threshold}},
  \code{\link{nmr_detect_peaks_plot}},
  \code{\link{nmr_detect_peaks_tune_snr}},
  \code{\link{nmr_detect_peaks}},
  \code{\link{nmr_exclude_region}},
  \code{\link{nmr_integrate_peak_positions}},
  \code{\link{nmr_integrate_regions}},
  \code{\link{nmr_interpolate_1D}},
  \code{\link{nmr_meta_add}},
  \code{\link{nmr_meta_export}},
  \code{\link{nmr_meta_get_column}},
  \code{\link{nmr_meta_get}}, \code{\link{nmr_normalize}},
  \code{\link{nmr_pca_build_model}},
  \code{\link{nmr_pca_outliers_filter}},
  \code{\link{nmr_pca_outliers_plot}},
  \code{\link{nmr_pca_outliers_robust}},
  \code{\link{nmr_pca_outliers}},
  \code{\link{nmr_ppm_resolution}},
  \code{\link{plot.nmr_dataset_1D}},
  \code{\link{plot_webgl}}, \code{\link{rdCV_PLS_RF}},
  \code{\link{save_files_to_rDolphin}},
  \code{\link{to_ASICS}}, \code{\link{to_ChemoSpec}},
  \code{\link{validate_nmr_dataset_1D}}
}
\concept{import/export functions}
\concept{nmr_dataset_1D functions}
