% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset_1D.R
\name{validate_nmr_dataset_1D}
\alias{validate_nmr_dataset_1D}
\title{Validate 1D nmr datasets}
\usage{
validate_nmr_dataset_1D(nmr_dataset_1D)
}
\arguments{
\item{nmr_dataset_1D}{An \link{nmr_dataset_1D} object}
}
\value{
The \link{nmr_dataset_1D} unchanged

This function is useful for its side-effects. Stopping in case of error
}
\description{
Validate 1D nmr datasets
}
