% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset_peak_table.R
\name{validate_nmr_dataset_peak_table}
\alias{validate_nmr_dataset_peak_table}
\title{Validate nmr_dataset_peak_table objects}
\usage{
validate_nmr_dataset_peak_table(nmr_dataset_peak_table)
}
\arguments{
\item{nmr_dataset_peak_table}{An \link{nmr_dataset_peak_table} object}
}
\value{
The \link{nmr_dataset_peak_table} unchanged
}
\description{
Validate nmr_dataset_peak_table objects
}
\examples{
pt <- new_nmr_dataset_peak_table(
    peak_table = matrix(c(1, 2), nrow = 1, dimnames = list("10", c("ppm_1.4", "ppm_1.6"))),
    metadata = list(external = data.frame(NMRExperiment = "10"))
)
pt_validated <- validate_nmr_dataset_peak_table(pt)

}
\seealso{
Other class helper functions:
\code{\link[=format.nmr_dataset]{format.nmr_dataset()}},
\code{\link[=format.nmr_dataset_1D]{format.nmr_dataset_1D()}},
\code{\link[=format.nmr_dataset_peak_table]{format.nmr_dataset_peak_table()}},
\code{\link[=is.nmr_dataset_1D]{is.nmr_dataset_1D()}},
\code{\link[=is.nmr_dataset_peak_table]{is.nmr_dataset_peak_table()}},
\code{\link[=new_nmr_dataset]{new_nmr_dataset()}},
\code{\link[=new_nmr_dataset_1D]{new_nmr_dataset_1D()}},
\code{\link[=new_nmr_dataset_peak_table]{new_nmr_dataset_peak_table()}},
\code{\link[=print.nmr_dataset]{print.nmr_dataset()}},
\code{\link[=print.nmr_dataset_1D]{print.nmr_dataset_1D()}},
\code{\link[=print.nmr_dataset_peak_table]{print.nmr_dataset_peak_table()}},
\code{\link[=validate_nmr_dataset]{validate_nmr_dataset()}},
\code{\link[=validate_nmr_dataset_family]{validate_nmr_dataset_family()}}
}
\concept{class helper functions}
