% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlpsNMR.R
\docType{package}
\name{AlpsNMR-package}
\alias{AlpsNMR}
\alias{AlpsNMR-package}
\title{AlpsNMR: Automated spectraL Processing System for NMR}
\description{
AlpsNMR allows you to import NMR spectra into R and provides automated and efficient signal processing for untargeted NMR metabolomics.
}
\details{
The following functions can be combined with the pipe. They
create or modify the \link{nmr_dataset} object.
\itemize{
\item \code{\link[=nmr_read_samples_dir]{nmr_read_samples_dir()}} or \code{\link[=nmr_read_samples]{nmr_read_samples()}}
\item \code{\link[=nmr_interpolate_1D]{nmr_interpolate_1D()}}
\item \code{\link[=nmr_exclude_region]{nmr_exclude_region()}}
\item \code{\link[=nmr_normalize]{nmr_normalize()}}
\item \link[=plot.nmr_dataset_1D]{plot()}
}

There are also functions to extract the metadata and submit the samples to
irods, see the example below.

The \link{nmr_dataset} object is essentially a list, so it is easy to access
its components for further analysis.
}
\examples{
\dontrun{
library(AlpsNMR)
sample_names <- c("dataset/10/", "dataset/20/", "dataset/30/")
my_nmr_dataset <- nmr_read_samples(sample_names) \%>\%
  nmr_interpolate_1D(axis = c(0.4, 10)) \%>\%
  nmr_exclude_region(exclude = list(water = c(4.6, 5))) \%>\%
  nmr_normalize(method = "pqn") \%>\%
  plot
}
}
\author{
\strong{Maintainer}: Luis Fernandez \email{lfernandez@ibecbarcelona.eu} (0000-0001-9790-6287)

Authors:
\itemize{
  \item Ivan Montoliu Roura \email{Ivan.MontoliuRoura@rd.nestle.com}
  \item Sergio Oller Moreno \email{soller@ibecbarcelona.eu} (0000-0002-8994-1549)
  \item Francisco Madrid Gambin \email{fmadrid@ibecbarcelona.eu} (0000-0001-9333-0014)
}

Other contributors:
\itemize{
  \item Nestlé Institute of Health Sciences [copyright holder]
  \item Institute for Bioengineering of Catalonia [copyright holder]
}

}
