% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier_MUVR.R
\name{AUC_model}
\alias{AUC_model}
\title{Deprecated function
Extracts AUC value}
\usage{
AUC_model(MVObj)
}
\arguments{
\item{MVObj}{a MUVR model}
}
\value{
the AUC value of the middle model
}
\description{
The function extracts the AUC value from the middle \code{MUVR} model
}
\examples{
message("MUVR is not compatible with Bioconductor, 
use bp_kfold_VIP_analysis method instead")
}
