% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_rDolphin_blood.R
\name{Automatic_targeted_profiling}
\alias{Automatic_targeted_profiling}
\title{Automatic targeted profiling}
\usage{
Automatic_targeted_profiling(
  imported_data,
  ROI = imported_data$ROI,
  optimization = TRUE,
  spectra_to_profile = NULL
)
}
\arguments{
\item{imported_data}{an \code{rDolphin_object} created with \code{to_rDolphin} function}

\item{ROI}{a ROI file containing a targeted list of metabolites to fit}

\item{optimization}{By default TRUE. If TRUE, signals parameters are optimized for profiling quality}

\item{spectra_to_profile}{By default NA. If NA, all spectra are considered. Otherwise, a vector of selected spectra.}
}
\value{
A Spectra object from the rDolphin package

A list with \code{final_output} containing the ROI-template-metabolites, their intensities, their relative quantification and quality indicators,
and \code{reproducibility_data} (to reproduce the automatic profiling). Use \code{save_profiling_output} to save the these files into your computer.
}
\description{
Automatic quantification of metabolites for all experiments using the information located in the ROI patterns file
}
\examples{
\dontrun{
# Run after apply to_rDolphin and get an rDolphin spectra object
# Set the directory in which rDolphin files will be saved
output_dir_10_rDolphin <- file.path(your_path, "10-rDolphin")
fs::dir_create(output_dir_10_rDolphin)

# Generate the files (for plasma/serum)
files_rDolphin = files_to_rDolphin_blood(nmr_dataset_0_10_ppm)

# Save the files
save_files_to_rDolphin(files_rDolphin, output_dir_10_rDolphin)

# Build the rDolphin object. Do not forget to set the directory
setwd(output_dir_10_rDolphin)
rDolphin_object = to_rDolphin("Parameters.csv")

# Visualize your spectra
rDolphin_plot(rDolphin_object)

# Run the main profiling function (it takes a while)
targeted_profiling = Automatic_targeted_profiling(rDolphin_object)

# Save results
save_profiling_output(targeted_profiling, output_dir_10_rDolphin)

save_profiling_plots(output_dir_10_rDolphin, targeted_profiling$final_output,
targeted_profiling$reproducibility_data)

#Additionally, you can run some stats
intensities = targeted_profiling$final_output$intensity
group = as.factor(rDolphin_object$Metadata$type)
model_PLS <- rdCV_PLS_RF(X = intensities, Y = group)
}
}
\seealso{
Other import/export functions: 
\code{\link{files_to_rDolphin_blood}()},
\code{\link{files_to_rDolphin_cell}()},
\code{\link{files_to_rDolphin_urine}()},
\code{\link{load_and_save_functions}},
\code{\link{nmr_data}()},
\code{\link{nmr_meta_export}()},
\code{\link{nmr_read_bruker_fid}()},
\code{\link{nmr_read_samples}()},
\code{\link{nmr_zip_bruker_samples}()},
\code{\link{pipe_load_samples}()},
\code{\link{save_files_to_rDolphin}()},
\code{\link{save_profiling_output}()},
\code{\link{to_ASICS}()},
\code{\link{to_ChemoSpec}()},
\code{\link{to_rDolphin}()}

Other nmr_dataset_1D functions: 
\code{\link{[.nmr_dataset_1D}()},
\code{\link{computes_peak_width_ppm}()},
\code{\link{file_lister}()},
\code{\link{files_to_rDolphin_blood}()},
\code{\link{files_to_rDolphin_cell}()},
\code{\link{files_to_rDolphin_urine}()},
\code{\link{get_integration_with_metadata}()},
\code{\link{is.nmr_dataset_1D}()},
\code{\link{load_and_save_functions}},
\code{\link{new_nmr_dataset_1D}()},
\code{\link{nmr_align_find_ref}()},
\code{\link{nmr_align}()},
\code{\link{nmr_baseline_removal}()},
\code{\link{nmr_baseline_threshold}()},
\code{\link{nmr_detect_peaks_plot}()},
\code{\link{nmr_detect_peaks_tune_snr}()},
\code{\link{nmr_detect_peaks}()},
\code{\link{nmr_exclude_region}()},
\code{\link{nmr_integrate_peak_positions}()},
\code{\link{nmr_integrate_regions}()},
\code{\link{nmr_interpolate_1D}()},
\code{\link{nmr_meta_add}()},
\code{\link{nmr_meta_export}()},
\code{\link{nmr_meta_get_column}()},
\code{\link{nmr_meta_get}()},
\code{\link{nmr_normalize}()},
\code{\link{nmr_pca_build_model}()},
\code{\link{nmr_pca_outliers_filter}()},
\code{\link{nmr_pca_outliers_plot}()},
\code{\link{nmr_pca_outliers_robust}()},
\code{\link{nmr_pca_outliers}()},
\code{\link{nmr_ppm_resolution}()},
\code{\link{plot.nmr_dataset_1D}()},
\code{\link{plot_webgl}()},
\code{\link{rdCV_PLS_RF_ML}()},
\code{\link{rdCV_PLS_RF}()},
\code{\link{save_files_to_rDolphin}()},
\code{\link{to_ASICS}()},
\code{\link{to_ChemoSpec}()},
\code{\link{to_rDolphin}()},
\code{\link{validate_nmr_dataset_1D}()}
}
\concept{import/export functions}
\concept{nmr_dataset_1D functions}
