% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier_MUVR.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Confusion matrix of the MUVR model}
\usage{
confusion_matrix(MVObj, model = "mid")
}
\arguments{
\item{MVObj}{A MUVR object (classification analysis)}

\item{model}{min, mid or max model}
}
\value{
A confusion matrix of the model comparing actual vs predicted class
}
\description{
The function makes a confusion matrix from a MUVR model
}
\examples{
\dontrun{
# Build a model with the X data from your nmr object and your class
MVObj <- rdCV_PLS_RF(nmr_data(nmr_peak_table),
Y = nmr_peak_table_completed$Timepoint)

# Model performance
confusion_matrix(MVObj)
}

}
