% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier_MUVR.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Deprecated function
Confusion matrix of the MUVR model}
\usage{
confusion_matrix(MVObj, model = "mid")
}
\arguments{
\item{MVObj}{a MUVR model}

\item{model}{What type of model to plot ('min', 'mid' or 'max'). Defaults to 'mid'.}
}
\value{
A confusion matrix of the model comparing actual vs predicted class
}
\description{
After creating a model with the \code{rdCV_PLS_RF} function, you can run
\code{confusion_matrix} on the model to make a confusion matrix from MUVR. This
gives information about the model performance (e.g. classification rate).
}
\examples{
message("Deprecated. MUVR is not compatible with Bioconductor, 
use bp_kfold_VIP_analysis method instead")
}
