% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_MTBLS242_demo.R
\name{download_demo}
\alias{download_demo}
\title{Download the MTBLS242 dataset}
\usage{
download_demo(to = ".")
}
\arguments{
\item{to}{A directory}
}
\value{
A folder with demo samples
}
\description{
The function downloads the NMR dataset from Gralka et al., 2015. DOI: 10.3945/ajcn.115.110536.
}
\examples{
\dontrun{ #it takes too much time
download_demo(to = ".")
}
}
\references{
\url{https://doi.org/10.3945/ajcn.115.110536}
}
