% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_MTBLS242.R
\name{download_MTBLS242}
\alias{download_MTBLS242}
\title{Download MTBLS242}
\usage{
download_MTBLS242(dest_dir = "MTBLS242")
}
\arguments{
\item{dest_dir}{Directory where the dataset should be saved}
}
\value{
A data frame with the downloaded samples
}
\description{
Downloads part of the MTBLS242 dataset (few hundred megabytes)
}
\details{
NMR dataset from Gralka et al., 2015. DOI: 10.3945/ajcn.115.110536.
}
\examples{
\dontrun{
download_MTBLS242("./MTBLS242")
}
}
