% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset.R
\name{filter.nmr_dataset}
\alias{filter.nmr_dataset}
\alias{filter.nmr_dataset_1D}
\title{Extract samples from the \link{nmr_dataset} based on metadata column criteria}
\usage{
\method{filter}{nmr_dataset}(.data, ...)

\method{filter}{nmr_dataset_1D}(.data, ...)
}
\arguments{
\item{.data}{An \link{nmr_dataset} object}

\item{...}{conditions, as in \link{dplyr}}
}
\value{
The same object, with the matching rows
}
\description{
Extract samples from the \link{nmr_dataset} based on metadata column criteria
}
