% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter.nmr_dataset}
\alias{filter.nmr_dataset}
\alias{filter.nmr_dataset_1D}
\alias{filter.nmr_dataset_peak_table}
\title{Keep samples based on metadata column criteria}
\usage{
\method{filter}{nmr_dataset}(.data, ...)

\method{filter}{nmr_dataset_1D}(.data, ...)

\method{filter}{nmr_dataset_peak_table}(.data, ...)
}
\arguments{
\item{.data}{An \link{nmr_dataset} object}

\item{...}{conditions, as in \link{dplyr}}
}
\value{
The same object, with the matching rows
}
\description{
Keep samples based on metadata column criteria
}
\seealso{
Other nmr_dataset manipulation functions: \code{\link{is.nmr_dataset}}

Other subsetting functions: \code{\link{[.nmr_dataset_1D}},
  \code{\link{[.nmr_dataset_peak_table}},
  \code{\link{[.nmr_dataset}},
  \code{\link{nmr_pca_outliers_filter}}

Other subsetting functions: \code{\link{[.nmr_dataset_1D}},
  \code{\link{[.nmr_dataset_peak_table}},
  \code{\link{[.nmr_dataset}},
  \code{\link{nmr_pca_outliers_filter}}

Other subsetting functions: \code{\link{[.nmr_dataset_1D}},
  \code{\link{[.nmr_dataset_peak_table}},
  \code{\link{[.nmr_dataset}},
  \code{\link{nmr_pca_outliers_filter}}
}
\concept{nmr_dataset manipulation functions}
\concept{nmr_dataset_1D manipulation functions}
\concept{nmr_dataset_peak_table manipulation functions}
\concept{subsetting functions}
