% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset_family.R
\name{filter.nmr_dataset_family}
\alias{filter.nmr_dataset_family}
\title{Keep samples based on metadata column criteria}
\usage{
\method{filter}{nmr_dataset_family}(.data, ...)
}
\arguments{
\item{.data}{An \link{nmr_dataset_family} object}

\item{...}{conditions, as in \link{dplyr}}
}
\value{
The same object, with the matching rows
}
\description{
Keep samples based on metadata column criteria
}
\examples{
\dontrun{
## example 1
placebo_samples <- filter(nmr_dataset, Group == "placebo")

## example 2
test_samples <- nmr_dataset \%>\% filter(nmr_peak_table$metadata$external$Group == "placebo")
}
}
\seealso{
Other subsetting functions: 
\code{\link{[.nmr_dataset_1D}()},
\code{\link{[.nmr_dataset_peak_table}()},
\code{\link{[.nmr_dataset}()},
\code{\link{nmr_pca_outliers_filter}()}
}
\concept{nmr_dataset_family manipulation functions}
\concept{subsetting functions}
