% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset.R
\name{format.nmr_dataset}
\alias{format.nmr_dataset}
\title{Format for nmr_dataset}
\usage{
\method{format}{nmr_dataset}(x, ...)
}
\arguments{
\item{x}{an \link{nmr_dataset} object}

\item{...}{for future use}
}
\value{
Format for nmr_dataset
}
\description{
Format for nmr_dataset
}
\examples{
dir_to_demo_dataset <- system.file("dataset-demo", package = "AlpsNMR")
dataset <- nmr_read_samples_dir(dir_to_demo_dataset)
format(dataset)

}
\seealso{
Other class helper functions: 
\code{\link{format.nmr_dataset_1D}()},
\code{\link{format.nmr_dataset_peak_table}()},
\code{\link{is.nmr_dataset_1D}()},
\code{\link{is.nmr_dataset_peak_table}()},
\code{\link{new_nmr_dataset_1D}()},
\code{\link{new_nmr_dataset_peak_table}()},
\code{\link{new_nmr_dataset}()},
\code{\link{print.nmr_dataset_1D}()},
\code{\link{print.nmr_dataset_peak_table}()},
\code{\link{print.nmr_dataset}()},
\code{\link{validate_nmr_dataset_family}()},
\code{\link{validate_nmr_dataset_peak_table}()},
\code{\link{validate_nmr_dataset}()}
}
\concept{class helper functions}
