% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_identify_regions.R
\docType{data}
\name{HMDB_blood}
\alias{HMDB_blood}
\title{The Human Metabolome DataBase multiplet table: blood metabolites normally found in NMR-based metabolomics}
\description{
The Human Metabolome DataBase multiplet table: blood metabolites normally found in NMR-based metabolomics
}
\examples{
data("HMDB_blood")
HMDB_blood[HMDB_blood$Metabolite == "1-Methylhistidine", ]
}
\references{
\url{https://hmdb.ca/}
}
\keyword{data}
