% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset_1D.R
\name{is.nmr_dataset_1D}
\alias{is.nmr_dataset_1D}
\title{Object is of \link{nmr_dataset_1D} class}
\usage{
is.nmr_dataset_1D(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object is an \code{nmr_dataset_1D}, \code{FALSE} otherwise
}
\description{
Object is of \link{nmr_dataset_1D} class
}
