% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset_peak_table.R
\name{is.nmr_dataset_peak_table}
\alias{is.nmr_dataset_peak_table}
\title{Object is of \link{nmr_dataset_peak_table} class}
\usage{
is.nmr_dataset_peak_table(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object is an \code{nmr_dataset_peak_table}, \code{FALSE} otherwise
}
\description{
Object is of \link{nmr_dataset_peak_table} class
}
