% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset_peak_table.R
\name{is.nmr_dataset_peak_table}
\alias{is.nmr_dataset_peak_table}
\title{Object is of \link{nmr_dataset_peak_table} class}
\usage{
is.nmr_dataset_peak_table(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object is an \code{nmr_dataset_peak_table}, \code{FALSE} otherwise
}
\description{
Object is of \link{nmr_dataset_peak_table} class
}
\seealso{
Other nmr_dataset_peak_table functions: \code{\link{[.nmr_dataset_peak_table}},
  \code{\link{load_and_save_functions}},
  \code{\link{new_nmr_dataset_peak_table}},
  \code{\link{nmr_meta_add}},
  \code{\link{nmr_meta_export}},
  \code{\link{nmr_meta_get_column}},
  \code{\link{nmr_meta_get}},
  \code{\link{validate_nmr_dataset_peak_table}}

Other class helper functions: \code{\link{is.nmr_dataset_1D}},
  \code{\link{new_nmr_dataset_1D}},
  \code{\link{new_nmr_dataset_peak_table}},
  \code{\link{new_nmr_dataset}},
  \code{\link{validate_nmr_dataset_1D}},
  \code{\link{validate_nmr_dataset_family}},
  \code{\link{validate_nmr_dataset_peak_table}},
  \code{\link{validate_nmr_dataset}}
}
\concept{class helper functions}
\concept{nmr_dataset_peak_table functions}
