% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset_load_save.R
\name{load_and_save_functions}
\alias{load_and_save_functions}
\alias{nmr_dataset_load}
\alias{nmr_dataset_save}
\title{Functions to load and save nmr_dataset objects}
\usage{
nmr_dataset_load(file_name)

nmr_dataset_save(nmr_dataset, file_name, ...)
}
\arguments{
\item{file_name}{The file name to load or save to}

\item{nmr_dataset}{An object from the \link{nmr_dataset_family}}

\item{...}{Additional arguments passed to \link{saveRDS}.}
}
\description{
Functions to load and save nmr_dataset objects
}
\seealso{
Other nmr_dataset functions: \code{\link{[.nmr_dataset}},
  \code{\link{new_nmr_dataset}},
  \code{\link{nmr_interpolate_1D}},
  \code{\link{nmr_meta_add}},
  \code{\link{nmr_meta_export}},
  \code{\link{nmr_meta_get_column}},
  \code{\link{nmr_meta_get}},
  \code{\link{nmr_ppm_resolution}},
  \code{\link{nmr_read_samples}},
  \code{\link{validate_nmr_dataset}}

Other nmr_dataset_1D functions: \code{\link{Automatic_targeted_profiling}},
  \code{\link{[.nmr_dataset_1D}},
  \code{\link{computes_peak_width_ppm}},
  \code{\link{files_to_rDolphin}},
  \code{\link{get_integration_with_metadata}},
  \code{\link{is.nmr_dataset_1D}},
  \code{\link{new_nmr_dataset_1D}},
  \code{\link{nmr_align_find_ref}},
  \code{\link{nmr_align}},
  \code{\link{nmr_baseline_removal}},
  \code{\link{nmr_baseline_threshold}},
  \code{\link{nmr_detect_peaks_plot}},
  \code{\link{nmr_detect_peaks_tune_snr}},
  \code{\link{nmr_detect_peaks}},
  \code{\link{nmr_exclude_region}},
  \code{\link{nmr_integrate_peak_positions}},
  \code{\link{nmr_integrate_regions}},
  \code{\link{nmr_interpolate_1D}},
  \code{\link{nmr_meta_add}},
  \code{\link{nmr_meta_export}},
  \code{\link{nmr_meta_get_column}},
  \code{\link{nmr_meta_get}}, \code{\link{nmr_normalize}},
  \code{\link{nmr_pca_build_model}},
  \code{\link{nmr_pca_outliers_filter}},
  \code{\link{nmr_pca_outliers_plot}},
  \code{\link{nmr_pca_outliers_robust}},
  \code{\link{nmr_pca_outliers}},
  \code{\link{nmr_ppm_resolution}},
  \code{\link{plot.nmr_dataset_1D}},
  \code{\link{plot_webgl}}, \code{\link{rdCV_PLS_RF_ML}},
  \code{\link{rdCV_PLS_RF}},
  \code{\link{save_files_to_rDolphin}},
  \code{\link{to_ASICS}}, \code{\link{to_ChemoSpec}},
  \code{\link{to_rDolphin}},
  \code{\link{validate_nmr_dataset_1D}}

Other nmr_dataset_peak_table functions: \code{\link{[.nmr_dataset_peak_table}},
  \code{\link{is.nmr_dataset_peak_table}},
  \code{\link{new_nmr_dataset_peak_table}},
  \code{\link{nmr_meta_add}},
  \code{\link{nmr_meta_export}},
  \code{\link{nmr_meta_get_column}},
  \code{\link{nmr_meta_get}},
  \code{\link{validate_nmr_dataset_peak_table}}

Other import/export functions: \code{\link{Automatic_targeted_profiling}},
  \code{\link{files_to_rDolphin}}, \code{\link{nmr_data}},
  \code{\link{nmr_meta_export}},
  \code{\link{nmr_read_bruker_fid}},
  \code{\link{nmr_read_samples}},
  \code{\link{nmr_zip_bruker_samples}},
  \code{\link{pipe_load_samples}},
  \code{\link{save_files_to_rDolphin}},
  \code{\link{save_profiling_output}},
  \code{\link{to_ASICS}}, \code{\link{to_ChemoSpec}},
  \code{\link{to_rDolphin}}
}
\concept{import/export functions}
\concept{nmr_dataset functions}
\concept{nmr_dataset_1D functions}
\concept{nmr_dataset_peak_table functions}
