% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier_MUVR.R
\name{MUVR_model_plot}
\alias{MUVR_model_plot}
\title{Deprecated function
Model plot}
\usage{
MUVR_model_plot(MVObj, model = "mid", factCols, sampLabels, ylim = NULL)
}
\value{
A plot with the model performance
}
\description{
Plot the model (a \code{MUVR} object) obtained from \code{rdCV_PLS_RF} function. For
more information about the multivariate modelling with minimally biased
variable selection (MUVR) from the \code{MUVR} package, see Shi et al., 2018 (DOI:
10.1093/bioinformatics/bty710).
}
\examples{
message("MUVR is not compatible with Bioconductor, 
use bp_kfold_VIP_analysis method instead")
}
