% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier_MUVR.R
\name{MUVR_model_plot}
\alias{MUVR_model_plot}
\title{Model plot}
\usage{
MUVR_model_plot(MVObj, model = "mid", factCols, sampLabels, ylim = NULL)
}
\arguments{
\item{MVObj}{An `MVobject` obtained from the MVWrap function}

\item{model}{What type of model to plot ('min', 'mid' or 'max'). Defaults to 'mid'.}

\item{factCols}{An optional vector with colors for the factor levels (in the same order as the levels)}

\item{sampLabels}{Sample labels (optional; implemented for classification)}

\item{ylim}{Optional for imposing y-limits for regression and classification analysis}
}
\value{
A plot with the model performance
}
\description{
Plot the model (a \code{MUVR} object) obtained from \code{rdCV_PLS_RF} function. For
more information about the multivariate modelling with minimally biased
variable selection (MUVR) from the \code{MUVR} package, see Shi et al., 2018 (DOI:
10.1093/bioinformatics/bty710).
}
\examples{
\dontrun{
# 1.Build a model with the X data from your nmr object and your class:
MVObj <- rdCV_PLS_RF(nmr_data(nmr_peak_table),
Y = nmr_peak_table_completed$Timepoint)


# 2.Model performance
confusion_matrix(MVObj)

# 3.Plotting the model
MUVR_model_plot(MVObj)

# 4.Permutation test
permutations <- permutation_test_model(MVObj, nPerm = 50)

# 5.Plotting permutation test results
permutation_test_plot(MVObj, permutations, model = "Mid", type = "t")

# 6.p-Value
p.value <- p_value_perm(MVObj$miss[["mid"]], permutations[, "Mid"])

# 7.Significant variables
VIPs <- model_VIP(MVObj)

# 8.Identification
results <- nmr_identify_regions_blood(ppm_VIP_vector(VIPs))
}
}
