% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset_peak_table.R
\name{new_nmr_dataset_peak_table}
\alias{new_nmr_dataset_peak_table}
\title{Creates a new nmr_dataset_peak_table object from scratch}
\usage{
new_nmr_dataset_peak_table(peak_table, metadata)
}
\arguments{
\item{peak_table}{A numeric matrix with one NMR spectrum on each row}

\item{metadata}{A list of data frames with at least the \code{NMRExperiment} column}
}
\description{
Creates a new nmr_dataset_peak_table object from scratch
}
\seealso{
Other nmr_dataset_peak_table functions: \code{\link{[.nmr_dataset_peak_table}},
  \code{\link{is.nmr_dataset_peak_table}},
  \code{\link{load_and_save_functions}},
  \code{\link{nmr_meta_add}},
  \code{\link{nmr_meta_export}},
  \code{\link{nmr_meta_get_column}},
  \code{\link{nmr_meta_get}},
  \code{\link{validate_nmr_dataset_peak_table}}

Other class helper functions: \code{\link{is.nmr_dataset_1D}},
  \code{\link{is.nmr_dataset_peak_table}},
  \code{\link{new_nmr_dataset_1D}},
  \code{\link{new_nmr_dataset}},
  \code{\link{validate_nmr_dataset_1D}},
  \code{\link{validate_nmr_dataset_family}},
  \code{\link{validate_nmr_dataset_peak_table}},
  \code{\link{validate_nmr_dataset}}
}
\concept{class helper functions}
\concept{nmr_dataset_peak_table functions}
