% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset_peak_table.R
\name{new_nmr_dataset_peak_table}
\alias{new_nmr_dataset_peak_table}
\title{Creates a new nmr_dataset_peak_table object from scratch}
\usage{
new_nmr_dataset_peak_table(peak_table, metadata)
}
\arguments{
\item{peak_table}{A numeric matrix with one NMR spectrum on each row}

\item{metadata}{A list of data frames with at least the \code{NMRExperiment} column}
}
\description{
Creates a new nmr_dataset_peak_table object from scratch
}
