% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset_peak_table.R
\name{new_nmr_dataset_peak_table}
\alias{new_nmr_dataset_peak_table}
\title{Creates a new nmr_dataset_peak_table object from scratch}
\usage{
new_nmr_dataset_peak_table(peak_table, metadata)
}
\arguments{
\item{peak_table}{A numeric matrix with one NMR spectrum on each row}

\item{metadata}{A list of data frames with at least the \code{NMRExperiment} column}
}
\value{
Creates a new nmr_dataset_peak_table object from scratch
}
\description{
Creates a new nmr_dataset_peak_table object from scratch
}
\examples{
\dontrun{
 Error: metadata should be a list 
#' dir_to_demo_dataset <- system.file("dataset-demo", package = "AlpsNMR")
dataset <- nmr_read_samples_dir(dir_to_demo_dataset)
dataset_1D <- nmr_interpolate_1D(dataset, axis = c(min = -0.5, max = 10, by = 2.3E-4))
metadata <- nmr_meta_get(dataset)
peak_table <- nmr_data(dataset_1D)
new <- new_nmr_dataset_peak_table(peak_table, metadata)
}
\dontrun{
Error: multiples errores
metadata_1D <- list(external = data.frame(NMRExperiment = c("10", "20")))
# Sample 10 and Sample 20 can have different lengths (due to different setups)
data_fields_1D <- list(data_1r = list(runif(16), runif(32)))
# Each sample has its own axis list, with one element (because this example is 1D)
axis_1D <- list(list(1:16), list(1:32))
nmr_dataset <- new_nmr_dataset(metadata_1D, data_fields_1D, axis_1D)
nmr_dataset <- nmr_interpolate_1D(nmr_dataset, axis = c(min = -0.5, max = 10, by = 2.3E-4))

# 1.Peak detection in the dataset.
peak_data <- nmr_detect_peaks(nmr_dataset,
                              nDivRange_ppm = 0.1, # Size of detection segments
                              scales = seq(1, 16, 2),
                              baselineThresh = NULL, # Minimum peak intensity
                              SNR.Th = 4) # Signal to noise ratio

# 2.Find the reference spectrum to align with.
NMRExp_ref <- nmr_align_find_ref(nmr_dataset, peak_data)

# 3.Spectra alignment using the ref spectrum and a maximum alignment shift
nmr_dataset <- nmr_align(nmr_dataset, # the dataset
                         peak_data, # detected peaks
                         NMRExp_ref = NMRExp_ref, # ref spectrum
                         maxShift_ppm = 0.0015, # max alignment shift
                         acceptLostPeak = FALSE) # lost peaks

# 4.PEAK INTEGRATION (please, consider previous normalization step).
# First we take the peak table from the reference spectrum
peak_data_ref <- filter(peak_data, NMRExperiment == NMRExp_ref)

# Then we integrate spectra considering the peaks from the ref spectrum
nmr_peak_table <- nmr_integrate_peak_positions(
                      samples = nmr_dataset,
                      peak_pos_ppm = peak_data_ref$ppm,
                      peak_width_ppm = NULL)
new_nmr_dataset <- new_nmr_dataset_peak_table(nmr_peak_table, metadata_1D)
}
}
\seealso{
Other nmr_dataset_peak_table functions: 
\code{\link{[.nmr_dataset_peak_table}()},
\code{\link{is.nmr_dataset_peak_table}()},
\code{\link{load_and_save_functions}},
\code{\link{nmr_meta_add}()},
\code{\link{nmr_meta_export}()},
\code{\link{nmr_meta_get_column}()},
\code{\link{nmr_meta_get}()},
\code{\link{validate_nmr_dataset_peak_table}()}

Other class helper functions: 
\code{\link{format_nmr_dataset_1D}},
\code{\link{format_nmr_dataset}},
\code{\link{is.nmr_dataset_1D}()},
\code{\link{is.nmr_dataset_peak_table}()},
\code{\link{new_nmr_dataset_1D}()},
\code{\link{new_nmr_dataset}()},
\code{\link{print_nmr_dataset_1D}},
\code{\link{print_nmr_dataset}},
\code{\link{validate_nmr_dataset_family}()},
\code{\link{validate_nmr_dataset_peak_table}()},
\code{\link{validate_nmr_dataset}()}
}
\concept{class helper functions}
\concept{nmr_dataset_peak_table functions}
