% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_get_metadata.R
\name{nmr_add_metadata}
\alias{nmr_add_metadata}
\title{Add metadata to an nmr_dataset object}
\usage{
nmr_add_metadata(nmr_data, metadata, by = "NMRExperiment")
}
\arguments{
\item{nmr_data}{an \link{nmr_dataset} object}

\item{metadata}{A data frame with metadata to add}

\item{by}{A column name of both the \code{nmr_dataset$metadata$external} and the metadata
data.frame. If you want to merge two columns with different headers you can
use a named character vector \code{c("NMRExperiment" = "ExperimentNMR")} where
the left side is the column name of the \code{nmr_dataset$metadata$external} and the right side is
the column name of the metadata data frame.}
}
\value{
The nmr_dataset object with the added metadata
}
\description{
This is useful to add metadata to datasets that can be later used for
plotting spectra or further analysis (PCA...).
}
\examples{
# Load a demo dataset with four samples:
dataset <- system.file("dataset-demo", package = "NIHSnmr")
nmr_dataset <- nmr_read_samples_dir(dataset)

# At first we just have the NMRExperiment column
print(nmr_dataset$metadata$external)
# The first excel file contains an "NMRExperiment" column:
first_excel_fn <- system.file("dataset-demo", "first_excel.xlsx", package = "NIHSnmr")
first_excel <- readxl::read_excel(first_excel_fn)
print(first_excel)
# We can link the SubjectID column of the first excel into the dataset
nmr_dataset <- nmr_add_metadata(nmr_dataset, first_excel, by = "NMRExperiment")
print(nmr_dataset$metadata$external)
# The second excel can use the SubjectID:
second_excel_fn <- system.file("dataset-demo", "second_excel.xlsx", package = "NIHSnmr")
second_excel <- readxl::read_excel(second_excel_fn)
print(second_excel)
# Add the metadata by its SubjectID:
nmr_dataset <- nmr_add_metadata(nmr_dataset, second_excel, by = "SubjectID")
# The final loaded metadata:
print(nmr_dataset$metadata$external)

}
