% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_detect_peaks_align.R
\name{nmr_align_find_ref}
\alias{nmr_align_find_ref}
\title{Find alignment reference}
\usage{
nmr_align_find_ref(nmr_dataset, peak_data)
}
\arguments{
\item{nmr_dataset}{An \link{nmr_dataset_1D}}

\item{peak_data}{The detected peak data given by \link{nmr_detect_peaks}.}
}
\value{
The NMRExperiment of the reference sample
}
\description{
Find alignment reference
}
