% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_baseline_threshold.R
\name{nmr_baseline_threshold}
\alias{nmr_baseline_threshold}
\title{Threshold estimation for peak detection}
\usage{
nmr_baseline_threshold(nmr_dataset, range_without_peaks = c(9.5, 10))
}
\arguments{
\item{nmr_dataset}{An \link{nmr_dataset_1D}.}

\item{range_without_peaks}{A vector with two doubles describing a range without peaks suitable for baseline detection}
}
\value{
Numerical. A threshold value in intensity below that no peak is detected.
}
\description{
Estimates a robust baseline threshold value for peak detection on an \link{nmr_dataset_1D} object.
This is performed computing the median and the mean absolute deviation of each spectrum
in the 9.5-10 ppm range. The threshold is computed as the \code{median+3*mad}.
}
\examples{
dir_to_demo_dataset <- system.file("dataset-demo", package = "AlpsNMR")
dataset <- nmr_read_samples_dir(dir_to_demo_dataset)
dataset_1D <- nmr_interpolate_1D(dataset, axis = c(min = -0.5, max = 10, by = 2.3E-4))
bl_threshold <- nmr_baseline_threshold(dataset_1D)

}
\seealso{
Other peak detection functions: 
\code{\link{Pipelines}},
\code{\link{nmr_detect_peaks_plot_overview}()},
\code{\link{nmr_detect_peaks_plot}()},
\code{\link{nmr_detect_peaks_tune_snr}()},
\code{\link{nmr_detect_peaks}()},
\code{\link{nmr_identify_regions_blood}()},
\code{\link{nmr_identify_regions_cell}()},
\code{\link{nmr_identify_regions_urine}()},
\code{\link{nmr_integrate_regions}()}
}
\concept{peak detection functions}
