% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_baseline_threshold.R
\name{nmr_baseline_threshold_plot}
\alias{nmr_baseline_threshold_plot}
\title{Plot the baseline thresholds}
\usage{
nmr_baseline_threshold_plot(
  nmr_dataset,
  thresholds,
  NMRExperiment = "all",
  chemshift_range = c(9.5, 10),
  ...
)
}
\arguments{
\item{nmr_dataset}{An \link{nmr_dataset_1D} object}

\item{thresholds}{A named vector. The values are baseline thresholds. The names are NMRExperiments.}

\item{NMRExperiment}{The NMRExperiments to plot (Use \code{"all"} to plot all of them)}

\item{chemshift_range}{The range to plot, as a first check use the \code{range_without_peaks} from \link{nmr_baseline_threshold}}

\item{...}{arguments passed to \link[ggplot2:aes_]{ggplot2::aes_string}.}
}
\value{
A plot.
}
\description{
If you have a lot of samples you can make the plot window bigger (or
use "\verb{ ```\{r fig.height=10, fig.width=10\}}" in notebooks), or choose some NMRExperiments.
}
