% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batman_helpers.R
\name{nmr_batman}
\alias{nmr_batman}
\alias{nmr_batman_options}
\alias{nmr_batman_write_options}
\alias{nmr_batman_export_dataset}
\alias{nmr_batman_multi_data_user_hmdb}
\alias{nmr_batman_multi_data_user}
\alias{nmr_batman_metabolites_list}
\title{Batman helpers}
\usage{
nmr_batman_options(ppmRange = matrix(c(3, 3.1, 3.6, 3.7, 3.9, 4, 4, 4.1, 6.95,
  7.05, 7.6, 7.7, 7.8, 7.9), ncol = 2, byrow = TRUE), specNo = "1",
  paraProc = 4L, negThresh = -0.5, scaleFac = 1e+06, downSamp = 1,
  hiresFlag = 1, randSeed = 100025L, nItBurnin = 200L,
  nItPostBurnin = 5000L, multFile = 2L, thinning = 50L, cfeFlag = 0,
  nItRerun = 5000L, startTemp = 1000, specFreq = 600, a = 1e-05,
  b = 1e-09, muMean = 1.1, muVar = 0.2, muVar_prop = 0.002,
  nuMVar = 0.0025, nuMVarProp = 0.1, tauMean = -0.05, tauPrec = 2,
  rdelta = 0.02, csFlag = 0)

nmr_batman_write_options(bopts, batman_dir = "BatmanInput",
  filename = "batmanOptions.txt")

nmr_batman_export_dataset(nmr_dataset, batman_dir = "BatmanInput",
  filename = "NMRdata.txt")

nmr_batman_multi_data_user_hmdb(batman_dir = "BatmanInput",
  filename = "multi_data_user.csv")

nmr_batman_multi_data_user(multiplet_table, batman_dir = "BatmanInput",
  filename = "multi_data_user.csv")

nmr_batman_metabolites_list(metabolite_names, batman_dir = "BatmanInput",
  filename = "metabolitesList.csv")
}
\arguments{
\item{ppmRange}{ppm ranges for analys}

\item{specNo}{Ranges of spectra number to be included}

\item{bopts}{Batman options}

\item{batman_dir}{Batman input directorye}

\item{filename}{Filename to use, inside \code{batman_dir}}

\item{metabolite_names}{A character vector of the metabolite names to consider}
}
\description{
These are helper functions to make Batman tests easier
}
