% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_data.R
\name{nmr_data}
\alias{nmr_data}
\title{Return the full spectra matrix}
\usage{
nmr_data(dataset, ...)
}
\arguments{
\item{dataset}{}

\item{...}{}
}
\value{
a matrix
}
\description{
Return the full spectra matrix
}
