% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_data.R
\name{nmr_data}
\alias{nmr_data}
\alias{nmr_data<-}
\alias{nmr_data<-.nmr_dataset_1D}
\title{Set/Return the full spectra matrix}
\usage{
nmr_data(nmr_dataset, ...)

nmr_data(nmr_dataset) <- value

\method{nmr_data}{nmr_dataset_1D}(nmr_dataset) <- value
}
\arguments{
\item{nmr_dataset}{An object from the \link{nmr_dataset_family} to get the raw data from}

\item{...}{Unused and left for future compatibility}

\item{value}{A matrix}
}
\value{
a matrix

The given nmr_dataset
}
\description{
Set/Return the full spectra matrix
}
\seealso{
Other import/export functions: \code{\link{Automatic_targeted_profiling}},
  \code{\link{files_to_rDolphin}},
  \code{\link{load_and_save_functions}},
  \code{\link{nmr_meta_export}},
  \code{\link{nmr_read_bruker_fid}},
  \code{\link{nmr_read_samples}},
  \code{\link{nmr_zip_bruker_samples}},
  \code{\link{pipe_load_samples}},
  \code{\link{save_files_to_rDolphin}},
  \code{\link{save_profiling_output}},
  \code{\link{to_ASICS}}, \code{\link{to_ChemoSpec}},
  \code{\link{to_rDolphin}}
}
\concept{import/export functions}
