% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_data.R
\name{nmr_data}
\alias{nmr_data}
\title{Return the full spectra matrix}
\usage{
nmr_data(nmr_dataset, ...)
}
\arguments{
\item{nmr_dataset}{The NMR object to get the raw data from}

\item{...}{Unused and left for future compatibility}
}
\value{
a matrix
}
\description{
Return the full spectra matrix
}
