% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset.R
\name{nmr_dataset}
\alias{nmr_dataset}
\title{nmr_dataset (S3 class)}
\description{
The \code{nmr_dataset} represents a set of NMR samples.
It is defined as an S3 class, and it can be treated as a regular list.
}
\details{
It currently has the following elements:

\describe{
  \item{metadata}{The metadata fields. A data frame with one row per sample}
  \item{axis}{A list with length equal to the dimensionality of the data.
              For 1D spectra it is a list with one numeric vector}
  \item{data_*}{Data arrays with the actual spectra. The first index represents
                the sample, the rest of the
                indices match the length of each \code{axis}. Typically
                \code{data_1r} is a matrix with one sample on each row
                and the chemical shifts in the columns.}
  \item{processing}{The processing steps performed on the object, such as
                    interpolation, exclusion or normalization.}
}
}

