% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset_family.R
\name{nmr_dataset_family}
\alias{nmr_dataset_family}
\title{nmr_dataset like objects (S3 classes)}
\description{
The NIHSnmr package defines and uses several objects to manage NMR Data.
}
\details{
These objects share some structure and functions, so it makes sense to have
an abstract class to ensure that the shared structures are compatible
}
\seealso{
\link[=load_and_save_functions]{Functions to save and load these objects}

Other NIHSnmr dataset objects: \code{\link{nmr_dataset_1D}},
  \code{\link{nmr_dataset}}
}
\concept{NIHSnmr dataset objects}
