% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIHSnmr.R
\name{nmr_dataset_load}
\alias{nmr_dataset_load}
\title{Load a nmr_dataset from a file}
\usage{
nmr_dataset_load(file_name)
}
\arguments{
\item{file_name}{The file name with the nmr_dataset}
}
\value{
the nmr_dataset object stored in file_name or NULL
}
\description{
Loads a nmr_dataset saved with nmr_save_dataset
}
