% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_detect_peaks_align.R
\name{nmr_detect_peaks}
\alias{nmr_detect_peaks}
\title{Peak detection for NMR}
\usage{
nmr_detect_peaks(nmr_dataset, nDivRange_ppm = 0.1, scales = seq(1, 16,
  2), baselineThresh = 0.01, SNR.Th = -1)
}
\arguments{
\item{nmr_dataset}{An \link{nmr_dataset_1D}.}

\item{nDivRange_ppm}{Segment size, in ppms, to divide the spectra and search for peaks.}

\item{scales}{The parameter of peakDetectionCWT function of MassSpecWavelet package, look it up in the original function.}

\item{baselineThresh}{It will remove all peaks under an intensity set by baselineThresh.}

\item{SNR.Th}{The parameter of peakDetectionCWT function of MassSpecWavelet package, look it up in the original function. If you set -1, the function will itself re-compute this value.}
}
\value{
A data frame with the NMRExperiment, the sample index, the position in ppm and index and the peak intensity
}
\description{
Detects peaks on an \link{nmr_dataset_1D} object, using \link[speaq:detectSpecPeaks]{speaq::detectSpecPeaks}.
\code{detectSpecPeaks} divides the whole spectra into smaller segments and uses
\link[MassSpecWavelet:peakDetectionCWT]{MassSpecWavelet::peakDetectionCWT} for peak detection.
}
