% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_detect_peaks_align.R
\name{nmr_detect_peaks}
\alias{nmr_detect_peaks}
\title{Peak detection for NMR}
\usage{
nmr_detect_peaks(nmr_dataset, nDivRange_ppm = 0.1, scales = seq(1, 16,
  2), baselineThresh = 0, SNR.Th = 3)
}
\arguments{
\item{nmr_dataset}{An \link{nmr_dataset_1D}.}

\item{nDivRange_ppm}{Segment size, in ppms, to divide the spectra and search for peaks.}

\item{scales}{The parameter of peakDetectionCWT function of MassSpecWavelet package, look it up in the original function.}

\item{baselineThresh}{It will remove all peaks under an intensity set by baselineThresh.}

\item{SNR.Th}{The parameter of peakDetectionCWT function of MassSpecWavelet package, look it up in the original function. If you set -1, the function will itself re-compute this value.}
}
\value{
A data frame with the NMRExperiment, the sample index, the position in ppm and index and the peak intensity
}
\description{
Detects peaks on an \link{nmr_dataset_1D} object, using \link[speaq:detectSpecPeaks]{speaq::detectSpecPeaks}.
\code{detectSpecPeaks} divides the whole spectra into smaller segments and uses
\link[MassSpecWavelet:peakDetectionCWT]{MassSpecWavelet::peakDetectionCWT} for peak detection.
}
\seealso{
\link{nmr_align} for peak alignment with the detected peak table

Other peak detection functions: \code{\link{nmr_detect_peaks_plot}},
  \code{\link{nmr_detect_peaks_tune_snr}},
  \code{\link{nmr_integrate_regions}},
  \code{\link{pipe_peakdet_align}},
  \code{\link{regions_from_peak_table}}

Other nmr_dataset_1D functions: \code{\link{[.nmr_dataset_1D}},
  \code{\link{is.nmr_dataset_1D}},
  \code{\link{load_and_save_functions}},
  \code{\link{new_nmr_dataset_1D}},
  \code{\link{nmr_align_find_ref}},
  \code{\link{nmr_align}},
  \code{\link{nmr_detect_peaks_plot}},
  \code{\link{nmr_detect_peaks_tune_snr}},
  \code{\link{nmr_exclude_region}},
  \code{\link{nmr_integrate_peak_positions}},
  \code{\link{nmr_integrate_regions}},
  \code{\link{nmr_interpolate_1D}},
  \code{\link{nmr_meta_add}},
  \code{\link{nmr_meta_export}},
  \code{\link{nmr_meta_get_column}},
  \code{\link{nmr_meta_get}}, \code{\link{nmr_normalize}},
  \code{\link{nmr_pca_build_model}},
  \code{\link{nmr_pca_outliers_filter}},
  \code{\link{nmr_pca_outliers_plot}},
  \code{\link{nmr_pca_outliers_robust}},
  \code{\link{nmr_pca_outliers}},
  \code{\link{nmr_ppm_resolution}},
  \code{\link{plot.nmr_dataset_1D}},
  \code{\link{plot_webgl}}, \code{\link{to_ASICS}},
  \code{\link{to_ChemoSpec}},
  \code{\link{validate_nmr_dataset_1D}}
}
\concept{nmr_dataset_1D functions}
\concept{peak detection functions}
