% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_detect_peaks_align.R
\name{nmr_detect_peaks_plot}
\alias{nmr_detect_peaks_plot}
\title{Plot peak detection results}
\usage{
nmr_detect_peaks_plot(nmr_dataset, peak_data, NMRExperiment, ...)
}
\arguments{
\item{NMRExperiment}{a single NMR experiment to plot}

\item{...}{Arguments passed to \link{plot.nmr_dataset_1D} (\code{chemshift_range}, \code{...})}
}
\description{
Plot peak detection results
}
\seealso{
Other peak detection functions: \code{\link{nmr_detect_peaks}}
}
\concept{peak detection functions}
