% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_detect_peaks_align.R
\name{nmr_detect_peaks_plot_peaks}
\alias{nmr_detect_peaks_plot_peaks}
\title{Plot multiple peaks from a peak list}
\usage{
nmr_detect_peaks_plot_peaks(
  nmr_dataset,
  peak_data,
  peak_ids,
 
    caption = "{peak_id} (NMRExp. {NMRExperiment}, γ = {gamma},\\narea = {area}, nrmse = {norm_rmse})"
)
}
\arguments{
\item{nmr_dataset}{The \code{nmr_dataset_1D} object with the spectra}

\item{peak_data}{A data frame, the peak list}

\item{peak_ids}{The peak ids to plot}

\item{caption}{The caption for each subplot}
}
\value{
A plot object
}
\description{
Plot multiple peaks from a peak list
}
