% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_detect_peaks_align.R
\name{nmr_detect_peaks_tune_snr}
\alias{nmr_detect_peaks_tune_snr}
\title{Diagnose SNR threshold in peak detection}
\usage{
nmr_detect_peaks_tune_snr(
  ds,
  NMRExperiment = NULL,
  SNR_thresholds = seq(from = 2, to = 6, by = 0.1),
  ...
)
}
\arguments{
\item{ds}{An \link{nmr_dataset_1D} dataset}

\item{NMRExperiment}{A string with the single NMRExperiment used explore the SNR thresholds. If not given, use the first one.}

\item{SNR_thresholds}{A numeric vector with the SNR thresholds to explore}

\item{...}{
  Arguments passed on to \code{\link[=nmr_detect_peaks]{nmr_detect_peaks}}
  \describe{
    \item{\code{nmr_dataset}}{An \link{nmr_dataset_1D}.}
    \item{\code{nDivRange_ppm}}{Segment size, in ppms, to divide the spectra and search
for peaks.}
    \item{\code{baselineThresh}}{It will remove all peaks under an intensity set by
\code{baselineThresh}. If you set it to \code{NULL}, \code{nmr_detect_peaks()} will
automatically estimate the baseline on the region given by the
\code{range_without_peaks} argument.}
    \item{\code{range_without_peaks}}{A numeric vector of length two with a region without peaks, only used when \code{baselineThresh = NULL}}
    \item{\code{verbose}}{Logical (\code{TRUE} or \code{FALSE}). Show informational messages, such as the estimated baseline}
    \item{\code{scales}}{The parameter of peakDetectionCWT function of MassSpecWavelet package, look it up in the original function.}
    \item{\code{SNR.Th}}{The parameter of peakDetectionCWT function of MassSpecWavelet package, look it up in the original function. If you set -1, the function will itself re-compute this value.}
  }}
}
\value{
A list with the following elements:
\itemize{
\item \code{peaks_detected}: A data frame with the columns from the \link{nmr_detect_peaks} output and an additional column
\code{SNR_threshold} with the threshold used on each row.
\item \code{num_peaks_per_region}: A summary of the \code{peaks_detected} table, with the number of peaks detected on
each chemical shift region
\item \code{plot_num_peaks_per_region}: A visual representation of \code{num_peaks_per_region}
\item \code{plot_spectrum_and_detections}: A visual representation of the spectrum and the peaks detected with each
SNR threshold. Use \link[plotly:ggplotly]{plotly::ggplotly} or \link{plot_interactive} on this to zoom and explore the results.
}
}
\description{
Diagnose SNR threshold in peak detection
}
\seealso{
nmr_detect_peaks

Other peak detection functions: 
\code{\link{Pipelines}},
\code{\link{nmr_baseline_threshold}()},
\code{\link{nmr_detect_peaks_plot_overview}()},
\code{\link{nmr_detect_peaks_plot}()},
\code{\link{nmr_detect_peaks}()},
\code{\link{nmr_identify_regions_blood}()},
\code{\link{nmr_identify_regions_cell}()},
\code{\link{nmr_identify_regions_urine}()},
\code{\link{nmr_integrate_regions}()},
\code{\link{regions_from_peak_table}()}

Other nmr_dataset_1D functions: 
\code{\link{[.nmr_dataset_1D}()},
\code{\link{computes_peak_width_ppm}()},
\code{\link{file_lister}()},
\code{\link{files_to_rDolphin}()},
\code{\link{format.nmr_dataset_1D}()},
\code{\link{get_integration_with_metadata}()},
\code{\link{is.nmr_dataset_1D}()},
\code{\link{load_and_save_functions}},
\code{\link{new_nmr_dataset_1D}()},
\code{\link{nmr_align_find_ref}()},
\code{\link{nmr_align}()},
\code{\link{nmr_baseline_removal}()},
\code{\link{nmr_baseline_threshold}()},
\code{\link{nmr_detect_peaks_plot}()},
\code{\link{nmr_detect_peaks}()},
\code{\link{nmr_exclude_region}()},
\code{\link{nmr_integrate_peak_positions}()},
\code{\link{nmr_integrate_regions}()},
\code{\link{nmr_interpolate_1D}()},
\code{\link{nmr_meta_add}()},
\code{\link{nmr_meta_export}()},
\code{\link{nmr_meta_get_column}()},
\code{\link{nmr_meta_get}()},
\code{\link{nmr_normalize}()},
\code{\link{nmr_pca_build_model}()},
\code{\link{nmr_pca_outliers_filter}()},
\code{\link{nmr_pca_outliers_plot}()},
\code{\link{nmr_pca_outliers_robust}()},
\code{\link{nmr_pca_outliers}()},
\code{\link{nmr_ppm_resolution}()},
\code{\link{plot.nmr_dataset_1D}()},
\code{\link{plot_webgl}()},
\code{\link{print.nmr_dataset_1D}()},
\code{\link{rdCV_PLS_RF_ML}()},
\code{\link{rdCV_PLS_RF}()},
\code{\link{save_files_to_rDolphin}()},
\code{\link{to_ChemoSpec}()},
\code{\link{validate_nmr_dataset}()}
}
\concept{nmr_dataset_1D functions}
\concept{peak detection functions}
