% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_diagnose.R
\name{nmr_diagnose}
\alias{nmr_diagnose}
\alias{nmr_diagnose<-}
\title{Set/Get diagnostic information}
\usage{
nmr_diagnose(nmr_dataset)

nmr_diagnose(nmr_dataset) <- value
}
\arguments{
\item{nmr_dataset}{An \link{nmr_dataset_family} object}

\item{value}{The diagnostic we want to set}
}
\value{
Diagnostic information, usually a list with data or plots
}
\description{
Set or get diagnostic information
}
