% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_diagnose.R
\name{nmr_diagnose}
\alias{nmr_diagnose}
\title{Deprecated function}
\usage{
nmr_diagnose(samples)
}
\arguments{
\item{samples}{An nmr_dataset}
}
\description{
Deprecated function
}
\seealso{
nmr_normalize
}
