% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIHSnmr.R
\name{nmr_exclude_region}
\alias{nmr_exclude_region}
\title{Exclude region from samples}
\usage{
nmr_exclude_region(samples, exclude = list(water = c(4.7, 5)))
}
\arguments{
\item{samples}{A \link{nmr_dataset} object}

\item{exclude}{A list with regions to be zeroed. Typically:
\code{exclude = list(water = c(4.7, 5.0))}}
}
\value{
The \link{nmr_dataset} object, with the regions excluded
}
\description{
Sets to zero a given region (for instance to remove the water peak)
}
