% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_exclude_region.R
\name{nmr_exclude_region}
\alias{nmr_exclude_region}
\alias{nmr_exclude_region.nmr_dataset}
\alias{nmr_exclude_region.nmr_dataset_1D}
\title{Exclude region from samples}
\usage{
nmr_exclude_region(samples, exclude = list(water = c(4.7, 5)))

\method{nmr_exclude_region}{nmr_dataset}(samples, exclude = list(water =
  c(4.7, 5)))

\method{nmr_exclude_region}{nmr_dataset_1D}(samples, exclude = list(water
  = c(4.7, 5)))
}
\arguments{
\item{samples}{A \link{nmr_dataset} or \link{nmr_dataset_1D} object}

\item{exclude}{A list with regions to be zeroed. Typically:
\code{exclude = list(water = c(4.7, 5.0))}}
}
\value{
The same object, with the regions excluded
}
\description{
Excludes a given region (for instance to remove the water peak)
}
