% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nihs_specific.R
\name{nmr_get_irods_meta}
\alias{nmr_get_irods_meta}
\title{Get metadata from an nmr_dataset for irods}
\usage{
nmr_get_irods_meta(nmr_data)
}
\arguments{
\item{nmr_data}{The \code{\link{nmr_dataset}} object}
}
\value{
A data frame (tibble) with the columns used by NIHSrods
}
\description{
Get metadata from an nmr_dataset for irods
}

