% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_get_metadata.R
\name{nmr_get_metadata}
\alias{nmr_get_metadata}
\title{Get metadata}
\usage{
nmr_get_metadata(samples, columns = NULL)
}
\arguments{
\item{samples}{a \link{nmr_dataset} object}

\item{columns}{Columns to get. By default gets all the columns.}
}
\value{
a data frame with the injection metadata
}
\description{
Get metadata
}
