% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_integrate_regions.R
\name{nmr_integrate_peak_positions}
\alias{nmr_integrate_peak_positions}
\title{Integrate peak positions}
\usage{
nmr_integrate_peak_positions(samples, peak_pos_ppm, peak_width_ppm, ...)
}
\arguments{
\item{samples}{A \link{nmr_dataset} object}

\item{peak_pos_ppm}{The peak positions, in ppm}

\item{peak_width_ppm}{The peak widths (or a single peak width for all peaks)}

\item{...}{Arguments passed on to \code{nmr_integrate_regions}
\describe{
  \item{samples}{A \link{nmr_dataset} object}
  \item{regions}{A named list. Each element of the list is a region,
given as a named numeric vector of length two with the range
to integrate. The name of the region will be the name of the
column}
}}
}
\value{
An \link{nmr_dataset_peak_table} object
}
\description{
Integrate peak positions
}
\seealso{
Other peak integration functions: \code{\link{nmr_integrate_regions}},
  \code{\link{pipe_peak_integration}}

Other nmr_dataset_1D functions: \code{\link{[.nmr_dataset_1D}},
  \code{\link{is.nmr_dataset_1D}},
  \code{\link{load_and_save_functions}},
  \code{\link{new_nmr_dataset_1D}},
  \code{\link{nmr_align_find_ref}},
  \code{\link{nmr_align}},
  \code{\link{nmr_detect_peaks_plot}},
  \code{\link{nmr_detect_peaks}},
  \code{\link{nmr_exclude_region}},
  \code{\link{nmr_integrate_regions}},
  \code{\link{nmr_interpolate_1D}},
  \code{\link{nmr_meta_add}},
  \code{\link{nmr_meta_export}},
  \code{\link{nmr_meta_get_column}},
  \code{\link{nmr_meta_get}}, \code{\link{nmr_normalize}},
  \code{\link{nmr_pca_build_model}},
  \code{\link{nmr_pca_outliers_filter}},
  \code{\link{nmr_pca_outliers_plot}},
  \code{\link{nmr_pca_outliers_robust}},
  \code{\link{nmr_pca_outliers}},
  \code{\link{nmr_ppm_resolution}},
  \code{\link{plot.nmr_dataset_1D}},
  \code{\link{plot_webgl}}, \code{\link{to_ASICS}},
  \code{\link{to_ChemoSpec}},
  \code{\link{validate_nmr_dataset_1D}}
}
\concept{nmr_dataset_1D functions}
\concept{peak integration functions}
