% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_integrate_regions.R
\name{nmr_integrate_peak_positions}
\alias{nmr_integrate_peak_positions}
\title{Integrate peak positions}
\usage{
nmr_integrate_peak_positions(samples, peak_pos_ppm, peak_width_ppm,
  fix_baseline = TRUE)
}
\arguments{
\item{samples}{A \link{nmr_dataset} object}

\item{peak_pos_ppm}{The peak positions, in ppm}

\item{peak_width_ppm}{The peak widths (or a single peak width for all peaks)}

\item{fix_baseline}{A logical. If \code{TRUE} it removes the baseline. See details
below

The integration is very naïve and consists of the sum of the intensities
in the given ppm range.

If \code{fix_baseline} is \code{TRUE}, then the region boundaries are used to estimate
a baseline. The baseline is estimated "connecting the boundaries with a straight
line". Only when the spectrum is above the baseline the area is integrated
(negative contributions due to the baseline estimation are ignored).}
}
\value{
A data frame with the NMRExperiment column and one additional column
for each peak position.
}
\description{
Integrate peak positions
}
