% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_integrate_regions.R
\name{nmr_integrate_peak_positions}
\alias{nmr_integrate_peak_positions}
\title{Integrate peak positions}
\usage{
nmr_integrate_peak_positions(
  samples,
  peak_pos_ppm,
  peak_width_ppm = 0.006,
  ...
)
}
\arguments{
\item{samples}{A \link{nmr_dataset} object}

\item{peak_pos_ppm}{The peak positions, in ppm}

\item{peak_width_ppm}{The peak widths (or a single peak width for all peaks)}

\item{...}{
  Arguments passed on to \code{\link[=nmr_integrate_regions]{nmr_integrate_regions}}
  \describe{
    \item{\code{regions}}{A named list. Each element of the list is a region,
given as a named numeric vector of length two with the range
to integrate. The name of the region will be the name of the
column}
  }}
}
\value{
Integrate peak positions
}
\description{
The function allows the integration of a given ppm vector with a specific width.
}
\examples{
dir_to_demo_dataset <- system.file("dataset-demo", package = "AlpsNMR")
nmr_dataset <- nmr_read_samples_dir(dir_to_demo_dataset)
nmr_dataset <- nmr_interpolate_1D(nmr_dataset, axis = c(min = -0.5, max = 10, by = 2.3E-4))

# 1.Peak detection in the dataset.
peak_data <- nmr_detect_peaks(nmr_dataset,
                              nDivRange_ppm = 0.1, # Size of detection segments
                              scales = seq(1, 16, 2),
                              baselineThresh = 0, # Minimum peak intensity
                              SNR.Th = 4) # Signal to noise ratio

# 2.Find the reference spectrum to align with.
NMRExp_ref <- nmr_align_find_ref(nmr_dataset, peak_data)

# 3.Spectra alignment using the ref spectrum and a maximum alignment shift
nmr_dataset <- nmr_align(nmr_dataset, # the dataset
                         peak_data, # detected peaks
                         NMRExp_ref = NMRExp_ref, # ref spectrum
                         maxShift_ppm = 0.0015, # max alignment shift
                         acceptLostPeak = FALSE) # lost peaks

# 4.PEAK INTEGRATION (please, consider previous normalization step).
# First we take the peak table from the reference spectrum
peak_data_ref <- filter(peak_data, NMRExperiment == NMRExp_ref)

# Then we integrate spectra considering the peaks from the ref spectrum
nmr_peak_table <- nmr_integrate_peak_positions(
                      samples = nmr_dataset,
                      peak_pos_ppm = peak_data_ref$ppm,
                      peak_width_ppm = NULL)

#If you wanted the final peak table before machine learning you can run
nmr_peak_table_completed <- get_integration_with_metadata(nmr_peak_table)

}
\seealso{
Other peak integration functions: 
\code{\link{computes_peak_width_ppm}()},
\code{\link{get_integration_with_metada}},
\code{\link{nmr_identify_regions_blood}()},
\code{\link{nmr_identify_regions_cell}()},
\code{\link{nmr_identify_regions_urine}()},
\code{\link{nmr_integrate_regions}()},
\code{\link{pipe_peak_integration}()}

Other nmr_dataset_1D functions: 
\code{\link{[.nmr_dataset_1D}()},
\code{\link{computes_peak_width_ppm}()},
\code{\link{file_lister}()},
\code{\link{files_to_rDolphin_blood}()},
\code{\link{files_to_rDolphin_cell}()},
\code{\link{files_to_rDolphin_urine}()},
\code{\link{format.nmr_dataset_1D}()},
\code{\link{get_integration_with_metada}},
\code{\link{is.nmr_dataset_1D}()},
\code{\link{load_and_save_functions}},
\code{\link{new_nmr_dataset_1D}()},
\code{\link{nmr_align_find_ref}()},
\code{\link{nmr_align}()},
\code{\link{nmr_baseline_removal}()},
\code{\link{nmr_baseline_threshold}()},
\code{\link{nmr_detect_peaks_plot}()},
\code{\link{nmr_detect_peaks_tune_snr}()},
\code{\link{nmr_detect_peaks}()},
\code{\link{nmr_exclude_region}()},
\code{\link{nmr_integrate_regions}()},
\code{\link{nmr_interpolate_1D}()},
\code{\link{nmr_meta_add}()},
\code{\link{nmr_meta_export}()},
\code{\link{nmr_meta_get_column}()},
\code{\link{nmr_meta_get}()},
\code{\link{nmr_normalize}()},
\code{\link{nmr_pca_build_model}()},
\code{\link{nmr_pca_outliers_filter}()},
\code{\link{nmr_pca_outliers_plot}()},
\code{\link{nmr_pca_outliers_robust}()},
\code{\link{nmr_pca_outliers}()},
\code{\link{nmr_ppm_resolution}()},
\code{\link{plot.nmr_dataset_1D}()},
\code{\link{plot_webgl}()},
\code{\link{print.nmr_dataset_1D}()},
\code{\link{save_files_to_rDolphin}()},
\code{\link{to_ASICS}()},
\code{\link{to_ChemoSpec}()},
\code{\link{validate_nmr_dataset}()}
}
\concept{nmr_dataset_1D functions}
\concept{peak integration functions}
