% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIHSnmr.R
\name{nmr_integrate_regions}
\alias{nmr_integrate_regions}
\title{Integrate regions}
\usage{
nmr_integrate_regions(samples, regions)
}
\arguments{
\item{samples}{A \link{nmr_dataset} object}

\item{regions}{A named list. Each element of the list is a region,
given as a named numeric vector of length two with the range
to integrate. The name of the region will be the name of the
column

The integration is very naïve and consists of the sum of the intensities
in the given ppm range.}
}
\value{
A data frame with the NMRExperiment column and one additional column
for each given region.
}
\description{
Integrate given regions and return a data frame with them
}
