% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_interpolate.R
\name{nmr_interpolate_1D}
\alias{nmr_interpolate_1D}
\alias{nmr_interpolate_1D.nmr_dataset}
\title{Interpolate a set of 1D NMR Spectra}
\usage{
nmr_interpolate_1D(samples, axis = c(min = 0.2, max = 10, by = 8e-04))

\method{nmr_interpolate_1D}{nmr_dataset}(samples, axis = c(min = 0.2, max
  = 10, by = 8e-04))
}
\arguments{
\item{samples}{An NMR dataset}

\item{axis}{The ppm axis range and optionally the ppm step}
}
\description{
Interpolate a set of 1D NMR Spectra
}
\seealso{
Other nmr_dataset functions: \code{\link{[.nmr_dataset}},
  \code{\link{load_and_save_functions}},
  \code{\link{new_nmr_dataset}},
  \code{\link{nmr_meta_add}},
  \code{\link{nmr_meta_export}},
  \code{\link{nmr_meta_get_column}},
  \code{\link{nmr_meta_get}},
  \code{\link{nmr_ppm_resolution}},
  \code{\link{nmr_read_samples}},
  \code{\link{validate_nmr_dataset}}

Other nmr_dataset_1D functions: \code{\link{[.nmr_dataset_1D}},
  \code{\link{is.nmr_dataset_1D}},
  \code{\link{load_and_save_functions}},
  \code{\link{new_nmr_dataset_1D}},
  \code{\link{nmr_align_find_ref}},
  \code{\link{nmr_align}},
  \code{\link{nmr_detect_peaks_plot}},
  \code{\link{nmr_detect_peaks_tune_snr}},
  \code{\link{nmr_detect_peaks}},
  \code{\link{nmr_exclude_region}},
  \code{\link{nmr_integrate_peak_positions}},
  \code{\link{nmr_integrate_regions}},
  \code{\link{nmr_meta_add}},
  \code{\link{nmr_meta_export}},
  \code{\link{nmr_meta_get_column}},
  \code{\link{nmr_meta_get}}, \code{\link{nmr_normalize}},
  \code{\link{nmr_pca_build_model}},
  \code{\link{nmr_pca_outliers_filter}},
  \code{\link{nmr_pca_outliers_plot}},
  \code{\link{nmr_pca_outliers_robust}},
  \code{\link{nmr_pca_outliers}},
  \code{\link{nmr_ppm_resolution}},
  \code{\link{plot.nmr_dataset_1D}},
  \code{\link{plot_webgl}}, \code{\link{to_ASICS}},
  \code{\link{to_ChemoSpec}},
  \code{\link{validate_nmr_dataset_1D}}
}
\concept{nmr_dataset functions}
\concept{nmr_dataset_1D functions}
