% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_interpolate.R
\name{nmr_interpolate_1D}
\alias{nmr_interpolate_1D}
\alias{nmr_interpolate_1D.nmr_dataset}
\title{Interpolate a set of 1D NMR Spectra}
\usage{
nmr_interpolate_1D(samples, axis = c(min = 0.2, max = 10, by = 8e-04))

\method{nmr_interpolate_1D}{nmr_dataset}(samples, axis = c(min = 0.2, max = 10, by = 8e-04))
}
\arguments{
\item{samples}{An NMR dataset}

\item{axis}{The ppm axis range and optionally the ppm step. Set it to \code{NULL} for autodetection}
}
\value{
Interpolate a set of 1D NMR Spectra
}
\description{
Interpolate a set of 1D NMR Spectra
}
\examples{
dir_to_demo_dataset <- system.file("dataset-demo", package = "AlpsNMR")
dataset <- nmr_read_samples_dir(dir_to_demo_dataset)
dataset_1D <- nmr_interpolate_1D(dataset, axis = c(min = -0.5, max = 10, by = 2.3E-4))

dir_to_demo_dataset <- system.file("dataset-demo", package = "AlpsNMR")
dataset <- nmr_read_samples_dir(dir_to_demo_dataset)
dataset_1D <- nmr_interpolate_1D(dataset, axis = c(min = -0.5, max = 10, by = 2.3E-4))

}
\seealso{
Other nmr_dataset functions: 
\code{\link{[.nmr_dataset}()},
\code{\link{format.nmr_dataset}()},
\code{\link{load_and_save_functions}},
\code{\link{new_nmr_dataset}()},
\code{\link{nmr_meta_add}()},
\code{\link{nmr_meta_export}()},
\code{\link{nmr_meta_get_column}()},
\code{\link{nmr_meta_get}()},
\code{\link{nmr_ppm_resolution}()},
\code{\link{nmr_read_samples}()},
\code{\link{print.nmr_dataset}()},
\code{\link{validate_nmr_dataset}()}

Other nmr_dataset_1D functions: 
\code{\link{[.nmr_dataset_1D}()},
\code{\link{computes_peak_width_ppm}()},
\code{\link{file_lister}()},
\code{\link{files_to_rDolphin}()},
\code{\link{format.nmr_dataset_1D}()},
\code{\link{get_integration_with_metadata}()},
\code{\link{is.nmr_dataset_1D}()},
\code{\link{load_and_save_functions}},
\code{\link{new_nmr_dataset_1D}()},
\code{\link{nmr_align_find_ref}()},
\code{\link{nmr_align}()},
\code{\link{nmr_baseline_removal}()},
\code{\link{nmr_baseline_threshold}()},
\code{\link{nmr_detect_peaks_plot}()},
\code{\link{nmr_detect_peaks_tune_snr}()},
\code{\link{nmr_detect_peaks}()},
\code{\link{nmr_exclude_region}()},
\code{\link{nmr_integrate_peak_positions}()},
\code{\link{nmr_integrate_regions}()},
\code{\link{nmr_meta_add}()},
\code{\link{nmr_meta_export}()},
\code{\link{nmr_meta_get_column}()},
\code{\link{nmr_meta_get}()},
\code{\link{nmr_normalize}()},
\code{\link{nmr_pca_build_model}()},
\code{\link{nmr_pca_outliers_filter}()},
\code{\link{nmr_pca_outliers_plot}()},
\code{\link{nmr_pca_outliers_robust}()},
\code{\link{nmr_pca_outliers}()},
\code{\link{nmr_ppm_resolution}()},
\code{\link{plot.nmr_dataset_1D}()},
\code{\link{plot_webgl}()},
\code{\link{print.nmr_dataset_1D}()},
\code{\link{rdCV_PLS_RF_ML}()},
\code{\link{rdCV_PLS_RF}()},
\code{\link{save_files_to_rDolphin}()},
\code{\link{to_ChemoSpec}()},
\code{\link{validate_nmr_dataset}()}
}
\concept{nmr_dataset functions}
\concept{nmr_dataset_1D functions}
