% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_meta.R
\name{nmr_meta_add}
\alias{nmr_meta_add}
\title{Add metadata to an nmr_dataset object}
\usage{
nmr_meta_add(nmr_data, metadata, by = "NMRExperiment")
}
\arguments{
\item{nmr_data}{an \link{nmr_dataset} object}

\item{metadata}{A data frame with metadata to add}

\item{by}{A column name of both the \code{nmr_dataset$metadata$external} and the metadata
data.frame. If you want to merge two columns with different headers you can
use a named character vector \code{c("NMRExperiment" = "ExperimentNMR")} where
the left side is the column name of the \code{nmr_dataset$metadata$external} and the right side is
the column name of the metadata data frame.}
}
\value{
The nmr_dataset object with the added metadata
}
\description{
This is useful to add metadata to datasets that can be later used for
plotting spectra or further analysis (PCA...).
}
\examples{
# Load a demo dataset with four samples:
dataset <- system.file("dataset-demo", package = "NIHSnmr")
nmr_dataset <- nmr_read_samples_dir(dataset)

# At first we just have the NMRExperiment column
print(nmr_meta_get(nmr_dataset, groups = "external"))
# Get a table with NMRExperiment -> SubjectID
dummy_metadata <- system.file("dataset-demo", "dummy_metadata.xlsx", package = "NIHSnmr")
NMRExp_SubjID <- readxl::read_excel(dummy_metadata, sheet = 1)

print(NMRExp_SubjID)
# We can link the SubjectID column of the first excel into the dataset
nmr_dataset <- nmr_meta_add(nmr_dataset, NMRExp_SubjID, by = "NMRExperiment")
print(nmr_meta_get(nmr_dataset, groups = "external"))
# The second excel can use the SubjectID:
SubjID_Age <- readxl::read_excel(dummy_metadata, sheet = 2)
print(SubjID_Age)
# Add the metadata by its SubjectID:
nmr_dataset <- nmr_meta_add(nmr_dataset, SubjID_Age, by = "SubjectID")
# The final loaded metadata:
print(nmr_meta_get(nmr_dataset, groups = "external"))

}
\seealso{
Other metadata functions: \code{\link{nmr_meta_export}},
  \code{\link{nmr_meta_get_column}},
  \code{\link{nmr_meta_get}},
  \code{\link{pipe_add_metadata}}

Other metadata management functions: \code{\link{nmr_meta_export}},
  \code{\link{nmr_meta_get_column}},
  \code{\link{nmr_meta_get}}

Other nmr_dataset functions: \code{\link{[.nmr_dataset}},
  \code{\link{load_and_save_functions}},
  \code{\link{new_nmr_dataset}},
  \code{\link{nmr_interpolate_1D}},
  \code{\link{nmr_meta_export}},
  \code{\link{nmr_meta_get_column}},
  \code{\link{nmr_meta_get}},
  \code{\link{nmr_ppm_resolution}},
  \code{\link{nmr_read_samples}},
  \code{\link{validate_nmr_dataset}}

Other nmr_dataset_1D functions: \code{\link{[.nmr_dataset_1D}},
  \code{\link{is.nmr_dataset_1D}},
  \code{\link{load_and_save_functions}},
  \code{\link{new_nmr_dataset_1D}},
  \code{\link{nmr_align_find_ref}},
  \code{\link{nmr_align}},
  \code{\link{nmr_detect_peaks_plot}},
  \code{\link{nmr_detect_peaks}},
  \code{\link{nmr_exclude_region}},
  \code{\link{nmr_integrate_peak_positions}},
  \code{\link{nmr_integrate_regions}},
  \code{\link{nmr_interpolate_1D}},
  \code{\link{nmr_meta_export}},
  \code{\link{nmr_meta_get_column}},
  \code{\link{nmr_meta_get}}, \code{\link{nmr_normalize}},
  \code{\link{nmr_pca_build_model}},
  \code{\link{nmr_pca_outliers_filter}},
  \code{\link{nmr_pca_outliers_plot}},
  \code{\link{nmr_pca_outliers_robust}},
  \code{\link{nmr_pca_outliers}},
  \code{\link{nmr_ppm_resolution}},
  \code{\link{plot.nmr_dataset_1D}},
  \code{\link{plot_webgl}}, \code{\link{to_ASICS}},
  \code{\link{to_ChemoSpec}},
  \code{\link{validate_nmr_dataset_1D}}

Other nmr_dataset_peak_table functions: \code{\link{[.nmr_dataset_peak_table}},
  \code{\link{is.nmr_dataset_peak_table}},
  \code{\link{load_and_save_functions}},
  \code{\link{new_nmr_dataset_peak_table}},
  \code{\link{nmr_meta_export}},
  \code{\link{nmr_meta_get_column}},
  \code{\link{nmr_meta_get}},
  \code{\link{validate_nmr_dataset_peak_table}}

Other import/export functions: \code{\link{load_and_save_functions}},
  \code{\link{nmr_data}}, \code{\link{nmr_meta_export}},
  \code{\link{nmr_read_bruker_fid}},
  \code{\link{nmr_read_samples}},
  \code{\link{nmr_zip_bruker_samples}},
  \code{\link{pipe_load_samples}}, \code{\link{to_ASICS}},
  \code{\link{to_ChemoSpec}}
}
\concept{import/export functions}
\concept{metadata functions}
\concept{metadata management functions}
\concept{nmr_dataset functions}
\concept{nmr_dataset_1D functions}
\concept{nmr_dataset_peak_table functions}
