% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_meta.R
\name{nmr_meta_export}
\alias{nmr_meta_export}
\title{Export Metadata to an Excel file}
\usage{
nmr_meta_export(
  nmr_dataset,
  xlsx_file,
  groups = c("info", "orig", "title", "external")
)
}
\arguments{
\item{nmr_dataset}{An \link{nmr_dataset_family} object}

\item{xlsx_file}{"The .xlsx excel file"}

\item{groups}{A character vector. Use \code{"external"} for the external metadata or
the default for a more generic solution}
}
\value{
The Excel file name
}
\description{
Export Metadata to an Excel file
}
\examples{
dir_to_demo_dataset <- system.file("dataset-demo", package = "AlpsNMR")
dataset <- nmr_read_samples_dir(dir_to_demo_dataset)
#nmr_meta_export(dataset, "metadata.xlsx")

}
\seealso{
Other metadata functions: 
\code{\link{Pipelines}},
\code{\link{nmr_meta_add}()},
\code{\link{nmr_meta_get_column}()},
\code{\link{nmr_meta_get}()}

Other nmr_dataset functions: 
\code{\link{[.nmr_dataset}()},
\code{\link{format.nmr_dataset}()},
\code{\link{load_and_save_functions}},
\code{\link{new_nmr_dataset}()},
\code{\link{nmr_interpolate_1D}()},
\code{\link{nmr_meta_add}()},
\code{\link{nmr_meta_get_column}()},
\code{\link{nmr_meta_get}()},
\code{\link{nmr_ppm_resolution}()},
\code{\link{nmr_read_samples}()},
\code{\link{print.nmr_dataset}()},
\code{\link{validate_nmr_dataset}()}

Other nmr_dataset_1D functions: 
\code{\link{[.nmr_dataset_1D}()},
\code{\link{computes_peak_width_ppm}()},
\code{\link{file_lister}()},
\code{\link{files_to_rDolphin}()},
\code{\link{format.nmr_dataset_1D}()},
\code{\link{get_integration_with_metadata}()},
\code{\link{is.nmr_dataset_1D}()},
\code{\link{load_and_save_functions}},
\code{\link{new_nmr_dataset_1D}()},
\code{\link{nmr_align_find_ref}()},
\code{\link{nmr_align}()},
\code{\link{nmr_baseline_removal}()},
\code{\link{nmr_baseline_threshold}()},
\code{\link{nmr_detect_peaks_plot}()},
\code{\link{nmr_detect_peaks_tune_snr}()},
\code{\link{nmr_detect_peaks}()},
\code{\link{nmr_exclude_region}()},
\code{\link{nmr_integrate_peak_positions}()},
\code{\link{nmr_integrate_regions}()},
\code{\link{nmr_interpolate_1D}()},
\code{\link{nmr_meta_add}()},
\code{\link{nmr_meta_get_column}()},
\code{\link{nmr_meta_get}()},
\code{\link{nmr_normalize}()},
\code{\link{nmr_pca_build_model}()},
\code{\link{nmr_pca_outliers_filter}()},
\code{\link{nmr_pca_outliers_plot}()},
\code{\link{nmr_pca_outliers_robust}()},
\code{\link{nmr_pca_outliers}()},
\code{\link{nmr_ppm_resolution}()},
\code{\link{plot.nmr_dataset_1D}()},
\code{\link{plot_webgl}()},
\code{\link{print.nmr_dataset_1D}()},
\code{\link{rdCV_PLS_RF_ML}()},
\code{\link{rdCV_PLS_RF}()},
\code{\link{save_files_to_rDolphin}()},
\code{\link{to_ChemoSpec}()},
\code{\link{validate_nmr_dataset}()}

Other nmr_dataset_peak_table functions: 
\code{\link{[.nmr_dataset_peak_table}()},
\code{\link{format.nmr_dataset_peak_table}()},
\code{\link{is.nmr_dataset_peak_table}()},
\code{\link{load_and_save_functions}},
\code{\link{new_nmr_dataset_peak_table}()},
\code{\link{nmr_meta_add}()},
\code{\link{nmr_meta_get_column}()},
\code{\link{nmr_meta_get}()},
\code{\link{print.nmr_dataset_peak_table}()},
\code{\link{validate_nmr_dataset_peak_table}()}

Other import/export functions: 
\code{\link{Pipelines}},
\code{\link{files_to_rDolphin}()},
\code{\link{load_and_save_functions}},
\code{\link{nmr_data}()},
\code{\link{nmr_read_bruker_fid}()},
\code{\link{nmr_read_samples}()},
\code{\link{nmr_zip_bruker_samples}()},
\code{\link{save_files_to_rDolphin}()},
\code{\link{save_profiling_output}()},
\code{\link{to_ChemoSpec}()}
}
\concept{import/export functions}
\concept{metadata functions}
\concept{nmr_dataset functions}
\concept{nmr_dataset_1D functions}
\concept{nmr_dataset_peak_table functions}
