% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_get_metadata.R
\name{nmr_meta_get}
\alias{nmr_meta_get}
\title{Get metadata}
\usage{
nmr_meta_get(samples, columns = NULL, groups = NULL)
}
\arguments{
\item{samples}{a \link{nmr_dataset} object}

\item{columns}{Columns to get. By default gets all the columns.}

\item{groups}{Groups to get. Groups are predefined of columns. Typically
\code{"external"} for metadata added with \link{nmr_add_metadata}.

Both \code{groups} and \code{columns} can't be given simultaneously.}
}
\value{
a data frame with the injection metadata
}
\description{
Get metadata
}
