% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_meta.R
\name{nmr_meta_get}
\alias{nmr_meta_get}
\title{Get metadata}
\usage{
nmr_meta_get(samples, columns = NULL, groups = NULL)
}
\arguments{
\item{samples}{a \link{nmr_dataset} object}

\item{columns}{Columns to get. By default gets all the columns.}

\item{groups}{Groups to get. Groups are predefined of columns. Typically
\code{"external"} for metadata added with \link{nmr_meta_add}.

Both \code{groups} and \code{columns} can't be given simultaneously.}
}
\value{
a data frame with the injection metadata
}
\description{
Get metadata
}
\seealso{
Other metadata functions: \code{\link{nmr_meta_add}},
  \code{\link{nmr_meta_export}},
  \code{\link{nmr_meta_get_column}},
  \code{\link{pipe_add_metadata}}

Other metadata management functions: \code{\link{nmr_meta_add}},
  \code{\link{nmr_meta_export}},
  \code{\link{nmr_meta_get_column}}

Other nmr_dataset functions: \code{\link{[.nmr_dataset}},
  \code{\link{load_and_save_functions}},
  \code{\link{new_nmr_dataset}},
  \code{\link{nmr_interpolate_1D}},
  \code{\link{nmr_meta_add}},
  \code{\link{nmr_meta_export}},
  \code{\link{nmr_meta_get_column}},
  \code{\link{nmr_ppm_resolution}},
  \code{\link{nmr_read_samples}},
  \code{\link{validate_nmr_dataset}}

Other nmr_dataset_1D functions: \code{\link{[.nmr_dataset_1D}},
  \code{\link{is.nmr_dataset_1D}},
  \code{\link{load_and_save_functions}},
  \code{\link{new_nmr_dataset_1D}},
  \code{\link{nmr_align_find_ref}},
  \code{\link{nmr_align}},
  \code{\link{nmr_detect_peaks_plot}},
  \code{\link{nmr_detect_peaks}},
  \code{\link{nmr_exclude_region}},
  \code{\link{nmr_integrate_peak_positions}},
  \code{\link{nmr_integrate_regions}},
  \code{\link{nmr_interpolate_1D}},
  \code{\link{nmr_meta_add}},
  \code{\link{nmr_meta_export}},
  \code{\link{nmr_meta_get_column}},
  \code{\link{nmr_normalize}},
  \code{\link{nmr_pca_build_model}},
  \code{\link{nmr_pca_outliers_filter}},
  \code{\link{nmr_pca_outliers_plot}},
  \code{\link{nmr_pca_outliers_robust}},
  \code{\link{nmr_pca_outliers}},
  \code{\link{nmr_ppm_resolution}},
  \code{\link{plot.nmr_dataset_1D}},
  \code{\link{plot_webgl}}, \code{\link{to_ASICS}},
  \code{\link{to_ChemoSpec}},
  \code{\link{validate_nmr_dataset_1D}}

Other nmr_dataset_peak_table functions: \code{\link{[.nmr_dataset_peak_table}},
  \code{\link{is.nmr_dataset_peak_table}},
  \code{\link{load_and_save_functions}},
  \code{\link{new_nmr_dataset_peak_table}},
  \code{\link{nmr_meta_add}},
  \code{\link{nmr_meta_export}},
  \code{\link{nmr_meta_get_column}},
  \code{\link{validate_nmr_dataset_peak_table}}
}
\concept{metadata functions}
\concept{metadata management functions}
\concept{nmr_dataset functions}
\concept{nmr_dataset_1D functions}
\concept{nmr_dataset_peak_table functions}
