% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_meta.R
\name{nmr_meta_get_column}
\alias{nmr_meta_get_column}
\title{Get a single metadata column}
\usage{
nmr_meta_get_column(samples, column = "NMRExperiment")
}
\arguments{
\item{samples}{a \link{nmr_dataset} object}

\item{column}{A column to get}
}
\value{
A vector with the column
}
\description{
Get a single metadata column
}
\seealso{
Other metadata functions: \code{\link{nmr_meta_add}},
  \code{\link{nmr_meta_export}}, \code{\link{nmr_meta_get}}
}
\concept{metadata functions}
