% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_meta.R
\name{nmr_meta_groups}
\alias{nmr_meta_groups}
\title{Get the names of metadata groups}
\usage{
nmr_meta_groups(samples)
}
\arguments{
\item{samples}{a \link{nmr_dataset_family} object}
}
\value{
A character vector with group names
}
\description{
Get the names of metadata groups
}
\examples{
dir_to_demo_dataset <- system.file("dataset-demo", package = "AlpsNMR")
dataset <- nmr_read_samples_dir(dir_to_demo_dataset)
metadata_column <- nmr_meta_get_column(dataset)
}
\seealso{
Other metadata functions: 
\code{\link{Pipelines}},
\code{\link{nmr_meta_add}()},
\code{\link{nmr_meta_export}()},
\code{\link{nmr_meta_get_column}()},
\code{\link{nmr_meta_get}()}
}
\concept{metadata functions}
