% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_normalize.R
\name{nmr_normalize}
\alias{nmr_normalize}
\alias{nmr_normalize.nmr_dataset_1D}
\title{Normalize NMR samples}
\usage{
nmr_normalize(samples, method = c("area", "max", "value", "region",
  "pqn", "none"), values = NULL)

\method{nmr_normalize}{nmr_dataset_1D}(samples, method = c("area", "max",
  "value", "region", "pqn", "none"), values = NULL)
}
\arguments{
\item{samples}{A \link{nmr_dataset} object}

\item{method}{The criteria to be used for normalization}

\item{values}{If \code{method == "value"} then values is a list
with the normalization values. The list must be named as the
data fields to normalize. Typically would be something like:
\code{values = list(data_1r = c(val1, val2, val3))}.
If \code{method == "region"} then values is the chemical shift
region to integrate.}
}
\value{
The \link{nmr_dataset} object, with the samples normalized
}
\description{
Normalize NMR samples
}
\seealso{
Other nmr_dataset_1D functions: \code{\link{[.nmr_dataset_1D}},
  \code{\link{is.nmr_dataset_1D}},
  \code{\link{load_and_save_functions}},
  \code{\link{new_nmr_dataset_1D}},
  \code{\link{nmr_align_find_ref}},
  \code{\link{nmr_align}},
  \code{\link{nmr_detect_peaks_plot}},
  \code{\link{nmr_detect_peaks}},
  \code{\link{nmr_exclude_region}},
  \code{\link{nmr_integrate_peak_positions}},
  \code{\link{nmr_integrate_regions}},
  \code{\link{nmr_interpolate_1D}},
  \code{\link{nmr_meta_add}},
  \code{\link{nmr_meta_export}},
  \code{\link{nmr_meta_get_column}},
  \code{\link{nmr_meta_get}},
  \code{\link{nmr_pca_build_model}},
  \code{\link{nmr_pca_outliers_filter}},
  \code{\link{nmr_pca_outliers_plot}},
  \code{\link{nmr_pca_outliers_robust}},
  \code{\link{nmr_pca_outliers}},
  \code{\link{nmr_ppm_resolution}},
  \code{\link{plot.nmr_dataset_1D}},
  \code{\link{plot_webgl}}, \code{\link{to_ASICS}},
  \code{\link{to_ChemoSpec}},
  \code{\link{validate_nmr_dataset_1D}}
}
\concept{nmr_dataset_1D functions}
