% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_normalize.R
\name{nmr_normalize}
\alias{nmr_normalize}
\alias{nmr_normalize.nmr_dataset}
\alias{nmr_normalize.nmr_dataset_1D}
\title{Normalize NMR samples}
\usage{
nmr_normalize(samples, method = c("area", "max", "value", "region",
  "pqn", "none"), values = NULL)

\method{nmr_normalize}{nmr_dataset}(samples, method = c("area", "max",
  "value", "region", "pqn", "none"), values = NULL)

\method{nmr_normalize}{nmr_dataset_1D}(samples, method = c("area", "max",
  "value", "region", "pqn", "none"), values = NULL)
}
\arguments{
\item{samples}{A \link{nmr_dataset} object}

\item{method}{The criteria to be used for normalization}

\item{values}{If \code{method == "value"} then values is a list
with the normalization values. The list must be named as the
data fields to normalize. Typically would be something like:
\code{values = list(data_1r = c(val1, val2, val3))}.
If \code{method == "region"} then values is the chemical shift
region to integrate.}
}
\value{
The \link{nmr_dataset} object, with the samples normalized
}
\description{
Normalize NMR samples
}
