% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_helpers.R
\name{nmr_pca_build_model}
\alias{nmr_pca_build_model}
\title{Build a PCA on for an nmr_dataset}
\usage{
nmr_pca_build_model(nmr_data, ncomp = NULL, center = TRUE, scale = FALSE,
  ...)
}
\arguments{
\item{nmr_data}{a \code{\link{nmr_dataset}} object}

\item{ncomp}{integer, if data is complete \code{ncomp} decides the number of components and associated
    eigenvalues to display from the \code{pcasvd} algorithm and if the data has missing values, 
    \code{ncomp} gives the number of components to keep to perform the reconstitution of the data using
	the NIPALS algorithm. If \code{NULL}, function sets \code{ncomp = min(nrow(X), ncol(X))}}

\item{center}{a logical value indicating whether the variables should be shifted to be zero centered.
    Alternately, a vector of length equal the number of columns of \code{X} can be supplied. 
    The value is passed to \code{\link{scale}}.}

\item{scale}{a logical value indicating whether the variables should be scaled to have
    unit variance before the analysis takes place. The default is \code{FALSE} for consistency with \code{prcomp} 
    function, but in general scaling is advisable. Alternatively, a vector of length equal the number of 
    columns of \code{X} can be supplied. The value is passed to \code{\link{scale}}.}

\item{...}{Additional arguments passed on to \link[mixOmics:pca]{mixOmics::pca}}
}
\value{
A PCA model as given by \link[mixOmics:pca]{mixOmics::pca} with two additional attributes:
\itemize{
\item \code{nmr_data_axis} containing the full ppm axis
\item \code{nmr_included} with the data points included in the model
These attributes are used internally by NIHSnmr to create loading plots
}
}
\description{
This function builds a PCA model with all the NMR spectra. Regions with
zero values (excluded regions) or near-zero variance regions are automatically
excluded from the analysis.
}
