% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_helpers.R
\name{nmr_pca_outliers_filter}
\alias{nmr_pca_outliers_filter}
\title{Exclude outliers}
\usage{
nmr_pca_outliers_filter(nmr_dataset, pca_outliers)
}
\arguments{
\item{nmr_dataset}{An \link{nmr_dataset_1D} object}

\item{pca_outliers}{The output from \code{\link[=nmr_pca_outliers]{nmr_pca_outliers()}}}
}
\value{
An \link{nmr_dataset_1D} without the detected outliers
}
\description{
Exclude outliers
}
