% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_helpers.R
\name{nmr_pca_outliers_plot}
\alias{nmr_pca_outliers_plot}
\title{Plot for outlier detection diagnostic}
\usage{
nmr_pca_outliers_plot(nmr_dataset, pca_outliers, ...)
}
\arguments{
\item{nmr_dataset}{An \link{nmr_dataset_1D} object}

\item{pca_outliers}{The output from \code{\link[=nmr_pca_outliers]{nmr_pca_outliers()}}}

\item{...}{Additional parameters passed on to \code{\link[ggplot2:aes_string]{ggplot2::aes_string()}}}
}
\value{
A plot for the outlier detection
}
\description{
Plot for outlier detection diagnostic
}
