% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_helpers.R
\name{nmr_pca_plots}
\alias{nmr_pca_plots}
\alias{nmr_pca_plot_variance}
\alias{nmr_pca_scoreplot}
\alias{nmr_pca_loadingplot}
\title{Plotting functions for PCA}
\usage{
nmr_pca_plot_variance(pca_model)

nmr_pca_scoreplot(nmr_data, pca_model, comp = 1:2, ...)

nmr_pca_loadingplot(pca_model, comp)
}
\arguments{
\item{pca_model}{A PCA model trained with \link{nmr_pca_build_model}}

\item{nmr_data}{a \code{\link{nmr_dataset}} object}

\item{comp}{Components to represent}

\item{...}{Additional aesthetics passed on to \link[ggplot2:aes]{ggplot2::aes} (use bare unquoted names)}
}
\description{
Plotting functions for PCA
}
