% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_peak_clustering.R
\name{nmr_peak_clustering}
\alias{nmr_peak_clustering}
\title{Peak clustering}
\usage{
nmr_peak_clustering(peak_data, peak2peak_dist = NULL)
}
\arguments{
\item{peak_data}{The peak list}

\item{peak2peak_dist}{The distances obtained with \link{nmr_get_peak_distances}.
If NULL it is computed from \code{peak_data}}
}
\value{
A list including: The \code{peak_data} with an
additional "cluster" column, cluster (the hierarchical cluster),
num_clusters (an estimation of the number of clusters)
}
\description{
Peak clustering
}
