% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_peak_clustering.R
\name{nmr_peak_clustering_plot}
\alias{nmr_peak_clustering_plot}
\title{Plot clustering results}
\usage{
nmr_peak_clustering_plot(
  dataset,
  peak_list_clustered,
  NMRExperiments,
  chemshift_range
)
}
\arguments{
\item{dataset}{The \link{nmr_dataset_1D} object}

\item{peak_list_clustered}{A peak list table with a clustered column}

\item{NMRExperiments}{Two and only two experiments to compare in the plot}

\item{chemshift_range}{A region, make it so it does not cover a huge range (maybe 1ppm or less)}
}
\value{
A plot of the two experiments in the given chemshift range, with
lines connecting peaks identified as the same and dots showing peaks without pairs
}
\description{
Plot clustering results
}
