% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_detect_peaks_align.R
\name{nmr_ppm_resolution}
\alias{nmr_ppm_resolution}
\alias{nmr_ppm_resolution.nmr_dataset}
\alias{nmr_ppm_resolution.nmr_dataset_1D}
\title{Get the ppm resolution of the dataset}
\usage{
nmr_ppm_resolution(nmr_dataset)

\method{nmr_ppm_resolution}{nmr_dataset}(nmr_dataset)

\method{nmr_ppm_resolution}{nmr_dataset_1D}(nmr_dataset)
}
\arguments{
\item{nmr_dataset}{An object containing NMR samples}
}
\value{
A number (the ppm resolution, measured in ppms)
}
\description{
Get the ppm resolution of the dataset
}
