% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nihs_specific.R
\name{nmr_prepare_zip_files}
\alias{nmr_prepare_zip_files}
\title{Create one zip file for each sample}
\usage{
nmr_prepare_zip_files(meta_irods, workdir, overwrite = FALSE, ...)
}
\arguments{
\item{meta_irods}{Data frame given by \code{\link{nmr_get_irods_meta}}}

\item{workdir}{Directory to store zip files}

\item{overwrite}{Should existing zip files be overwritten?}

\item{...}{Passed to \code{\link[utils]{zip}}}
}
\value{
A meta_irods data frame with an additional \code{src_path} column
        containing the zip file path location.
}
\description{
Create one zip file for each sample
}

