% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset.R
\name{nmr_read_samples}
\alias{nmr_read_samples}
\alias{nmr_read_samples_dir}
\title{Read NMR samples}
\usage{
nmr_read_samples_dir(samples_dir, pulse_sequence = NULL,
  metadata_only = FALSE, ...)

nmr_read_samples(sample_names, pulse_sequence = NULL,
  metadata_only = FALSE, ...)
}
\arguments{
\item{samples_dir}{A directory that contains multiple samples}

\item{pulse_sequence}{If it is set to a pulse sequence
("NOESY", "JRES", "CPMG"...) it will only load
the samples that match that pulse sequence.}

\item{metadata_only}{A logical, to load only metadata (default: \code{FALSE})}

\item{...}{Arguments passed to \code{\link[=read_bruker_sample]{read_bruker_sample()}} for data loading}

\item{sample_names}{A character vector with file or directory names.}
}
\value{
a \link{nmr_dataset} object
}
\description{
These functions load samples from files and return a \link{nmr_dataset}.
}
