% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset.R
\name{nmr_read_samples}
\alias{nmr_read_samples}
\alias{nmr_read_samples_dir}
\title{Read NMR samples}
\usage{
nmr_read_samples_dir(
  samples_dir,
  format = "bruker",
  pulse_sequence = NULL,
  metadata_only = FALSE,
  ...
)

nmr_read_samples(
  sample_names,
  format = "bruker",
  pulse_sequence = NULL,
  metadata_only = FALSE,
  ...
)
}
\arguments{
\item{samples_dir}{A directory that contains multiple samples}

\item{format}{Either "bruker" or "jdx"}

\item{pulse_sequence}{If it is set to a pulse sequence
("NOESY", "JRES", "CPMG"...) it will only load
the samples that match that pulse sequence.}

\item{metadata_only}{A logical, to load only metadata (default: \code{FALSE})}

\item{...}{Arguments passed to \code{\link[=read_bruker_sample]{read_bruker_sample()}} for data loading}

\item{sample_names}{A character vector with file or directory names.}
}
\value{
a \link{nmr_dataset} object
}
\description{
These functions load samples from files and return a \link{nmr_dataset}.
}
\examples{
dir_to_demo_dataset <- system.file("dataset-demo", package = "AlpsNMR")
dataset <- nmr_read_samples_dir(dir_to_demo_dataset)

dir_to_demo_dataset <- system.file("dataset-demo", package = "AlpsNMR")
zip_files <- fs::dir_ls(dir_to_demo_dataset, glob = "*.zip")
dataset <- nmr_read_samples(sample_names = zip_files)

}
\seealso{
Other nmr_dataset functions: 
\code{\link{[.nmr_dataset}()},
\code{\link{format.nmr_dataset}()},
\code{\link{load_and_save_functions}},
\code{\link{new_nmr_dataset}()},
\code{\link{nmr_interpolate_1D}()},
\code{\link{nmr_meta_add}()},
\code{\link{nmr_meta_export}()},
\code{\link{nmr_meta_get_column}()},
\code{\link{nmr_meta_get}()},
\code{\link{nmr_ppm_resolution}()},
\code{\link{print.nmr_dataset}()},
\code{\link{validate_nmr_dataset}()}

Other import/export functions: 
\code{\link{Pipelines}},
\code{\link{files_to_rDolphin}()},
\code{\link{load_and_save_functions}},
\code{\link{nmr_data}()},
\code{\link{nmr_meta_export}()},
\code{\link{nmr_read_bruker_fid}()},
\code{\link{nmr_zip_bruker_samples}()},
\code{\link{save_files_to_rDolphin}()},
\code{\link{save_profiling_output}()},
\code{\link{to_ChemoSpec}()}
}
\concept{import/export functions}
\concept{nmr_dataset functions}
