% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruker.R
\name{nmr_zip_bruker_samples}
\alias{nmr_zip_bruker_samples}
\title{Create one zip file for each brucker sample path}
\usage{
nmr_zip_bruker_samples(path, workdir, overwrite = FALSE, ...)
}
\arguments{
\item{path}{Character vector with sample directories}

\item{workdir}{Directory to store zip files}

\item{overwrite}{Should existing zip files be overwritten?}

\item{...}{Passed to \code{\link[utils]{zip}}}
}
\value{
A character vector of the same length as path, with the zip file names
}
\description{
Create one zip file for each brucker sample path
}
