% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_rDolphin_blood.R
\docType{data}
\name{Parameters_blood}
\alias{Parameters_blood}
\title{to rDolphin}
\description{
Parameters for blood (plasma/serum) samples profiling
}
\details{
The template \code{Parameters_blood} contains the chosen normalization approach (by default, PQN), the Spectometer Frequency (by default, 600.04MHz),
alignment (by default, TSP 0.00 ppm), bucket resolution (by default, 0.00023)
}
\references{
\url{github.com/danielcanueto/rDolphin}
}
\keyword{data}
