% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_data_analysis.R
\name{permutation_test_model}
\alias{permutation_test_model}
\title{Permutation test}
\usage{
permutation_test_model(
  dataset,
  y_column,
  identity_column,
  external_val,
  internal_val,
  data_analysis_method,
  nPerm = 50,
  nRep,
  nOuter,
  varRatio,
  parallel
)
}
\value{
A permutation matrix with permuted values
}
\description{
Make permutations with data and default settings from an nmr_data_analysis_method
}
\examples{
\dontrun{
P = permutation_test_model(model)
}
}
