% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier_MUVR.R
\name{permutation_test_model}
\alias{permutation_test_model}
\title{Permutation test on MUVR}
\usage{
permutation_test_model(MVObj, nPerm = 50, nRep, nOuter, varRatio,
  parallel)
}
\arguments{
\item{MVObj}{A MUVR obvject}

\item{nPerm}{number of permutations to run}

\item{nRep}{number of repetitions for each permutation (defaults to value of actual model)}

\item{nOuter}{number of outer validation segments for each permutation (defaults to value of actual model)}

\item{varRatio}{varRatio for each permutation (defaults to value of actual model)}

\item{parallel}{whether to run calculations using parallel processing - which requires registered backend (defaults to value of actual model)}
}
\value{
A permutation matrix with permuted values
}
\description{
Make permutations with data and default settings from an actual MUVR object
}
\examples{
\dontrun{
P = permutation_test_model(model)
}
}
