% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier_MUVR.R
\name{permutation_test_plot}
\alias{permutation_test_plot}
\title{Permutation test plot}
\usage{
permutation_test_plot(MVObj, permMatrix, model = "mid", type = type,
  pos, xlab = NULL, xlim, ylim = NULL, breaks = "Sturges",
  main = NULL)
}
\arguments{
\item{MVObj}{A MUVR object}

\item{permMatrix}{A permutation fitness outcome from MUVR::permutations()}

\item{model}{'Min', 'Mid', or 'Max'}

\item{type}{'t' (default; for Student's t) or 'non' for "non-parametric" (i.e. rank) studen'ts}

\item{pos}{which side of actual to put p-value on}

\item{xlab}{optional xlabel}

\item{xlim}{optional x-range}

\item{ylim}{otional y-range}

\item{breaks}{optional custom histogram breaks (defaults to 'sturges')}

\item{main}{optional plot title (or TRUE for autoname)}
}
\value{
A plot with the comparison between the actual model versus the permuted models
}
\description{
Plot permutation test using actual model and permutated models
}
\examples{
\dontrun{
P = permutation_test_model(model)
permutation_test_plot (model, P)
}

}
