% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier_MUVR.R
\name{permutation_test_plot}
\alias{permutation_test_plot}
\title{Permutation test plot}
\usage{
permutation_test_plot(MVObj, permMatrix, model = "mid")
}
\arguments{
\item{MVObj}{A MUVR object}

\item{permMatrix}{A permutation fitness outcome from MUVR::permutations()}

\item{model}{'Min', 'Mid', or 'Max'}
}
\value{
A plot with the comparison between the actual model versus the permuted models
}
\description{
Plot permutation test using actual model and permutated models
}
\examples{
\dontrun{
P = permutation_test_model(model)
permutation_test_plot (model, P)
}

}
