% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier_MUVR.R
\name{permutation_test_plot}
\alias{permutation_test_plot}
\title{Permutation test plot}
\usage{
permutation_test_plot(
  MVObj,
  permMatrix,
  model = "mid",
  type = type,
  pos,
  xlab = "Number of misclassifications",
  xlim,
  ylim = NULL,
  breaks = "Sturges",
  main = "Permutation test"
)
}
\arguments{
\item{MVObj}{A MUVR object}

\item{permMatrix}{A permutation fitness outcome from MUVR::permutations()}

\item{model}{'Min', 'Mid', or 'Max'}

\item{type}{'t' (default; for Student's t) or 'non' for "non-parametric" (i.e. rank) studen'ts}

\item{pos}{which side of actual to put p-value on}

\item{xlab}{optional xlabel}

\item{xlim}{optional x-range}

\item{ylim}{otional y-range}

\item{breaks}{optional custom histogram breaks (defaults to 'sturges')}

\item{main}{optional plot title (or TRUE for autoname)}
}
\value{
A plot with the comparison between the actual model versus the permuted models
}
\description{
Plot permutation test using actual model and permutated models
}
\examples{
\dontrun{

# 1.Build a model with the X data from your nmr object and your class:
MVObj <- rdCV_PLS_RF(nmr_data(nmr_peak_table),
Y = nmr_peak_table_completed$Timepoint)


# 2.Model performance
confusion_matrix(MVObj)

# 3.Plotting the model
MUVR_model_plot(MVObj)

# 4.Permutation test
permutations <- permutation_test_model(MVObj, nPerm = 50)

# 5.Plotting permutation test results
permutation_test_plot(MVObj, permutations, model = "Mid", type = "t")

# 6.p-Value
p.value <- p_value_perm(MVObj$miss[["mid"]], permutations[, "Mid"])

# 7.Significant variables
VIPs <- model_VIP(MVObj)
 
# 8.Identification
results <- nmr_identify_regions_blood(ppm_VIP_vector(VIPs))
}

}
