% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipe_add_metadata}
\alias{pipe_add_metadata}
\title{Pipeline: Add Metadata}
\usage{
pipe_add_metadata(nmr_dataset_rds, excel_file, output_dir)
}
\arguments{
\item{nmr_dataset_rds}{The nmr_dataset.rds file name coming from previous nodes}

\item{excel_file}{An excel file name. See details for the requirements

The excel file can have one or more sheets. The excel sheets need to be as
simple as possible: One header column on the first row and values below.

Each of the sheets contain metadata that has to be integrated. The merge
(technically a left join) is done using the first column of each sheet as key.

In practical terms this means that the first sheet of the excel file MUST
start with an "NMRExperiment" column, and as many additional columns to add
(e.g. FluidXBarcode, SampleCollectionDate,    TimePoint and SubjectID).

The second sheet can have as the first column any of the already added columns,
for instance the "SubjectID", and any additional columns (e.g. Gender, Age).

The first column on each sheet, named the key column, MUST have unique values.
For instance, a sheet starting with "SubjectID" MUST specify each subject ID
only once (without repetitions).}

\item{output_dir}{The output directory for this pipe element}
}
\value{
This function saves the result to the output directory
}
\description{
Pipeline: Add Metadata
}
\examples{
dataset <- system.file("dataset-demo", package = "AlpsNMR")
excel_file <- system.file("dataset-demo", 
                          "dummy_metadata.xlsx", 
                          package = "AlpsNMR")
nmr_dataset <- nmr_read_samples_dir(dataset)
nmr_dataset <- nmr_interpolate_1D(nmr_dataset, 
                                  axis = c(min = -0.5, max = 10, by = 2.3E-4))
nmr_dataset_rds <- tempfile(fileext = ".rds")
nmr_dataset_save(nmr_dataset, nmr_dataset_rds)
output_dir <- tempdir()
pipe_add_metadata(nmr_dataset_rds = nmr_dataset_rds, output_dir = output_dir,
                                    excel_file = excel_file)
# Check out: output_dir

}
\seealso{
Other pipeline functions: 
\code{\link{Pipe_normalization}},
\code{\link{pipe_exclude_regions}()},
\code{\link{pipe_filter_samples}()},
\code{\link{pipe_interpolate_1D}()},
\code{\link{pipe_load_samples}()},
\code{\link{pipe_outlier_detection}()},
\code{\link{pipe_pakdet_align}},
\code{\link{pipe_peak_integration}()}

Other metadata functions: 
\code{\link{nmr_meta_add}()},
\code{\link{nmr_meta_export}()},
\code{\link{nmr_meta_get_column}()},
\code{\link{nmr_meta_get}()}
}
\concept{metadata functions}
\concept{pipeline functions}
