% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipe_exclude_regions}
\alias{pipe_exclude_regions}
\title{Pipeline: Exclude regions}
\usage{
pipe_exclude_regions(nmr_dataset_rds, exclude, output_dir)
}
\arguments{
\item{nmr_dataset_rds}{The nmr_dataset.rds file name coming from previous nodes}

\item{exclude}{A list with regions to be removed Typically:
\code{exclude = list(water = c(4.7, 5.0))}}

\item{output_dir}{The output directory for this pipe element}
}
\value{
This function saves the result to the output directory
}
\description{
Pipeline: Exclude regions
}
\examples{
dataset <- system.file("dataset-demo", package = "AlpsNMR")
excel_file <- system.file("dataset-demo", 
                          "dummy_metadata.xlsx", 
                          package = "AlpsNMR")
nmr_dataset <- nmr_read_samples_dir(dataset)
nmr_dataset <- nmr_interpolate_1D(nmr_dataset, 
                                  axis = c(min = -0.5, max = 10, by = 2.3E-4))
nmr_dataset_rds <- tempfile(fileext = ".rds")
nmr_dataset_save(nmr_dataset, nmr_dataset_rds)
exclude_regions <- list(water = c(5.1, 4.5))
output_dir <- tempdir()
pipe_exclude_regions(nmr_dataset_rds, exclude_regions, output_dir)

}
\seealso{
Other pipeline functions: 
\code{\link{Pipe_normalization}},
\code{\link{pipe_add_metadata}()},
\code{\link{pipe_filter_samples}()},
\code{\link{pipe_interpolate_1D}()},
\code{\link{pipe_load_samples}()},
\code{\link{pipe_outlier_detection}()},
\code{\link{pipe_pakdet_align}},
\code{\link{pipe_peak_integration}()}
}
\concept{pipeline functions}
