% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipe_exclude_regions}
\alias{pipe_exclude_regions}
\title{Pipeline: Exclude regions}
\usage{
pipe_exclude_regions(nmr_dataset_rds, exclude, output_dir)
}
\arguments{
\item{nmr_dataset_rds}{The nmr_dataset.rds file name coming from previous nodes}

\item{exclude}{A list with regions to be zeroed. Typically:
\code{exclude = list(water = c(4.7, 5.0))}}

\item{output_dir}{The output directory for this pipe element}
}
\value{
This function saves the result to the output directory
}
\description{
Pipeline: Exclude regions
}
\seealso{
Other pipeline functions: \code{\link{pipe_add_metadata}},
  \code{\link{pipe_interpolate_1D}},
  \code{\link{pipe_load_samples}}
}
\concept{pipeline functions}
