% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipe_filter_samples}
\alias{pipe_filter_samples}
\title{Pipeline: Filter samples according to metadata conditions}
\usage{
pipe_filter_samples(nmr_dataset_rds, conditions, output_dir)
}
\arguments{
\item{nmr_dataset_rds}{The nmr_dataset.rds file name coming from previous nodes}

\item{conditions}{A character vector with conditions to filter metadata.
The \code{conditions} parameter should be a character vector of
valid R logical conditions.
Some examples:
\itemize{
\item conditions <- 'Gender == "Female"'
\item conditions <- 'Cohort == "Chuv"'
\item conditions <- 'TimePoint \%in\% c("T0", "T31")'
\item conditions <- c(Cohort == "Chuv", 'TimePoint \%in\% c("T0", "T31")')
}

Only samples fullfilling all the given conditions are kept in further analysis.}

\item{output_dir}{The output directory for this pipe element}
}
\value{
Pipeline: Filter samples according to metadata conditions
}
\description{
Pipeline: Filter samples according to metadata conditions
}
\examples{
dataset <- system.file("dataset-demo", package = "AlpsNMR")
excel_file <- system.file("dataset-demo", 
                          "dummy_metadata.xlsx", 
                          package = "AlpsNMR")
nmr_dataset <- nmr_read_samples_dir(dataset)
nmr_dataset <- nmr_interpolate_1D(nmr_dataset, 
                                  axis = c(min = -0.5, max = 10, by = 2.3E-4))
nmr_dataset_rds <- tempfile(fileext = ".rds")
nmr_dataset_save(nmr_dataset, nmr_dataset_rds)
output_dir <- tempdir()
conditions <- 'NMRExperiment == "10"'
pipe_filter_samples(nmr_dataset_rds, conditions, output_dir)

}
\seealso{
Other pipeline functions: 
\code{\link{Pipe_normalization}},
\code{\link{pipe_add_metadata}()},
\code{\link{pipe_exclude_regions}()},
\code{\link{pipe_interpolate_1D}()},
\code{\link{pipe_load_samples}()},
\code{\link{pipe_outlier_detection}()},
\code{\link{pipe_pakdet_align}},
\code{\link{pipe_peak_integration}()}
}
\concept{pipeline functions}
