% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipe_interpolate_1D}
\alias{pipe_interpolate_1D}
\title{Pipeline: Interpolate 1D samples}
\usage{
pipe_interpolate_1D(nmr_dataset_rds, axis, output_dir)
}
\arguments{
\item{nmr_dataset_rds}{The nmr_dataset.rds file name coming from previous nodes}

\item{axis}{The ppm axis range and optionally the ppm step}

\item{output_dir}{The output directory for this pipe element}
}
\value{
This function saves the result to the output directory
}
\description{
Pipeline: Interpolate 1D samples
}
\examples{
dataset <- system.file("dataset-demo", package = "AlpsNMR")
excel_file <- system.file("dataset-demo", 
                          "dummy_metadata.xlsx", 
                          package = "AlpsNMR")
nmr_dataset <- nmr_read_samples_dir(dataset)
nmr_dataset_rds <- tempfile(fileext = ".rds")
nmr_dataset_save(nmr_dataset, nmr_dataset_rds)
output_dir <- tempdir()
pipe_interpolate_1D(nmr_dataset_rds, 
                    axis = c(min = -0.5, max = 10, by = 2.3E-4), 
                    output_dir)

}
\seealso{
Other pipeline functions: 
\code{\link{Pipe_normalization}},
\code{\link{pipe_add_metadata}()},
\code{\link{pipe_exclude_regions}()},
\code{\link{pipe_filter_samples}()},
\code{\link{pipe_load_samples}()},
\code{\link{pipe_outlier_detection}()},
\code{\link{pipe_pakdet_align}},
\code{\link{pipe_peak_integration}()}
}
\concept{pipeline functions}
