% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipe_load_samples}
\alias{pipe_load_samples}
\title{Pipeline: Load NMR samples}
\usage{
pipe_load_samples(samples_dir, glob = "*0", output_dir = NULL)
}
\arguments{
\item{samples_dir}{The directory where the samples are}

\item{glob}{A wildcard aka globbing pattern (e.g. \verb{*.csv}) passed on to \code{\link[=grep]{grep()}} to filter paths.}

\item{output_dir}{Directory where the nmr_dataset and the excel files will be saved}
}
\value{
This function saves the result to the output directory
}
\description{
Pipeline: Load NMR samples
}
\examples{
dir_to_demo_dataset <- system.file("dataset-demo", package = "AlpsNMR")
pipe_load_samples(dir_to_demo_dataset,
                  glob = "*.zip",
                  output_dir = "./pipe_output")

}
\seealso{
Other pipeline functions: 
\code{\link{Pipe_normalization}},
\code{\link{pipe_add_metadata}()},
\code{\link{pipe_exclude_regions}()},
\code{\link{pipe_filter_samples}()},
\code{\link{pipe_interpolate_1D}()},
\code{\link{pipe_outlier_detection}()},
\code{\link{pipe_pakdet_align}},
\code{\link{pipe_peak_integration}()}

Other import/export functions: 
\code{\link{files_to_rDolphin_blood}()},
\code{\link{files_to_rDolphin_cell}()},
\code{\link{files_to_rDolphin_urine}()},
\code{\link{load_and_save_functions}},
\code{\link{nmr_data}()},
\code{\link{nmr_meta_export}()},
\code{\link{nmr_read_bruker_fid}()},
\code{\link{nmr_read_samples}()},
\code{\link{nmr_zip_bruker_samples}()},
\code{\link{save_files_to_rDolphin}()},
\code{\link{save_profiling_output}()},
\code{\link{to_ASICS}()},
\code{\link{to_ChemoSpec}()}
}
\concept{import/export functions}
\concept{pipeline functions}
