% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipe_load_samples}
\alias{pipe_load_samples}
\title{Pipeline: Load NMR samples}
\usage{
pipe_load_samples(samples_dir, glob = "*0", output_dir = NULL)
}
\arguments{
\item{samples_dir}{The directory where the samples are}

\item{glob}{A wildcard aka globbing pattern (e.g. \code{*.csv}) passed on to \code{\link[=grep]{grep()}} to filter paths.}

\item{output_dir}{Directory where the nmr_dataset and the excel files will be saved}
}
\value{
This function saves the result to the output directory
}
\description{
Pipeline: Load NMR samples
}
\examples{
\dontrun{
pipe_load_samples("/dir/with/nmr/samples/",
                  glob = "*0",
                  output_dir = "/dir/to/save/output")
}
}
\seealso{
Other pipeline functions: \code{\link{pipe_add_metadata}},
  \code{\link{pipe_exclude_regions}},
  \code{\link{pipe_interpolate_1D}}
}
\concept{pipeline functions}
