% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{Pipe_normalization}
\alias{Pipe_normalization}
\alias{pipe_normalization}
\title{Pipe: Full spectra normalization}
\usage{
pipe_normalization(
  nmr_dataset_rds,
  internal_calibrant = NULL,
  output_dir = NULL
)
}
\arguments{
\item{nmr_dataset_rds}{The nmr_dataset.rds file name coming from previous nodes}

\item{internal_calibrant}{A ppm range where the internal calibrant is, or \code{NULL}.}

\item{output_dir}{The output directory for this pipe element}
}
\value{
Pipe: Full spectra normalization
}
\description{
Normalize the full spectra to the internal calibrant region, then exclude
that region and finally perform PQN normalization.
}
\details{
If there is no internal calibrant, only the PQN normalization is done.
}
\examples{
dataset <- system.file("dataset-demo", package = "AlpsNMR")
excel_file <- system.file("dataset-demo", 
                          "dummy_metadata.xlsx", 
                          package = "AlpsNMR")
nmr_dataset <- nmr_read_samples_dir(dataset)
nmr_dataset <- nmr_interpolate_1D(nmr_dataset, 
                                  axis = c(min = -0.5, max = 10, by = 2.3E-4))
nmr_dataset_rds <- tempfile(fileext = ".rds")
nmr_dataset_save(nmr_dataset, nmr_dataset_rds)
output_dir <- tempdir()
pipe_normalization(nmr_dataset_rds, output_dir = output_dir)

}
\seealso{
Other pipeline functions: 
\code{\link{pipe_add_metadata}()},
\code{\link{pipe_exclude_regions}()},
\code{\link{pipe_filter_samples}()},
\code{\link{pipe_interpolate_1D}()},
\code{\link{pipe_load_samples}()},
\code{\link{pipe_outlier_detection}()},
\code{\link{pipe_pakdet_align}},
\code{\link{pipe_peak_integration}()}
}
\concept{pipeline functions}
