% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipe_normalization}
\alias{pipe_normalization}
\title{Pipe: Full spectra normalization}
\usage{
pipe_normalization(
  nmr_dataset_rds,
  internal_calibrant = NULL,
  output_dir = NULL
)
}
\arguments{
\item{nmr_dataset_rds}{The nmr_dataset.rds file name coming from previous nodes}

\item{internal_calibrant}{A ppm range where the internal calibrant is, or \code{NULL}.}

\item{output_dir}{The output directory for this pipe element}
}
\description{
Normalize the full spectra to the internal calibrant region, then exclude
that region and finally perform PQN normalization.
}
\details{
If there is no internal calibrant, only the PQN normalization is done.
}
\seealso{
Other pipeline functions: 
\code{\link{pipe_add_metadata}()},
\code{\link{pipe_exclude_regions}()},
\code{\link{pipe_filter_samples}()},
\code{\link{pipe_interpolate_1D}()},
\code{\link{pipe_load_samples}()},
\code{\link{pipe_outlier_detection}()},
\code{\link{pipe_peak_integration}()},
\code{\link{pipe_peakdet_align}()}
}
\concept{pipeline functions}
