% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipe_outlier_detection}
\alias{pipe_outlier_detection}
\title{Pipeline: Remove blatant outliers}
\usage{
pipe_outlier_detection(nmr_dataset_rds, output_dir)
}
\arguments{
\item{nmr_dataset_rds}{The nmr_dataset.rds file name coming from previous nodes}

\item{output_dir}{The output directory for this pipe element}
}
\value{
This function saves the result to the output directory
}
\description{
Uses \link{nmr_pca_outliers_robust} to perform the detection of outliers
}
\seealso{
Other outlier detection functions: \code{\link{nmr_pca_outliers_filter}},
  \code{\link{nmr_pca_outliers_plot}},
  \code{\link{nmr_pca_outliers_robust}},
  \code{\link{nmr_pca_outliers}}

Other pipeline functions: \code{\link{pipe_add_metadata}},
  \code{\link{pipe_exclude_regions}},
  \code{\link{pipe_filter_samples}},
  \code{\link{pipe_interpolate_1D}},
  \code{\link{pipe_load_samples}},
  \code{\link{pipe_normalization}},
  \code{\link{pipe_peak_integration}},
  \code{\link{pipe_peakdet_align}}
}
\concept{outlier detection functions}
\concept{pipeline functions}
