% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipe_peak_integration}
\alias{pipe_peak_integration}
\title{Pipeline: Peak integration}
\usage{
pipe_peak_integration(nmr_dataset_rds, peak_det_align_dir, peak_width_ppm,
  output_dir)
}
\arguments{
\item{nmr_dataset_rds}{The nmr_dataset.rds file name coming from previous nodes}

\item{peak_det_align_dir}{Output directory from \link{pipe_peakdet_align}}

\item{peak_width_ppm}{A peak width in ppm}

\item{output_dir}{The output directory for this pipe element}
}
\description{
Pipeline: Peak integration
}
\seealso{
Other pipeline functions: \code{\link{pipe_add_metadata}},
  \code{\link{pipe_exclude_regions}},
  \code{\link{pipe_filter_samples}},
  \code{\link{pipe_interpolate_1D}},
  \code{\link{pipe_load_samples}},
  \code{\link{pipe_normalization}},
  \code{\link{pipe_outlier_detection}},
  \code{\link{pipe_peakdet_align}}

Other peak integration functions: \code{\link{computes_peak_width_ppm}},
  \code{\link{get_integration_with_metadata}},
  \code{\link{nmr_identify_regions_blood}},
  \code{\link{nmr_identify_regions_urine}},
  \code{\link{nmr_integrate_peak_positions}},
  \code{\link{nmr_integrate_regions}}
}
\concept{peak integration functions}
\concept{pipeline functions}
