% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipe_peakdet_align}
\alias{pipe_peakdet_align}
\title{Pipeline: Peak detection and Alignment}
\usage{
pipe_peakdet_align(nmr_dataset_rds, nDivRange_ppm = 0.1,
  scales = seq(1, 16, 2), baselineThresh = 0.01, SNR.Th = -1,
  maxShift_ppm = 0.0015, acceptLostPeak = FALSE, output_dir = NULL)
}
\arguments{
\item{nmr_dataset_rds}{The nmr_dataset.rds file name coming from previous nodes}

\item{nDivRange_ppm}{Segment size, in ppms, to divide the spectra and search for peaks.}

\item{scales}{The parameter of peakDetectionCWT function of MassSpecWavelet package, look it up in the original function.}

\item{baselineThresh}{It will remove all peaks under an intensity set by baselineThresh.}

\item{SNR.Th}{The parameter of peakDetectionCWT function of MassSpecWavelet package, look it up in the original function. If you set -1, the function will itself re-compute this value.}

\item{maxShift_ppm}{The maximum shift allowed, in ppm}

\item{acceptLostPeak}{This is an option for users, TRUE is the default value. If the users believe that all the peaks in the peak list are true positive, change it to FALSE.}

\item{output_dir}{The output directory for this pipe element}
}
\description{
Pipeline: Peak detection and Alignment
}
