% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_interactive}
\alias{plot_interactive}
\title{Plots in WebGL}
\usage{
plot_interactive(plt, html_filename)
}
\arguments{
\item{plt}{A plot created with plotly or ggplot2}

\item{html_filename}{The file name where the plot will be saved}
}
\value{
The html_filename
}
\description{
Plots in WebGL
}
\seealso{
Other plotting functions: \code{\link{plot.nmr_dataset_1D}}
}
\concept{plotting functions}
