% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_webgl}
\alias{plot_webgl}
\title{Plot a dataset into a HTML file}
\usage{
plot_webgl(nmr_dataset, html_filename, ...)
}
\arguments{
\item{nmr_dataset}{An \link{nmr_dataset_1D}}

\item{html_filename}{The output HTML filename to be created}

\item{...}{Arguments passed on to \code{plot.nmr_dataset_1D}
\describe{
  \item{x}{a \link{nmr_dataset_1D} object}
  \item{NMRExperiment}{A character vector with the NMRExperiments to include. Use "all" to include all experiments.}
  \item{quantile_plot}{If \code{TRUE} plot the 10\%, 50\%, 90\% percentiles of the spectra as reference.
If two numbers between 0 and 1 are given then a custom percentile can be plotted}
  \item{quantile_colors}{A vector with the colors for each of the quantiles}
  \item{interactive}{if \code{TRUE} return an interactive plotly plot, otherwise return a ggplot one.}
  \item{chemshift_range}{range of the chemical shifts to be included. Can be of length 3
to include the resolution in the third element (e.g. \code{c(0.2, 0.8, 0.05)})}
}}
}
\value{
the html filename created
}
\description{
Uses WebGL for performance
}
\seealso{
Other plotting nmr datasets: \code{\link{plot.nmr_dataset_1D}},
  \code{\link{plot_interactive}}
}
\concept{plotting nmr datasets}
