% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset.R
\name{print.nmr_dataset}
\alias{print.nmr_dataset}
\title{Print for nmr_dataset}
\usage{
\method{print}{nmr_dataset}(x, ...)
}
\arguments{
\item{x}{an \link{nmr_dataset} object}

\item{...}{for future use}
}
\value{
Print for nmr_dataset
}
\description{
Print for nmr_dataset
}
\examples{
dir_to_demo_dataset <- system.file("dataset-demo", package = "AlpsNMR")
dataset <- nmr_read_samples_dir(dir_to_demo_dataset)
print(dataset)

}
\seealso{
Other class helper functions: 
\code{\link{format.nmr_dataset_1D}()},
\code{\link{format.nmr_dataset_peak_table}()},
\code{\link{format.nmr_dataset}()},
\code{\link{is.nmr_dataset_1D}()},
\code{\link{is.nmr_dataset_peak_table}()},
\code{\link{new_nmr_dataset_1D}()},
\code{\link{new_nmr_dataset_peak_table}()},
\code{\link{new_nmr_dataset}()},
\code{\link{print.nmr_dataset_1D}()},
\code{\link{print.nmr_dataset_peak_table}()},
\code{\link{validate_nmr_dataset_family}()},
\code{\link{validate_nmr_dataset_peak_table}()},
\code{\link{validate_nmr_dataset}()}

Other nmr_dataset functions: 
\code{\link{[.nmr_dataset}()},
\code{\link{format.nmr_dataset}()},
\code{\link{load_and_save_functions}},
\code{\link{new_nmr_dataset}()},
\code{\link{nmr_interpolate_1D}()},
\code{\link{nmr_meta_add}()},
\code{\link{nmr_meta_export}()},
\code{\link{nmr_meta_get_column}()},
\code{\link{nmr_meta_get}()},
\code{\link{nmr_ppm_resolution}()},
\code{\link{nmr_read_samples}()},
\code{\link{validate_nmr_dataset}()}
}
\concept{class helper functions}
\concept{nmr_dataset functions}
