% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier_MUVR.R
\name{rdCV_PLS_RF}
\alias{rdCV_PLS_RF}
\title{Deprecated function}
\usage{
rdCV_PLS_RF(nmr_peak_table)
}
\arguments{
\item{nmr_peak_table}{an AlpsNMR integration object (2 classes)}
}
\value{
a MUVR model containing selection parameters, validation and fitness
}
\description{
Deprecated function
}
\examples{
message("Deprecated. MUVR is not compatible with Bioconductor, 
use bp_kfold_VIP_analysis method instead")
}
\references{
Shi,L. et al. (2018) Variable selection and validation in multivariate modelling. Bioinformatics.
}
\seealso{
nmr_data_analysis
Feature selection and validation in multivariate analysis

Statistical analysis and feature selection in a repeated double
cross-validation frame based on the partial least squares (PLS) or random
forest (RF) analyses using an algorithm for multivariate modelling with
minimally biased variable selection (MUVR) from the \code{MUVR} package. If your
work with a \code{nmr_peak_table} object from AlpsNMR, first you need to extract
the X data from the main nmr_dataset object (e.g. your peak table) with the
\code{nmr_data} function, otherwise you would try to set a list on the X. You also
need to set the class from this object, or just set it from another Y vector.

Other nmr_dataset_1D functions: 
\code{\link{[.nmr_dataset_1D}()},
\code{\link{computes_peak_width_ppm}()},
\code{\link{file_lister}()},
\code{\link{files_to_rDolphin}()},
\code{\link{format.nmr_dataset_1D}()},
\code{\link{get_integration_with_metadata}()},
\code{\link{is.nmr_dataset_1D}()},
\code{\link{load_and_save_functions}},
\code{\link{new_nmr_dataset_1D}()},
\code{\link{nmr_align_find_ref}()},
\code{\link{nmr_align}()},
\code{\link{nmr_baseline_removal}()},
\code{\link{nmr_baseline_threshold}()},
\code{\link{nmr_detect_peaks_plot}()},
\code{\link{nmr_detect_peaks_tune_snr}()},
\code{\link{nmr_detect_peaks}()},
\code{\link{nmr_exclude_region}()},
\code{\link{nmr_integrate_peak_positions}()},
\code{\link{nmr_integrate_regions}()},
\code{\link{nmr_interpolate_1D}()},
\code{\link{nmr_meta_add}()},
\code{\link{nmr_meta_export}()},
\code{\link{nmr_meta_get_column}()},
\code{\link{nmr_meta_get}()},
\code{\link{nmr_normalize}()},
\code{\link{nmr_pca_build_model}()},
\code{\link{nmr_pca_outliers_filter}()},
\code{\link{nmr_pca_outliers_plot}()},
\code{\link{nmr_pca_outliers_robust}()},
\code{\link{nmr_pca_outliers}()},
\code{\link{nmr_ppm_resolution}()},
\code{\link{plot.nmr_dataset_1D}()},
\code{\link{plot_webgl}()},
\code{\link{print.nmr_dataset_1D}()},
\code{\link{rdCV_PLS_RF_ML}()},
\code{\link{save_files_to_rDolphin}()},
\code{\link{to_ChemoSpec}()},
\code{\link{validate_nmr_dataset}()}
}
\concept{nmr_dataset_1D functions}
