% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nihs_specific.R
\name{read_exported_slims}
\alias{read_exported_slims}
\title{Reads an exported xlsx slims spreadsheet}
\usage{
read_exported_slims(file_name, sheet = NULL, all_character = FALSE)
}
\arguments{
\item{file_name}{The exported slims spreadsheet in excel format}

\item{sheet}{passed on to read_excel}

\item{all_character}{Read all columns as text}
}
\value{
a data frame with the slims information

Key columns on the output data frame are "cntn_cf_fluidx_barcode" and "cntn_cf_subject_id" "cntn_cf_fk_masterSample"
}
\description{
This function reads an xlsx slims spreadsheet and converts it into a data frame
}
\seealso{
Other NIHS specific infrastructure functions: \code{\link{nmr_get_irods_meta}},
  \code{\link{nmr_irods_search}},
  \code{\link{nmr_irods_validate_meta}},
  \code{\link{nmr_prepare_zip_files}},
  \code{\link{nmr_push_to_irods}},
  \code{\link{nmr_read_samples_irods}}
}
\concept{NIHS specific infrastructure functions}
