% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlpsNMR.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tidy}
\alias{rename}
\alias{filter}
\alias{\%>\%}
\alias{.DollarNames}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}, \code{\link[dplyr]{rename}}}

  \item{generics}{\code{\link[generics]{tidy}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{utils}{\code{\link[utils:rcompgen]{.DollarNames}}}
}}

