% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset_family.R, R/utils.R,
%   R/utils_future.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{filter}
\alias{\%>\%}
\alias{plan}
\alias{sequential}
\alias{multicore}
\alias{multiprocess}
\alias{multisession}
\alias{cluster}
\alias{remote}
\title{Objects exported from other packages}
\examples{
dir_to_demo_dataset <- system.file("dataset-demo", package = "AlpsNMR")
dataset <- nmr_read_samples_dir(dir_to_demo_dataset)
dataset_1D <- nmr_interpolate_1D(dataset, axis = c(min = -0.5, max = 10, by = 2.3E-4))

## example 1
sample_10 <- filter(dataset_1D, NMRExperiment == "10")

## example 2
#test_samples <- dataset_1D \%>\% filter(nmr_peak_table$metadata$external$Group == "placebo")

}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}}

  \item{future}{\code{\link[future]{cluster}}, \code{\link[future]{multicore}}, \code{\link[future]{multiprocess}}, \code{\link[future]{multisession}}, \code{\link[future]{plan}}, \code{\link[future]{remote}}, \code{\link[future]{sequential}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

