% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset_family.R, R/utils.R,
%   R/utils_future.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{filter}
\alias{\%>\%}
\alias{plan}
\alias{sequential}
\alias{multicore}
\alias{multiprocess}
\alias{multisession}
\alias{cluster}
\alias{remote}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}}

  \item{future}{\code{\link[future]{plan}}, \code{\link[future]{sequential}}, \code{\link[future]{multicore}}, \code{\link[future]{multiprocess}}, \code{\link[future]{multisession}}, \code{\link[future]{cluster}}, \code{\link[future]{remote}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

