% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_rDolphin_blood.R
\docType{data}
\name{ROI_blood}
\alias{ROI_blood}
\title{ROIs for blood (plasma/serum) samples}
\description{
The template ROI_blood contains the targeted list of metabolites to be quantified (blood samples)
}
\references{
\url{github.com/danielcanueto/rDolphin}
}
\keyword{data}
