% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_rDolphin_blood.R
\docType{data}
\name{ROI_urine}
\alias{ROI_urine}
\title{The template ROI_urine contains the targeted list of metabolites to be quantified (urine samples)}
\description{
The template ROI_urine contains the targeted list of metabolites to be quantified (urine samples)
}
\references{
\url{github.com/danielcanueto/rDolphin}
}
\keyword{data}
