% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_rDolphin_blood.R
\name{save_files_to_rDolphin}
\alias{save_files_to_rDolphin}
\title{Save files to rDoplhin}
\usage{
save_files_to_rDolphin(files_rDolphin, output_directory)
}
\arguments{
\item{files_rDolphin}{a list containing 4 elements from \code{files_to_rDolphin}
\itemize{
\item \code{meta_rDolphin}: metadata in rDolphin format,
\item \code{NMR_spectra}: spectra matrix
\item \code{ROI}: ROI template
\item \code{Parameters_blood}: parameters file
}}

\item{output_directory}{a directory in which the CSV files are saved}
}
\value{
CSV files containing:
}
\description{
The function saves the CSV files required by to_rDolphin and Automatic_targeted_profiling functions for metabolite profiling.
}
\examples{
\dontrun{
dataset <- system.file("dataset-demo", package = "AlpsNMR")
excel_file <- system.file("dataset-demo", "dummy_metadata.xlsx", package = "AlpsNMR")
nmr_dataset <- nmr_read_samples_dir(dataset)
files_rDolphin = files_to_rDolphin_blood(nmr_dataset)
output_directory = "."
save_files_to_rDolphin(files_rDolphin, output_directory)
}
}
\seealso{
Other import/export functions: 
\code{\link{Pipelines}},
\code{\link{files_to_rDolphin}()},
\code{\link{load_and_save_functions}},
\code{\link{nmr_data}()},
\code{\link{nmr_meta_export}()},
\code{\link{nmr_read_bruker_fid}()},
\code{\link{nmr_read_samples}()},
\code{\link{nmr_zip_bruker_samples}()},
\code{\link{save_profiling_output}()},
\code{\link{to_ChemoSpec}()}

Other nmr_dataset_1D functions: 
\code{\link{[.nmr_dataset_1D}()},
\code{\link{computes_peak_width_ppm}()},
\code{\link{file_lister}()},
\code{\link{files_to_rDolphin}()},
\code{\link{format.nmr_dataset_1D}()},
\code{\link{get_integration_with_metadata}()},
\code{\link{is.nmr_dataset_1D}()},
\code{\link{load_and_save_functions}},
\code{\link{new_nmr_dataset_1D}()},
\code{\link{nmr_align_find_ref}()},
\code{\link{nmr_align}()},
\code{\link{nmr_baseline_removal}()},
\code{\link{nmr_baseline_threshold}()},
\code{\link{nmr_detect_peaks_plot}()},
\code{\link{nmr_detect_peaks_tune_snr}()},
\code{\link{nmr_detect_peaks}()},
\code{\link{nmr_exclude_region}()},
\code{\link{nmr_integrate_peak_positions}()},
\code{\link{nmr_integrate_regions}()},
\code{\link{nmr_interpolate_1D}()},
\code{\link{nmr_meta_add}()},
\code{\link{nmr_meta_export}()},
\code{\link{nmr_meta_get_column}()},
\code{\link{nmr_meta_get}()},
\code{\link{nmr_normalize}()},
\code{\link{nmr_pca_build_model}()},
\code{\link{nmr_pca_outliers_filter}()},
\code{\link{nmr_pca_outliers_plot}()},
\code{\link{nmr_pca_outliers_robust}()},
\code{\link{nmr_pca_outliers}()},
\code{\link{nmr_ppm_resolution}()},
\code{\link{plot.nmr_dataset_1D}()},
\code{\link{plot_webgl}()},
\code{\link{print.nmr_dataset_1D}()},
\code{\link{rdCV_PLS_RF_ML}()},
\code{\link{rdCV_PLS_RF}()},
\code{\link{to_ChemoSpec}()},
\code{\link{validate_nmr_dataset}()}

Other to_rDolphin_blood functions: 
\code{\link{save_profiling_output}()}
}
\concept{import/export functions}
\concept{nmr_dataset_1D functions}
\concept{to_rDolphin_blood functions}
