% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_rDolphin_blood.R
\name{save_profiling_output}
\alias{save_profiling_output}
\title{Save rDoplhin output}
\usage{
save_profiling_output(targeted_profiling, output_directory)
}
\arguments{
\item{targeted_profiling}{A list from Automatic_targeted_profiling function}

\item{output_directory}{a directory in which the CSV files are saved}
}
\value{
rDolphin output from Automatic_targeted_profiling function:
\itemize{
\item metabolites_intensity
\item metabolites_quantification
\item ROI_profiles_used
\item chemical_shift
\item fitting_error
\item half_bandwidth
\item signal_area_ratio
}
}
\description{
The function saves the output from Automatic_targeted_profiling function in CSV format.
}
\examples{
\dontrun{
rDolphin_object = to_rDolphin(parameters)
targeted_profiling = Automatic_targeted_profiling(rDolphin)
save_profiling_output(targeted_profiling, output_directory)
}
}
\seealso{
Other import/export functions: 
\code{\link{Pipelines}},
\code{\link{files_to_rDolphin_blood}()},
\code{\link{files_to_rDolphin_cell}()},
\code{\link{files_to_rDolphin_urine}()},
\code{\link{load_and_save_functions}},
\code{\link{nmr_data}()},
\code{\link{nmr_meta_export}()},
\code{\link{nmr_read_bruker_fid}()},
\code{\link{nmr_read_samples}()},
\code{\link{nmr_zip_bruker_samples}()},
\code{\link{save_files_to_rDolphin}()},
\code{\link{to_ChemoSpec}()}

Other to_rDolphin_blood functions: 
\code{\link{files_to_rDolphin_blood}()},
\code{\link{save_files_to_rDolphin}()}
}
\concept{import/export functions}
\concept{to_rDolphin_blood functions}
