% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_rDolphin_blood.R
\name{save_profiling_plots}
\alias{save_profiling_plots}
\title{Save rDolphin output and profiling plots}
\usage{
save_profiling_plots(export_path, final_output, reproducibility_data,
  signals_to_plot = NA)
}
\arguments{
\item{export_path}{Path where plots are stored inside a created 'plots' folder}

\item{final_output}{List with quantifications and indicators of quality of quantification.}

\item{reproducibility_data}{List with necessary information to load quantifications on the Shiny GUI.}

\item{signals_to_plot}{Vector of indexes of signals in ROI data to plot. By default, NA and all figures are outputted.}
}
\value{
Plots of the fitted compounds in pdf files
}
\description{
The function writes the plots of the adjusted compounds.
}
\examples{
\dontrun{
## library(AlpsNMR)
Set the directory in which rDolphin files will be saved
output_dir_10_rDolphin <- file.path(your_path, "10-rDolphin")
fs::dir_create(output_dir_10_rDolphin)

# Generate the files (for plasma/serum)
files_rDolphin = files_to_rDolphin_blood(nmr_dataset_0_10_ppm)

# Save the files
save_files_to_rDolphin(files_rDolphin, output_dir_10_rDolphin)

# Build the rDolphin object. Do not forget to set the directory
setwd(output_dir_10_rDolphin) 
rDolphin_object = to_rDolphin("Parameters.csv")

# Visualize your spectra
rDolphin_plot(rDolphin_object)

# Run the main profiling function (it takes a while)
targeted_profiling = Automatic_targeted_profiling(rDolphin_object)

# Save results
save_profiling_output(targeted_profiling, output_dir_10_rDolphin)

save_profiling_plots(output_dir_10_rDolphin, targeted_profiling$final_output,
targeted_profiling$reproducibility_data)

#Additionally, you can run some stats
intensities = targeted_profiling$final_output$intensity
group = as.factor(rDolphin_object$Metadata$type)
model_PLS <- rdCV_PLS_RF(X = intensities, Y = group)
}

}
