% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_rDolphin_blood.R
\name{save_profiling_plots}
\alias{save_profiling_plots}
\title{Save rDolphin output and profiling plots}
\usage{
save_profiling_plots(export_path, final_output, reproducibility_data,
  signals_to_plot = NA)
}
\arguments{
\item{export_path}{Path where plots are stored inside a created 'plots' folder}

\item{final_output}{List with quantifications and indicators of quality of quantification.}

\item{reproducibility_data}{List with necessary information to load quantifications on the Shiny GUI.}

\item{signals_to_plot}{Vector of indexes of signals in ROI data to plot. By default, NA and all figures are outputted.}
}
\value{
Plots of the fitted compounds in pdf files
}
\description{
The function writes the plots of the adjusted compounds.
}
\examples{
\dontrun{
## library(AlpsNMR)
rDolphin_object = to_rDolphin(parameters)
targeted_profiling = Automatic_targeted_profiling(rDolphin_object)
write_plots(dir_10_rDolphin, targeted_profiling$final_output,
targeted_profiling$reproducibility_data)
}

}
