% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset.R
\name{[.nmr_dataset}
\alias{[.nmr_dataset}
\title{Extract parts of an nmr_dataset}
\usage{
\method{[}{nmr_dataset}(x, i)
}
\arguments{
\item{x}{an \link{nmr_dataset} object}

\item{i}{indices of the samples to keep}
}
\value{
an nmr_dataset with the extracted samples
}
\description{
Extract parts of an nmr_dataset
}
\examples{
\dontrun{
data <- nm_read_samples_dir("your_dir")
data2 <- data[1:3] # get the first 3 samples
}
}
