% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset.R
\name{[.nmr_dataset}
\alias{[.nmr_dataset}
\title{Extract parts of an nmr_dataset}
\usage{
\method{[}{nmr_dataset}(x, i)
}
\arguments{
\item{x}{an \link{nmr_dataset} object}

\item{i}{indices of the samples to keep}
}
\value{
an nmr_dataset with the extracted samples
}
\description{
Extract parts of an nmr_dataset
}
\examples{
dir_to_demo_dataset <- system.file("dataset-demo", package = "AlpsNMR")
dataset <- nmr_read_samples_dir(dir_to_demo_dataset)
dataset2 <- dataset[1:3] # get the first 3 samples

}
\seealso{
Other subsetting functions: 
\code{\link{[.nmr_dataset_1D}()},
\code{\link{[.nmr_dataset_peak_table}()},
\code{\link{filter.nmr_dataset_family}()},
\code{\link{nmr_pca_outliers_filter}()}

Other nmr_dataset functions: 
\code{\link{format.nmr_dataset}()},
\code{\link{load_and_save_functions}},
\code{\link{new_nmr_dataset}()},
\code{\link{nmr_interpolate_1D}()},
\code{\link{nmr_meta_add}()},
\code{\link{nmr_meta_export}()},
\code{\link{nmr_meta_get_column}()},
\code{\link{nmr_meta_get}()},
\code{\link{nmr_ppm_resolution}()},
\code{\link{nmr_read_samples}()},
\code{\link{print.nmr_dataset}()},
\code{\link{validate_nmr_dataset}()}
}
\concept{nmr_dataset functions}
\concept{subsetting functions}
