% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset_1D.R
\name{[.nmr_dataset_1D}
\alias{[.nmr_dataset_1D}
\title{Extract parts of an nmr_dataset_1D}
\usage{
\method{[}{nmr_dataset_1D}(x, i)
}
\arguments{
\item{x}{an \link{nmr_dataset_1D} object}

\item{i}{indices of the samples to keep}
}
\value{
an nmr_dataset_1D with the extracted samples
}
\description{
Extract parts of an nmr_dataset_1D
}
