% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset_1D.R
\name{[.nmr_dataset_1D}
\alias{[.nmr_dataset_1D}
\title{Extract parts of an nmr_dataset_1D}
\usage{
\method{[}{nmr_dataset_1D}(x, i)
}
\arguments{
\item{x}{an \link{nmr_dataset_1D} object}

\item{i}{indices of the samples to keep}
}
\value{
an nmr_dataset_1D with the extracted samples
}
\description{
Extract parts of an nmr_dataset_1D
}
\seealso{
Other subsetting functions: \code{\link{[.nmr_dataset_peak_table}},
  \code{\link{[.nmr_dataset}},
  \code{\link{filter.nmr_dataset}},
  \code{\link{nmr_pca_outliers_filter}}

Other nmr_dataset_1D functions: \code{\link{is.nmr_dataset_1D}},
  \code{\link{load_and_save_functions}},
  \code{\link{new_nmr_dataset_1D}},
  \code{\link{nmr_align_find_ref}},
  \code{\link{nmr_align}},
  \code{\link{nmr_detect_peaks_plot}},
  \code{\link{nmr_detect_peaks}},
  \code{\link{nmr_exclude_region}},
  \code{\link{nmr_integrate_peak_positions}},
  \code{\link{nmr_integrate_regions}},
  \code{\link{nmr_interpolate_1D}},
  \code{\link{nmr_meta_add}},
  \code{\link{nmr_meta_export}},
  \code{\link{nmr_meta_get_column}},
  \code{\link{nmr_meta_get}}, \code{\link{nmr_normalize}},
  \code{\link{nmr_pca_build_model}},
  \code{\link{nmr_pca_outliers_filter}},
  \code{\link{nmr_pca_outliers_plot}},
  \code{\link{nmr_pca_outliers_robust}},
  \code{\link{nmr_pca_outliers}},
  \code{\link{nmr_ppm_resolution}},
  \code{\link{plot.nmr_dataset_1D}},
  \code{\link{plot_webgl}}, \code{\link{to_ASICS}},
  \code{\link{to_ChemoSpec}},
  \code{\link{validate_nmr_dataset_1D}}
}
\concept{nmr_dataset_1D functions}
\concept{subsetting functions}
