% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset_peak_table.R
\name{[.nmr_dataset_peak_table}
\alias{[.nmr_dataset_peak_table}
\title{Extract parts of an nmr_dataset_peak_table}
\usage{
\method{[}{nmr_dataset_peak_table}(x, i)
}
\arguments{
\item{x}{an \link{nmr_dataset_peak_table} object}

\item{i}{indices of the samples to keep}
}
\value{
an nmr_dataset_peak_table with the extracted samples
}
\description{
Extract parts of an nmr_dataset_peak_table
}
