% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset_peak_table.R
\name{[.nmr_dataset_peak_table}
\alias{[.nmr_dataset_peak_table}
\title{Extract parts of an nmr_dataset_peak_table}
\usage{
\method{[}{nmr_dataset_peak_table}(x, i)
}
\arguments{
\item{x}{an \link{nmr_dataset_peak_table} object}

\item{i}{indices of the samples to keep}
}
\value{
an nmr_dataset_peak_table with the extracted samples
}
\description{
Extract parts of an nmr_dataset_peak_table
}
\seealso{
Other subsetting functions: \code{\link{[.nmr_dataset_1D}},
  \code{\link{[.nmr_dataset}},
  \code{\link{filter.nmr_dataset}},
  \code{\link{nmr_pca_outliers_filter}}

Other nmr_dataset_peak_table functions: \code{\link{is.nmr_dataset_peak_table}},
  \code{\link{load_and_save_functions}},
  \code{\link{new_nmr_dataset_peak_table}},
  \code{\link{nmr_meta_add}},
  \code{\link{nmr_meta_export}},
  \code{\link{nmr_meta_get_column}},
  \code{\link{nmr_meta_get}},
  \code{\link{validate_nmr_dataset_peak_table}}
}
\concept{nmr_dataset_peak_table functions}
\concept{subsetting functions}
