% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset.R
\name{validate_nmr_dataset}
\alias{validate_nmr_dataset}
\title{Validate nmr_dataset objects}
\usage{
validate_nmr_dataset(samples)
}
\arguments{
\item{samples}{An nmr_dataset object}
}
\description{
Validate nmr_dataset objects
}
\seealso{
Other class helper functions: \code{\link{is.nmr_dataset_1D}},
  \code{\link{is.nmr_dataset_peak_table}},
  \code{\link{new_nmr_dataset_1D}},
  \code{\link{new_nmr_dataset_peak_table}},
  \code{\link{new_nmr_dataset}},
  \code{\link{validate_nmr_dataset_1D}},
  \code{\link{validate_nmr_dataset_peak_table}}

Other nmr_dataset functions: \code{\link{[.nmr_dataset}},
  \code{\link{load_and_save_functions}},
  \code{\link{new_nmr_dataset}},
  \code{\link{nmr_interpolate_1D}},
  \code{\link{nmr_meta_add}},
  \code{\link{nmr_meta_export}},
  \code{\link{nmr_meta_get_column}},
  \code{\link{nmr_meta_get}},
  \code{\link{nmr_ppm_resolution}},
  \code{\link{nmr_read_samples}}
}
\concept{class helper functions}
\concept{nmr_dataset functions}
