% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset_1D.R
\name{validate_nmr_dataset_1D}
\alias{validate_nmr_dataset_1D}
\title{Validate 1D nmr datasets}
\usage{
validate_nmr_dataset_1D(nmr_dataset_1D)
}
\arguments{
\item{nmr_dataset_1D}{An \link{nmr_dataset_1D} object}
}
\value{
The \link{nmr_dataset_1D} unchanged

This function is useful for its side-effects. Stopping in case of error
}
\description{
Validate 1D nmr datasets
}
\seealso{
Other class helper functions: \code{\link{is.nmr_dataset_1D}},
  \code{\link{is.nmr_dataset_peak_table}},
  \code{\link{new_nmr_dataset_1D}},
  \code{\link{new_nmr_dataset_peak_table}},
  \code{\link{new_nmr_dataset}},
  \code{\link{validate_nmr_dataset_family}},
  \code{\link{validate_nmr_dataset_peak_table}},
  \code{\link{validate_nmr_dataset}}

Other nmr_dataset_1D functions: \code{\link{Automatic_targeted_profiling}},
  \code{\link{[.nmr_dataset_1D}},
  \code{\link{computes_peak_width_ppm}},
  \code{\link{files_to_rDolphin}},
  \code{\link{get_integration_with_metadata}},
  \code{\link{is.nmr_dataset_1D}},
  \code{\link{load_and_save_functions}},
  \code{\link{new_nmr_dataset_1D}},
  \code{\link{nmr_align_find_ref}},
  \code{\link{nmr_align}},
  \code{\link{nmr_baseline_removal}},
  \code{\link{nmr_baseline_threshold}},
  \code{\link{nmr_detect_peaks_plot}},
  \code{\link{nmr_detect_peaks_tune_snr}},
  \code{\link{nmr_detect_peaks}},
  \code{\link{nmr_exclude_region}},
  \code{\link{nmr_integrate_peak_positions}},
  \code{\link{nmr_integrate_regions}},
  \code{\link{nmr_interpolate_1D}},
  \code{\link{nmr_meta_add}},
  \code{\link{nmr_meta_export}},
  \code{\link{nmr_meta_get_column}},
  \code{\link{nmr_meta_get}}, \code{\link{nmr_normalize}},
  \code{\link{nmr_pca_build_model}},
  \code{\link{nmr_pca_outliers_filter}},
  \code{\link{nmr_pca_outliers_plot}},
  \code{\link{nmr_pca_outliers_robust}},
  \code{\link{nmr_pca_outliers}},
  \code{\link{nmr_ppm_resolution}},
  \code{\link{plot.nmr_dataset_1D}},
  \code{\link{plot_webgl}}, \code{\link{rdCV_PLS_RF}},
  \code{\link{save_files_to_rDolphin}},
  \code{\link{to_ASICS}}, \code{\link{to_ChemoSpec}},
  \code{\link{to_rDolphin}}
}
\concept{class helper functions}
\concept{nmr_dataset_1D functions}
