% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset_family.R
\name{validate_nmr_dataset_family}
\alias{validate_nmr_dataset_family}
\title{Validate nmr_dataset_family objects}
\usage{
validate_nmr_dataset_family(nmr_dataset_family)
}
\arguments{
\item{nmr_dataset_family}{An \link{nmr_dataset_family} object}
}
\value{
The \link{nmr_dataset_family} unchanged

This function is useful for its side-effects: Stopping in case of error
}
\description{
Validate nmr_dataset_family objects
}
\seealso{
Other class helper functions: 
\code{\link{is.nmr_dataset_1D}()},
\code{\link{is.nmr_dataset_peak_table}()},
\code{\link{new_nmr_dataset_1D}()},
\code{\link{new_nmr_dataset_peak_table}()},
\code{\link{new_nmr_dataset}()},
\code{\link{validate_nmr_dataset_1D}()},
\code{\link{validate_nmr_dataset_peak_table}()},
\code{\link{validate_nmr_dataset}()}
}
\concept{class helper functions}
\concept{nmr_dataset_family functions}
