% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset_peak_table.R
\name{validate_nmr_dataset_peak_table}
\alias{validate_nmr_dataset_peak_table}
\title{Validate nmr_dataset_peak_table objects}
\usage{
validate_nmr_dataset_peak_table(nmr_dataset_peak_table)
}
\arguments{
\item{nmr_dataset_peak_table}{An \link{nmr_dataset_peak_table} object}
}
\value{
The \link{nmr_dataset_peak_table} unchanged

This function is useful for its side-effects: Stopping in case of error
}
\description{
Validate nmr_dataset_peak_table objects
}
