% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_dataset_peak_table.R
\name{validate_nmr_dataset_peak_table}
\alias{validate_nmr_dataset_peak_table}
\title{Validate nmr_dataset_peak_table objects}
\usage{
validate_nmr_dataset_peak_table(nmr_dataset_peak_table)
}
\arguments{
\item{nmr_dataset_peak_table}{An \link{nmr_dataset_peak_table} object}
}
\value{
The \link{nmr_dataset_peak_table} unchanged
}
\description{
Validate nmr_dataset_peak_table objects
}
\seealso{
Other nmr_dataset_peak_table functions: 
\code{\link{[.nmr_dataset_peak_table}()},
\code{\link{format.nmr_dataset_peak_table}()},
\code{\link{is.nmr_dataset_peak_table}()},
\code{\link{load_and_save_functions}},
\code{\link{new_nmr_dataset_peak_table}()},
\code{\link{nmr_meta_add}()},
\code{\link{nmr_meta_export}()},
\code{\link{nmr_meta_get_column}()},
\code{\link{nmr_meta_get}()},
\code{\link{print.nmr_dataset_peak_table}()}

Other class helper functions: 
\code{\link{format.nmr_dataset_1D}()},
\code{\link{format.nmr_dataset_peak_table}()},
\code{\link{format.nmr_dataset}()},
\code{\link{is.nmr_dataset_1D}()},
\code{\link{is.nmr_dataset_peak_table}()},
\code{\link{new_nmr_dataset_1D}()},
\code{\link{new_nmr_dataset_peak_table}()},
\code{\link{new_nmr_dataset}()},
\code{\link{print.nmr_dataset_1D}()},
\code{\link{print.nmr_dataset_peak_table}()},
\code{\link{print.nmr_dataset}()},
\code{\link{validate_nmr_dataset_family}()},
\code{\link{validate_nmr_dataset}()}
}
\concept{class helper functions}
\concept{nmr_dataset_peak_table functions}
